% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{add_ids}
\alias{add_ids}
\title{Add numerical identifiers to ASVs}
\usage{
add_ids(
  read_count,
  input_asv_list = NULL,
  output_asv_list = "",
  sep = ",",
  quiet = T
)
}
\arguments{
\item{read_count}{Data frame or path to a CSV file with columns:
\code{asv}, \code{sample}, \code{replicate}, \code{read_count}.}

\item{input_asv_list}{Data frame or path to a CSV file containing \code{asv}–\code{asv_id}
pairs. Optional; used to harmonize \code{asv_id} values across datasets.}

\item{output_asv_list}{Character string specifying the output file containing
the updated \code{asv}–\code{asv_id} pairs. Optional.}

\item{sep}{Character string specifying the field separator used in input and
output CSV files.}

\item{quiet}{Logical. If \code{TRUE}, suppress informational messages and
only display warnings or errors.}
}
\value{
Data frame with an added \code{asv_id} column.
}
\description{
Add \code{asv_id} values to a data frame or CSV file containing an \code{asv} column.
Previously existing \code{asv}–\code{asv_id} pairs (from earlier datasets) can be taken
into account via \code{input_asv_list}.

If \code{output_asv_list} is provided, the \code{input_asv_list} is updated with newly
observed ASVs and their corresponding \code{asv_id} values.
}
\examples{
\dontrun{
add_ids(read_count_df, input_asv_list = asv_list)
}

}
