% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{add_output_filename}
\alias{add_output_filename}
\title{Add Output Filenames to a Files Table}
\usage{
add_output_filename(files, compress)
}
\arguments{
\item{files}{A data frame, containing one or more of
the following columns: \code{fasta}, \code{fastq_fw}, and \code{fastq_rv}.}

\item{compress}{Logical. If \code{TRUE}, output filenames are given a
\code{.gz} extension. If \code{FALSE}, any existing \code{.gz} extension
is removed.}
}
\value{
The input \code{files} object with additional output filename columns
named \code{fasta_out}, \code{fastq_fw_out}, and/or \code{fastq_rv_out},
corresponding to the input columns that are present.
}
\description{
Generates output filename columns for FASTA and FASTQ input files based on
the requested compression setting. Existing \code{.gz} extensions are first
removed and then added back when \code{compress} is \code{TRUE}.
}
\examples{
files <- data.frame(
  fasta = c("sample1.fasta.gz", "sample2.fasta"),
  fastq_fw = c("sample1_R1.fastq.gz", "sample2_R1.fastq")
)

add_output_filename(files, compress = FALSE)
add_output_filename(files, compress = TRUE)

}
\keyword{internal}
