% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{add_read_counts}
\alias{add_read_counts}
\title{Add read_count to df}
\usage{
add_read_counts(df, dir)
}
\arguments{
\item{df}{Data frame with a \code{fasta} column containing the names of FASTA files.}

\item{dir}{Character string specifying the directory containing the FASTA files.}
}
\value{
Data frame with an additional \code{read_count} column.
}
\description{
Count the number of reads in all FASTA files listed in the \code{fasta} column
of a data frame, and add a \code{read_count} column to it.
}
\examples{
\dontrun{
df <- add_read_counts(df, sorted_dir)
}

}
