% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{ASVspecificCutoff}
\alias{ASVspecificCutoff}
\title{Generate ASV-Specific Cutoff Values}
\usage{
ASVspecificCutoff(
  read_count,
  max_cutoff = 0.05,
  mock_composition = "",
  habitat_proportion = 0.5,
  by_replicate = FALSE,
  outfile = "",
  sep = ","
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv, sample, replicate, read_count.}

\item{max_cutoff}{Numeric; the maximum allowed cutoff value.}

\item{mock_composition}{Data frame or csv file with columns:
sample, action (keep/tolerate), asv.}

\item{habitat_proportion}{Numeric. Value between 0 and 1: for each asv, if the proportion
of reads in a habitat is below this cutoff,
it is considered as a false positive in all samples of the habitat.}

\item{by_replicate}{Logical; if TRUE, compute cutoffs separately by replicates.}

\item{outfile}{Character string: output file. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Data frame with the following columns:
asv_id, replicate (if by_replicate), cutoff
Lines contain only ASVs that have know false positive occurrences.
}
\description{
In some datasets, certain ASVs may occur in many samples with high read counts.
As a result, reads originating from tag-jump or inter sample contamination
may remain after applying a fixed cutoff value in the \emph{LFNvariant} function.
This function computes ASV-specific cutoff values for use in \emph{LFNvariant},
targeting ASVs known to have false-positive occurrences (as identified by
\emph{MakeKnownOccurrences}).

For each ASV, the function:
\itemize{
\item Identifies all false-positive occurrences.
\item Takes the maximum read count among these false positives.
\item Divides this value by the total read count of the ASV across the dataset,
or within each replicate if \code{by_replicate = TRUE}.
}

Because some false positives may not result from tag-jump contamination, the resulting
cutoff values can sometimes be excessively high. Therefore:
\itemize{
\item Filter the dataset as thoroughly as possible \emph{before} computing ASV-specific cutoffs.
\item Set a reasonable upper limit for these cutoffs using the \emph{max_cutoff} parameter.
}
}
\examples{
\dontrun{
cutoffs <- ASVspecificCutoff(read_count=read_count_df, 
    mock_composition="data/mock_composition.csv"
    )
}
}
\seealso{
\code{\link[=LFNvariant]{LFNvariant()}}
}
