% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_graph.R
\name{barplot_renkonen_distance}
\alias{barplot_renkonen_distance}
\title{barplot_renkonen_distance}
\usage{
barplot_renkonen_distance(
  df,
  sample_types = "",
  sep = ",",
  x_axis_label_size = 6
)
}
\arguments{
\item{df}{data frame with the following columns: sample1,sample2,replicate1,replicate2,renkonen_d (can be produced by make_renkonen_distance_matrix)}

\item{sample_types}{file with sample and sample_type (real/mock/negative) columns}

\item{sep}{separator used in csv files}

\item{x_axis_label_size}{size of labels in x axis}
}
\description{
Create barplot with renkonen distances between pairs of sample-replicates
If information is given on sample types, bars are colored in function of them
}
