% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_graph.R
\name{Barplot_RenkonenDistance}
\alias{Barplot_RenkonenDistance}
\title{Barplot_RenkonenDistance}
\usage{
Barplot_RenkonenDistance(
  df,
  sampleinfo = NULL,
  sep = ",",
  x_axis_label_size = 6,
  plotfile = ""
)
}
\arguments{
\item{df}{data frame with the following columns:
sample1,sample2,replicate1,replicate2,renkonen_d
(can be produced by make_renkonen_distance_matrix)}

\item{sampleinfo}{Data frame or CSV file with the following columns:
sample, sample_type (real/mock/negative)}

\item{sep}{separator used in csv files}

\item{x_axis_label_size}{size of labels in x axis}

\item{plotfile}{Character string: png file name for the output plot;
If empty, no file is written.}
}
\description{
Create barplot with renkonen distances between pairs of sample-replicates
If information is given on sample types, bars are colored in function of them
}
