% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{calculate_renkonen_dist}
\alias{calculate_renkonen_dist}
\title{calculate_renkonen_dist}
\usage{
calculate_renkonen_dist(df1, df2)
}
\arguments{
\item{df1}{data frame with asv and read_count columns}

\item{df2}{data frame with asv and read_count columns}
}
\description{
Calculate renkonen distance between two sample-replicate
}
