% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{check_one_to_one_relationship}
\alias{check_one_to_one_relationship}
\title{check_one_to_one_relationship}
\usage{
check_one_to_one_relationship(df)
}
\arguments{
\item{df}{data frame with the following variables: asv_id, asv (can have other columns as well)}
}
\description{
Check if in th input data frame there is a one to one relationship between unique asvs and unisq asv_ids. If yes, returns TRUE, otherwise quit the run
The same asv - asv_id combination can appear more than once in the data frame
}
