% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{CheckFileinfo}
\alias{CheckFileinfo}
\title{CheckFileinfo}
\usage{
CheckFileinfo(file, dir = "", file_type = "fastqinfo", sep = ",")
}
\arguments{
\item{file}{input file name or df to be checked}

\item{dir}{directory name containing the files in fastq_fw, fastq_rv and fasta columns in the input file}

\item{file_type}{\link{fastqinfo/fastainfo/sortedinfo/mock_composition/known_occurrences/read_count/read_count_sample/asv_list}}

\item{sep}{separator used in csv files}
}
\description{
Check the format and coherence of different file types:
Check if all obligatory columns are present (all file_type)
Check sample type, habitat homogeneity across replicates (fastqinfo, fastainfo, sortedinfo)
Check if fastq file pairs are coherent (e.g. 1 to 1 relation; fastqinfo)
Check if tag combinations are unique within a file(pair) (fastqinfo, fastainfo)
Check action (mock_composition, known_occurrences)
Check if 1 to 1 relation between asv_id ad asv (read_count,read_count_sample,asv_list)
}
