% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{ClassifyClusters}
\alias{ClassifyClusters}
\title{Classify clusters based on taxonomic agreement among their ASVs}
\usage{
ClassifyClusters(
  cluster,
  taxa,
  outfile = "",
  sep = ",",
  quiet = TRUE,
  taxlevels = c("domain", "phylum", "class", "order", "family", "genus", "species")
)
}
\arguments{
\item{cluster}{Data frame or csv file with the following variables:
asv_id, cluter_id}

\item{taxa}{Data frame or CSV file with the following variables:
asv_id, one columns per taxonomic level. (e.g. domain, phylum, class, order,
family, genus, species).
Missing values within the lineage are not allowed. If the taxonomic assignment
is incomplete due to low resolution, higher-level taxa should be recorded as \code{NA}.}

\item{outfile}{Character string: name of the output csv file}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{Logical: If TRUE, suppress informational messages and only
show warnings or errors.}

\item{taxlevels}{Character vector with names of the taxonomic levels}
}
\value{
Data frame with the following columns: cluster_id, class for each taxonomic level
}
\description{
Classifies each cluster according to the taxonomic assignment at a given level:
}
\details{
\itemize{
\item closed: ASVs in the cluster are assigned to the same taxon,
and all ASVs of that taxon belong exclusively to this cluster.
\item open: ASVs in the cluster are assigned to the same taxon,
but some ASVs of that taxon are found in other clusters.
\item hybrid: the cluster contains ASVs assigned to more than one taxon.
}
}
\examples{
\dontrun{
plot <- ClassifyClusters(read_count_df, 
                                      swarm_d=7, 
                                      vsearch_path="vsearch",
                                      num_threads=8)
}
}
