% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{Cluster_size}
\alias{Cluster_size}
\title{Cluster_size}
\usage{
Cluster_size(
  read_count_samples,
  id = 0.97,
  vsearch_path = "vsearch",
  outfile = "",
  sep = ","
)
}
\arguments{
\item{read_count_samples}{Data frame or csv file with the following variables:
asv_id, sample, read_count, asv.}

\item{id}{Real; Minimum identity between asv and centroid.}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
read_count_sample data frame, with ASV of the same cluster
and sample grouped to the same line
}
\description{
Cluster ASV in input data frame with the cluster_size command of vsearch.
Replace ASV by their centroids and regroup lines by centroid and sample.
}
\examples{
\dontrun{
clustered_df <- Cluster_size(read_count_samples_df, id=0.97)
}
}
