% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{cluster_vsearch_cluster_size}
\alias{cluster_vsearch_cluster_size}
\title{Cluster all input ASV by cluster_size function of vsearch}
\usage{
cluster_vsearch_cluster_size(
  read_count,
  identity = 0.97,
  vsearch_path = "vsearch",
  num_threads = 0,
  outfile = "",
  sep = ",",
  quiet = TRUE
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, asv, read_count.}

\item{identity}{Real; Identity threshold for clustering.}

\item{vsearch_path}{Character string: path to vsearch executable.}

\item{num_threads}{Positive integer: Number of CPUs.}

\item{outfile}{Character string: name of the output csv file}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{Logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
Data frame with the following columns: asv_id, cluster_id
}
\description{
Cluster all input ASV and return a data frame with asv_id, cluster_id
columns
}
\examples{
\dontrun{
plot <- PairwiseIdentityPlotPerSwarmD(read_count_df, 
                                      swarm_d=7, 
                                      vsearch_path="vsearch",
                                      num_threads=8)
}
}
