% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{ClusterASV}
\alias{ClusterASV}
\title{Cluster ASVs by Swarm or Vsearch}
\usage{
ClusterASV(
  read_count,
  group = TRUE,
  by_sample = FALSE,
  method = "swarm",
  path = "",
  num_threads = 0,
  swarm_d = 1,
  fastidious = T,
  identity = 0.97,
  outfile = "",
  sep = ",",
  quiet = T
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv, sample, replicate (optional), read_count.}

\item{group}{Logical; If TRUE, ASVs of the same cluster are pooled to one row.
if FALSE, input Df is simply completed by a cluser_id column.}

\item{by_sample}{Logical: run clustering separately for each sample.}

\item{method}{Character string: program to use for clustering: swarm or vsearch}

\item{path}{Character string: path to swarm or vsearch executables.}

\item{num_threads}{Positive integer: Number of CPUs.}

\item{swarm_d}{Positive integer: d parameter for swarm (if applicable).
Maximum number of differences allowed between two ASVs,
meaning that two ASVs will be grouped if they have d (or less) differences.}

\item{fastidious}{Logical: When clustering with swarm and working with d = 1,
perform a second clustering pass to reduce the number of small clusters.}

\item{identity}{real: (0-1), the identity threshold used for
clustering when the cluster_size algorithm of vsearch is used.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
read_count_df: same structure as the input, but ASVs of
the same cluster pooled to one row if group==TRUE, or rows kept as in the input
and an additional cluster_id column is added if group==FALSE.
}
\description{
This function runs Swarm or the cluster_size command of Vsearch
on the ASVs in the input data frame.
Each ASV is assigned to a cluster, and the function provides
two possible output formats, controlled by the argument group.
}
\details{
If group = TRUE, the function aggregates the ASVs belonging to the same
cluster into a single row. In this case, the asv_id and asv columns contain
the identifier and the sequence of the cluster's centroid.
read_count is summed over ASVs, samples and replicates are unchanged.

If group = FALSE, the function returns the original input data frame
with an additional column: cluster_id. Each row still corresponds to one ASV.

The clustering can be run on the whole data set at once,
or sample by sample (by_sample)
}
\examples{
\dontrun{
read_count_df <- ClusterASV(read_count=read_count_df, group=TRUE,method="vsearch",
by_sample=TRUE, path=swarm_path, 
num_threads=4)
}
}
