% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{ClusterSize}
\alias{ClusterSize}
\title{ClusterSize}
\usage{
ClusterSize(
  read_count,
  id = 0.97,
  vsearch_path = "vsearch",
  by_sample = FALSE,
  num_threads = 0,
  outfile = "",
  sep = ",",
  quiet = TRUE
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate (optional), read_count, asv.}

\item{id}{Real; Minimum identity between asv and centroid.}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{by_sample}{logical: run clustering separately for each sample.}

\item{num_threads}{Positive integer: Number of CPUs.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
read_count data frame, with ASV of the same cluster, sample, (replicate)
grouped to the same line
}
\description{
Cluster ASV in input data frame with the cluster_size command of vsearch.
Replace ASV by their centroids and regroup lines by centroid and sample.
}
\examples{
\dontrun{
clustered_df <- ClusterSize(read_count_df, id=0.97)
}
}
