% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{codon_stops_from_genetic_code}
\alias{codon_stops_from_genetic_code}
\title{codon_stops_from_genetic_code}
\usage{
codon_stops_from_genetic_code(genetic_code = 5)
}
\arguments{
\item{genetic_code}{genetic code number from https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=cgencodes}
}
\description{
Returns a vector of codon stops corresponding the the genetic code number in https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=cgencodes
}
