% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{compare_df}
\alias{compare_df}
\title{compare_df}
\usage{
compare_df(df1, df2, step = "")
}
\arguments{
\item{df1}{dataframe with columns: "asv", "plate","marker", "sample","replicate","read_count"}

\item{df2}{dataframe with columns: "asv", "plate","marker", "sample","replicate","read_count"}

\item{step}{string to include in the FAIL or PASS message}
}
\description{
Compare two dataframes of and returns teh full join ob the two
}
