% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{compress_file}
\alias{compress_file}
\title{compress_file}
\usage{
compress_file(filename = "", remove_input = F)
}
\arguments{
\item{filename}{name of the gzipped file, including path}

\item{remove_input}{\link{T/F} If TRUE the input file is deleted}
}
\description{
Compress input file to gzip format. Returns the output (compressed) filename
}
