% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{count_reads_dir}
\alias{count_reads_dir}
\title{count_reads_dir}
\usage{
count_reads_dir(
  dir,
  pattern = ".fastq",
  file_type = "",
  outfile = "",
  sep = ",",
  quiet = T
)
}
\arguments{
\item{dir}{name of the input directory}

\item{pattern}{take files from input dir, if patter (regular expression) is present in the file name}

\item{file_type}{\link{fasta/fastq} if not specified, the number of lines is returned, otherwise the nulber of sequences}

\item{outfile}{name of the output file; optional, it is not given results are returned in a data frame, but no file is written}

\item{sep}{separator used in csv files}

\item{quiet}{\link{T/F}; TRUE by default print only warnings and error to STDOUT}
}
\description{
For all files in input directory, where the file name contains pattern,
count the number of sequences in a fasta or fastq file, or  he number of line in other files
Efficient in linux-like systems, but slow on Windows for large files
Can handle gz compressed and uncompressed files, but not zipped files
}
