% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{count_seq2}
\alias{count_seq2}
\title{Count sequences in fasta}
\usage{
count_seq2(file)
}
\arguments{
\item{file}{Character string: input fasta file (including path).}
}
\value{
The number of sequences in the input file.
}
\description{
Count the number of sequences in the input fasta file.
}
\details{
Input can be uncompressed or gzip compressed file,
but other compression types are not supported.
}
\examples{
\dontrun{
n <- count_seq2(file="data/test.fasta")
}
}
