% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_functions.R
\name{count_seq_backup}
\alias{count_seq_backup}
\title{count_seq_backup}
\usage{
count_seq_backup(dir = "", file)
}
\arguments{
\item{dir}{directory that contains the input file}

\item{file}{fasta file, without it's path}
}
\description{
Count the number of sequences in the input fasta file. Input can be uncompressed, gz, or zip file
}
