% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{count_seq_bis}
\alias{count_seq_bis}
\title{count_seq_bis}
\usage{
count_seq_bis(file)
}
\arguments{
\item{file}{fasta file including path}
}
\description{
Count the number of sequences in the input fasta file. Input can be uncompressed of gz file, but zip file are not handled
}
