% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{decompress_file}
\alias{decompress_file}
\title{decompress_file}
\usage{
decompress_file(filename = "", remove_input = F)
}
\arguments{
\item{filename}{name of the gzipped file, including path}

\item{remove_input}{\link{T/F} If TRUE the input file is deleted}
}
\description{
Decompress the input gzipped file. Returns the output filename
}
