% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{delete_1_by_row}
\alias{delete_1_by_row}
\title{delete_1_by_row}
\usage{
delete_1_by_row(row)
}
\arguments{
\item{row}{Vector of taxids.}
}
\value{
Vector of taxids.
}
\description{
Delete all 1 from the beginning of a raw, shift the other values and
replace de missing ones at the end by NA
}
\examples{
delete_1_by_row(row)
}
