% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_graph.R
\name{density_plot_renkonen_distance}
\alias{density_plot_renkonen_distance}
\title{density_plot_renkonen_distance}
\usage{
density_plot_renkonen_distance(df)
}
\arguments{
\item{df}{data frame with the following columns: sample1,sample2,replicate1,replicate2,renkonen_d (can be produced by make_renkonen_distances)}
}
\description{
Create density plot with Renkonen distances between pairs of sample-replicates
}
