% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{dereplicate_reads}
\alias{dereplicate_reads}
\title{dereplicate_reads}
\usage{
dereplicate_reads(files)
}
\arguments{
\item{files}{vector with file names}
}
\description{
reads all input file in the input vector containing file names
demultiplex reads and count them
returns a data frame with asv, asv_id (arbitrary, valid only within the data frame) and read_count columns
}
