% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_helper_functions.R
\name{download_osf}
\alias{download_osf}
\title{Download from OSF}
\usage{
download_osf(filename, url, dest_dir = ".", untar = TRUE, quiet = FALSE)
}
\arguments{
\item{filename}{Character string specifying the name of the file to download.}

\item{url}{Character string; direct download URL of the OSF file.}

\item{dest_dir}{Character string; path to the directory where the file should be downloaded.}

\item{untar}{Logical; if \code{TRUE}, the archive is extracted after download.}

\item{quiet}{Logical; if \code{TRUE}, suppress informational messages and show only warnings or errors.}
}
\value{
Invisible full path to the directory where the file was downloaded.
}
\description{
Downloads a file from the Open Science Framework (OSF).
If the file is a \code{.tar.gz} archive, it can optionally be extracted after download.
}
\examples{
\dontrun{
download_osf(
  filename = "COInr_for_vtam_2025_05_23_dbV5.tar.gz",
  url = "https://osf.io/download/jyhz6/",
  dest_dir = "~/vtamR/OSF",
  untar = TRUE,
  quiet = FALSE
)
}

}
