% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_helper_functions.R
\name{download_zenodo}
\alias{download_zenodo}
\title{Download from Zenodo}
\usage{
download_zenodo(filename, url, dest_dir = ".", untar = TRUE, quiet = FALSE)
}
\arguments{
\item{filename}{Character string specifying the name of the file to download.}

\item{url}{Character string; URL of the Zenodo record (copy from browser address bar).}

\item{dest_dir}{Character string; path to the directory where the file should be downloaded.}

\item{untar}{Logical; if \code{TRUE}, the archive is extracted after download.}

\item{quiet}{Logical; if \code{TRUE}, suppress informational messages and show only warnings or errors.}
}
\value{
Invisible full path to the directory where the file was downloaded.
}
\description{
Downloads a file from Zenodo and optionally extracts it if it is a
\code{.tar.gz} archive.
}
\examples{
\dontrun{
download_zenodo(
  filename = "COInr_2025_05_23.tar.gz",
  url = "https://zenodo.org/records/15515860",
  dest_dir = "~/vtamR",
  untar = TRUE,
  quiet = FALSE
)
}

}
