% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{expand_rows}
\alias{expand_rows}
\title{Expand rows by splitting taxid field}
\usage{
expand_rows(row)
}
\arguments{
\item{row}{A data frame row containing the columns: \code{qseqid}, \code{pident},
\code{qcovhsp}, \code{staxids}.}
}
\value{
Data frame with one row per taxid, with columns: \code{qseqid}, \code{pident},
\code{qcovhsp}, \code{staxids}.
}
\description{
Split entries containing multiple taxonomy IDs into separate rows, ensuring
a one-to-one mapping between hits and taxids.
}
\examples{
\dontrun{
expand_rows(row)
}

}
