% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{FilerRenkonen}
\alias{FilerRenkonen}
\title{FilerRenkonen}
\usage{
FilerRenkonen(
  read_count_df,
  write_csv = F,
  outdir = "",
  renkonen_distance_quantile = 0.9,
  sep = ","
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}

\item{renkonen_distance_quantile}{quantile renkonen distance to determine cutoff value (among all distances, the lowest renkonen_distance_quantile proportion of values are considered as bellow cutoff)}
}
\description{
Filter out all replicates that have renkonen distances above cutoff to most other replicates of the sample.
Returns the filtered read_count_df data frame
}
