% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{filter_min_replicate}
\alias{filter_min_replicate}
\title{Filter occurrences by minimum number of replicates per sample}
\usage{
filter_min_replicate(
  read_count,
  cutoff = 2,
  outfile = NULL,
  sep = ",",
  log_file = NULL
)
}
\arguments{
\item{read_count}{Data frame or path to a CSV file with the following columns:
\code{asv_id}, \code{sample}, \code{replicate}, \code{read_count}, \code{asv}.}

\item{cutoff}{Positive integer specifying the minimum number of replicates
in which an ASV must be detected within a sample to be retained.}

\item{outfile}{Character string specifying the CSV file to write the output
data frame. If NULL, no file is written.}

\item{sep}{Character string specifying the field separator used in input and
output CSV files.}

\item{log_file}{Character string specifying the path to the CSV log file.
The path is resolved with the following priority:
\enumerate{
\item the \code{log_file} argument, if explicitly provided by the user;
\item the \code{vtamR.log_file} option set with
\code{options(vtamR.log_file = ...)}, if set;
\item \code{"vtamR_log.csv"} in the current working directory, used as a
last resort if neither of the above is set.
}
To disable logging entirely, set \code{log_file = NA}.}
}
\value{
Filtered \code{read_count} data frame.
}
\description{
Remove occurrences where an ASV is not present in at least \code{cutoff}
replicates within a sample.
}
\examples{
\dontrun{
filtered_read_count_df <- filter_min_replicate(read_count_df, cutoff = 3)
}

}
