% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{FilterCodonStop}
\alias{FilterCodonStop}
\title{FilterCodonStop}
\usage{
FilterCodonStop(read_count_df, outfile = "", genetic_code = 5, sep = ",")
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv_id, sample, replicate, read_count, asv}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{genetic_code}{genetic code number from https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=cgencodes}
}
\description{
Filter out all ASVs, if there is a codon stop in all three frames of the direct strand.
Returns the filtered read_count_df data frame.
}
