% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{FilterExternalContaminant}
\alias{FilterExternalContaminant}
\title{FilterExternalContaminant}
\usage{
FilterExternalContaminant(
  read_count,
  sample_types,
  outfile = "",
  conta_file = "",
  sep = ","
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{sample_types}{Data frame or csv file with at least the following columns:
sample,sample_type (negative/mock/real).}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{conta_file}{Character string: csv file name to print lines filtered out
from the input dataset.
If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Filtered read_count_df data frame.
}
\description{
Eliminate ASVs that are likely to be external contaminants, since
they have higher read count in at least one negative control, than in any of
the other samples.
}
\examples{
\dontrun{
filtered_read_count_df <- FilterExternalContaminant(read_count_df, sample_types=sortedinfo)
}
}
