% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{FilterIndel}
\alias{FilterIndel}
\title{FilterIndel}
\usage{
FilterIndel(read_count, outfile = "", sep = ",")
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Filtered read_count_df data frame.
}
\description{
Filter out all ASVs, if the modulo 3 of their length is not the same as
that of the majority of the ASVs.
}
\examples{
filtered_read_count_df <-FilterIndel(read_count_df)
}
