% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_functions.R
\name{FilterPCRerror_NW}
\alias{FilterPCRerror_NW}
\title{FilerPCRerror_NW}
\usage{
FilterPCRerror_NW(
  read_count_df,
  by_sample = T,
  write_csv = F,
  outdir = outdir,
  pcr_error_var_prop = 0.1,
  max_mismatch = 1,
  sample_prop = 0.8
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{by_sample}{T/F, if T ASVs are flagged as an PCR error separately for each sample}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}

\item{pcr_error_var_prop}{if the proportion of read_counts of two similar ASVs is bellow pcr_error_var_prop, the less abundant is flagged as a PCR error}

\item{max_mismatch}{maximum number of mismatches (gaps included) to consider two ASVs as similar}

\item{sample_prop}{if by_sample=T, the ASV must be flagged as an PCRerror in sample_prop of the cases to be eliminated}
}
\description{
Filter out all ASVs if they very similar (max_mismatch) to another more frequent ASV (pcr_error_var_prop)
The whole plate can be analysed et once (by_sampe=F)
}
