% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{FilterRenkonen}
\alias{FilterRenkonen}
\title{FilterRenkonen}
\usage{
FilterRenkonen(
  read_count_df,
  outfile = "",
  renkonen_distance_quantile = 0.9,
  sep = ","
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv_id, sample, replicate, read_count, asv}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{renkonen_distance_quantile}{quantile renkonen distance to determine cutoff value (among all distances, the lowest renkonen_distance_quantile proportion of values are considered as bellow cutoff)}
}
\description{
Filter out all replicates that have renkonen distances above cutoff to most other replicates of the sample.
Returns the filtered read_count_df data frame
}
