% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{flag_from_habitat}
\alias{flag_from_habitat}
\title{flag_from_habitat}
\usage{
flag_from_habitat(occurrence_df, fileinfo_df, habitat_proportion = 0.5)
}
\arguments{
\item{occurrence_df}{data frame with the following variables: asv, plate, marker, sample, mean_read_count, sample_type, habitat, action}

\item{fileinfo_df}{csv file with columns: plate, marker, sample, sample_type(mock/negative/real), habitat}

\item{habitat_proportion}{For each of ASVs, if the proportion of reads in a habitat is below this cutoff it is considered as an artifact in all samples of the habitat.}
}
\description{
Flag FP occurrences in samples based on the habitat the samples are from.
All ASVs present in more than one habitat are checked.
For each of these ASVs, if the proportion of reads in a habitat is below a cutoff (habitat_proportion),
it is considered as an artifact in all samples of the habitat.
}
