% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_functions.R
\name{flag_from_negative_controls}
\alias{flag_from_negative_controls}
\title{flag_from_negative_controls}
\usage{
flag_from_negative_controls(occurrence_df, fileinfo_df)
}
\arguments{
\item{occurrence_df}{data frame with the following variables: asv, plate, marker, sample, mean_read_count, sample_type, habitat, action}

\item{fileinfo_df}{csv file with columns: plate, marker, sample, sample_type(mock/negative/real), habitat}
}
\description{
flag all occurrences in all negative controls as "delete" (FP)
}
