% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_functions.R
\name{flagPCRerrors}
\alias{flagPCRerrors}
\title{flagPCRerrors}
\usage{
flagPCRerrors(unique_asv_df, pcr_error_var_prop = 0.1, max_mismatch = 1)
}
\arguments{
\item{pcr_error_var_prop}{if the proportion of read_counts of two similar ASVs is bellow pcr_error_var_prop, the less abundant is flagged as a PCR error}

\item{max_mismatch}{maximum number of mismatches (gaps included) to consider two ASVs as similar}

\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}
}
\description{
Flag unique ASVs if they can be a PCR error
}
