% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{get_lineage_ids}
\alias{get_lineage_ids}
\title{Get the complete taxonomic lineage using taxIDS}
\usage{
get_lineage_ids(taxids, tax_df)
}
\arguments{
\item{taxids}{Vector of taxIDs (taxonomic IDs).}

\item{tax_df}{Data frame with the following columns:
tax_id, parent_tax_id, rank, name_txt, taxlevel
(8: species, 7: genus, 6: family, 5: order, 4: class, 3: phylum, 2: kingdom, 1: superkingdom, 0: root).}
}
\value{
Data frame with each line starting by a taxID followed by a vector taxids
in its full lineage (starting by the highest taxonomic level).
}
\description{
Get the complete taxonomic lineage of each taxID in the input vector.
}
\examples{
taxids <- c(9606, 9593)
tax_df <- read.delim(taxonomy="xxxxxxxx", header=T, sep="\t", fill=T, quote="")
get_lineage_ids(taxids, tax_df=tax_df)
}
