% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_helper_functions.R
\name{get_program_version}
\alias{get_program_version}
\title{Identify the version of a third-party program}
\usage{
get_program_version(name, path)
}
\arguments{
\item{name}{Character string giving the name of the argument used to define
the program path. This is used to handle programs with different version
options (e.g. \code{blast_path} uses \code{-version}).}

\item{path}{Character string giving the path to the third-party program
executable.}
}
\value{
A character string containing the detected program version.
}
\description{
This function runs an external program with its version option and extracts
the version number from the returned message.
}
\examples{
\dontrun{
version <- get_program_version(
  name = "vsearch_path",
  path = "/usr/local/bin/vsearch"
)
}

}
