% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{get_read_counts}
\alias{get_read_counts}
\title{Add read_count to df}
\usage{
get_read_counts(df, dir)
}
\arguments{
\item{df}{Data frame with fasta column containing name of fasta files}

\item{dir}{Character string: directory name with the fasta files.}
}
\value{
Data frame with the read_count column added to the input.
}
\description{
Count the number of reads in all fatsa files in the fasta column of a df.
Add read_count column to df.
}
\examples{
\dontrun{
df <- get_read_counts(df, sorted_dir)
}
}
