% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{get_stop_codons}
\alias{get_stop_codons}
\title{Codon stop set for a genetic code}
\usage{
get_stop_codons(genetic_code = 5)
}
\arguments{
\item{genetic_code}{Integer specifying the NCBI genetic code number.}
}
\value{
Character vector of stop codons.
}
\description{
Return the set of stop codons corresponding to a given genetic code number
as defined by NCBI.

Genetic codes follow the NCBI translation tables:
https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=cgencodes
}
\examples{
\dontrun{
get_stop_codons(genetic_code = 1)
}

}
