% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_graph.R
\name{graph_read_count_by_variant}
\alias{graph_read_count_by_variant}
\title{graph_read_count_by_variant}
\usage{
graph_read_count_by_variant(read_count_df, min_read_count = 0, binwidth = 100)
}
\arguments{
\item{read_count_df}{data frame with asv and read_count columns}

\item{min_read_count}{filter out ASVs with less than min_read_count reads, before making the graph}

\item{binwidth}{width of the read count intervals}
}
\description{
Create histogram with the number of reads in each sample or sample-replicate;
If information is given on sample types, bars are colored in function of them
}
