% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{HistoryBy}
\alias{HistoryBy}
\title{HistoryBy}
\usage{
HistoryBy(dir, feature = "", value = "", sep = ",")
}
\arguments{
\item{dir}{directory containing the output files of intermediate filtering steps; file names should start with a number followed by an underscore (e.g. 5_LFNsampleReplicate.csv)}

\item{feature}{\link{asv_id/asv/sample/replicate/asv_id}}

\item{value}{values of feature that should selected. That output data frame contains all lines where value is present in feature in the input files}

\item{sep}{separator used in csv files}
}
\description{
Filters a feature (asv_id/asv/sample/replicate), in all output files of intermediate filtering steps to retain only lines corresponding a value
}
