% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{is_linux}
\alias{is_linux}
\title{Test if OS is Linux-like}
\usage{
is_linux()
}
\value{
Logical. \code{TRUE} if the \code{"sysname"} returned by
\code{Sys.info()} starts with one of: \code{linux}, \code{sunos},
\code{darwin}, \code{gnu}, or \code{unix}; otherwise \code{FALSE}.
}
\description{
Determine whether the operating system is Unix/Linux-like.
}
\examples{
\dontrun{
os_linux <- is_linux()
}

}
