% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{LFN_global_read_count}
\alias{LFN_global_read_count}
\title{LFN_global_read_count}
\usage{
LFN_global_read_count(read_count_df, cutoff = 10, outfile = "", sep = ",")
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv_id, sample, replicate, read_count, asv}

\item{cutoff}{minimum number of reads for an ASV in the whole dataset; default=10}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{sep}{separator in csv files; default: ","}
}
\description{
Eliminate ASVs with less than cutoff reads in the dataset.
Returns the filtered read_count_df data frame.
}
