% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{LFN_sample_replicate}
\alias{LFN_sample_replicate}
\title{LFN_sample_replicate}
\usage{
LFN_sample_replicate(read_count, cutoff = 0.001, outfile = "", sep = ",")
}
\arguments{
\item{read_count}{data frame or csv file with the following variables: asv_id, sample, replicate, read_count, asv}

\item{cutoff}{minimum proportion for an occurrence within a sample-replicate; default=0.001}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{sep}{separator in csv files; default: ","}
}
\description{
Eliminate occurrences where the read_count/sum(read_count of the sample-replicate) is less than cutoff.
Returns the filtered read_count_df data frame.
}
