% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{LFN_variant}
\alias{LFN_variant}
\title{LFN_variant}
\usage{
LFN_variant(
  read_count_df,
  cutoff = 0.001,
  by_replicate = FALSE,
  outfile = "",
  sep = ","
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv_id, sample, replicate, read_count, asv}

\item{cutoff}{minimum proportion for an occurrence within an asv or an asv-replicate; default=0.001}

\item{by_replicate}{T/F; default=FALSE}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{sep}{separator in csv files; default: ","}
}
\description{
If by_replicate=F: Eliminate occurrences where the read_count/sum(read_count of the asv) is less than cutoff.
If by_replicate=T: Eliminate occurrences where the read_count/sum(read_count of the asv in its replicate) is less than cutoff.
Returns the filtered read_count_df data frame.
}
