% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{LFN_variant_all}
\alias{LFN_variant_all}
\title{LFN_variant_all}
\usage{
LFN_variant_all(read_count_df, cutoff = 0.001, write_csv = F, outdir = NA)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{cutoff}{minimum proportion for an occurrence within an asv; default=0.001}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}
}
\description{
Eliminate occurrences where the read_count/sum(read_count of the asv) is less than cutoff
}
