% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_functions.R
\name{LFN_variant_backup}
\alias{LFN_variant_backup}
\title{LFN_variant}
\usage{
LFN_variant_backup(
  read_count_df,
  cutoff = 0.001,
  by_replicate = FALSE,
  write_csv = F,
  outdir = NA,
  sep = ","
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{cutoff}{minimum proportion for an occurrence within an asv or an asv-replicate; default=0.001}

\item{by_replicate}{T/F; default=FALSE}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}
}
\description{
If by_replicate=F: Eliminate occurrences where the read_count/sum(read_count of the asv) is less than cutoff.
If by_replicate=T: Eliminate occurrences where the read_count/sum(read_count of the asv with its replicate) is less than cutoff.
Returns the filtered read_count_df data frame.
}
