% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{LFNreadCount}
\alias{LFNreadCount}
\title{LFNreadCount}
\usage{
LFNreadCount(read_count, cutoff = 10, outfile = "", sep = ",")
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{cutoff}{Positive integer: minimum number of reads for an occurrence.
Bellow this cutoff, the occurrence (presence of the ASV in a sample-replicate)
is deleted.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Filtered read_count_df data frame.
}
\description{
Eliminate occurrences (presence of the ASV in a sample-replicate)
with less than cutoff reads.
}
\examples{
\dontrun{
filtered_read_count <- LFNreadCount(read_count_df, cutoff=20)
}
}
