% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{LFNsampleReplicate}
\alias{LFNsampleReplicate}
\title{LFNsampleReplicate}
\usage{
LFNsampleReplicate(read_count, cutoff = 0.001, outfile = "", sep = ",")
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{cutoff}{Real (0-1): minimum proportion of the read count of an
occurrence within the reads of its sample-replicate. Bellow this cutoff
the occurrence is deleted.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Filtered read_count_df data frame.
}
\description{
Eliminate occurrences (presence of the ASV in a sample-replicate)
where the \verb{read_count / sum(read_count of the sample-replicate)} is less than
cutoff.
}
\examples{
\dontrun{
filtered_read_count_df <- LFNsampleReplicate(read_count_df, cutoff=0.005)
}
}
