% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{LFNvariant}
\alias{LFNvariant}
\title{LFNvariant}
\usage{
LFNvariant(
  read_count,
  cutoff = NULL,
  asv_specific_cutoffs = NULL,
  by_replicate = FALSE,
  outfile = "",
  lost_asv_file = "",
  sep = ",",
  min_read_count_prop = 0.7
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{cutoff}{Numeric. Value between 0 and 1: minimum proportion of the read count of
an occurrence within all reads of the asv or asv-replicate. Bellow this cutoff
the occurrence is deleted.}

\item{asv_specific_cutoffs}{A data frame or CSV file specifying ASV-specific
cutoff values. May include a \code{replicate} column if \code{by_replicate = TRUE}.}

\item{by_replicate}{Logical; If TRUEn compare read count of the occurrence to the
read counts of the ASV-replicate.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{lost_asv_file}{Character string: name of the CSV file in which to save
information about ASVs whose read counts in the output fall below
\code{min_read_count_prop} of their counts in the input.}

\item{sep}{Field separator character in input and output csv files.}

\item{min_read_count_prop}{Minimum proportion of total reads retained for
each ASV after filtering. ASVs dropping below this threshold trigger a
warning.}
}
\value{
Filtered read_count_df data frame.
}
\description{
Filters out potential false positives arising from tag-jumps or low-level
inter-sample contamination.

When \emph{by_replicate = FALSE}, the function removes occurrences where the
ratio (read_count/read_count of the asv in the dataset) is
lower than \emph{cutoff}.

When \emph{by_replicate = TRUE}, the function removes occurrences where the
ratio (read_count/read_count of the ASV across all samples of the same replicate) is
lower than \emph{cutoff}.

A warning is issued if the total read count of an ASV is reduced below
\emph{min_read_count_prop} after filtering, as this may indicate that the cutoff
value is too high.

By default, a single global cutoff value (argument \emph{cutoff}) is applied to
all ASVs. Alternatively, ASV-specific cutoff values can be provided through
the \emph{asv_specific_cutoffs} argument, either as a data frame or a CSV file.
When both are provided, the global \emph{cutoff} is applied in cases where
an ASV-specific cutoff is not defined or is lower than the global cutoff.
}
\examples{
\dontrun{
read_count_df <- LFNvariant2(
  read_count = "read_counts.csv",
  cutoff = 0.001,
  asv_specific_cutoffs = "asv_cutoffs.csv",
  by_replicate = FALSE
)
}

}
