% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{make_known_occurrences}
\alias{make_known_occurrences}
\title{make_known_occurrences}
\usage{
make_known_occurrences(
  read_count_samples_df,
  fileinfo = "",
  mock_composition = "",
  sep = ",",
  known_occurrences = "",
  missing_occurrences = "",
  habitat_proportion = 0.5
)
}
\arguments{
\item{read_count_samples_df}{data frame with the following variables: asv_id, plate, marker, sample, read_count, asv}

\item{fileinfo}{csv file with columns: sample, sample_type(mock/negative/real), habitat, replicate, (optional: file)}

\item{mock_composition}{csv file with columns: sample, action (keep/tolerate), asv}

\item{sep}{separator used in csv files}

\item{known_occurrences}{name of the output file containing known occurrences (TP in mock, and FP)}

\item{missing_occurrences}{name of the output file containing the missing occurrences (FN); file is written if only the name has been specified}

\item{habitat_proportion}{for each asv, if the proportion of reads in a habitat is below this cutoff, it is considered as an artifact in all samples of the habitat}
}
\description{
Prepare a file that list all occurrences that are clearly a TP (expected variants in mock) or FP (unexpected variants in mocks, all variants in negative control samples, variants present in a wrong habitat)
returns a data frame with the number of False Positives, False Negatives, True Positives
}
