% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_functions.R
\name{make_known_occurrences_backup}
\alias{make_known_occurrences_backup}
\title{make_known_occurrences_backup}
\usage{
make_known_occurrences_backup(
  read_count_samples_df,
  fileinfo = "",
  mock_composition = "",
  sep = ",",
  out = "",
  missing_occurrences = "",
  habitat_proportion = 0.5
)
}
\arguments{
\item{read_count_samples_df}{data frame with the following variables: asv, plate, marker, sample, read_count}

\item{fileinfo}{csv file with columns: plate, marker, sample, sample_type(mock/negative/real), habitat, replicate, (optional: file)}

\item{mock_composition}{csv file with columns: plate, marker, sample, action (keep/tolerate), asv}

\item{sep}{separator used in fileinfo and mock_composition}

\item{out}{name of the output file containing known occurrences (TP in mock, and FP)}

\item{missing_occurrences}{name of the output file containing the missing occurrences (FN); file is written only the name has been specified}

\item{habitat_proportion}{for each asv, if the proportion of reads in a habitat is below this cutoff, it is considered as an artifact in all samples of the habitat}
}
\description{
Prepare a file that list all occurrences that are clearly a TP (expected variants in mock) or FP (unexpected variants in mocks, all variants in negative control samples, variants present in a wrong habitat)
}
