% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{make_ltg}
\alias{make_ltg}
\title{make_ltg}
\usage{
make_ltg(taxids, lineages, phit = 70)
}
\arguments{
\item{taxids}{vector taxIDs (taxonomic IDs); there can be duplicated values}

\item{lineages}{data frame: taxids in the first column followed by their lineages represented by taxids (starting at the lowest resolution)}

\item{phit}{Percentage of hit that should be included to the LTG; 70 by default}
}
\description{
Determine the Lowest Taxonomic Group (LTG) that contains phit percentage of the input taxids
}
