% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_functions.R
\name{make_missing_occurrences_backup}
\alias{make_missing_occurrences_backup}
\title{make_missing_occurrences_backup}
\usage{
make_missing_occurrences_backup(
  read_count_samples_df,
  mock_composition = "",
  sep = ",",
  out = ""
)
}
\arguments{
\item{read_count_samples_df}{data frame with the following variables: asv, plate, marker, sample, read_count}

\item{mock_composition}{csv file with columns: plate, marker, sample, action (keep/tolerate), asv}

\item{sep}{separator used in fileinfo and mock_composition}

\item{out}{name of the output file}
}
\description{
Prepare a file that list all expected occurrences that are missing (False negatives)
}
