% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{make_renkonen_df}
\alias{make_renkonen_df}
\title{make_renkonen_df}
\usage{
make_renkonen_df(read_count_df)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}
}
\description{
Calculate renkonen distances between all pairs  of replicates within samples.
Return a data frame with the following columns: sample, replicate1, replicate2, renkonen_d
}
