% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{make_renkonen_distances}
\alias{make_renkonen_distances}
\title{make_renkonen_distances}
\usage{
make_renkonen_distances(read_count_df, compare = "all")
}
\arguments{
\item{read_count_df}{data frame with asv, sample, replicate, and read_count columns}

\item{compare}{\link{all/within} calculate the Renkonen distance among all pairs of sample-replicates (all), only between replicates of the same sample (within)}
}
\description{
Calculate the Renkonen distance between pairs of sample-replicates
Returns a data frame with the following columns: sample1,sample2,replicate1,replicate2,renkonen_d,sample_comp (within, if sample1 and sample2 are identical, between otherwise)
}
