% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{MakeRenkonenDistances}
\alias{MakeRenkonenDistances}
\title{Make Renkonen Distances}
\usage{
MakeRenkonenDistances(read_count_df, compare_all = FALSE)
}
\arguments{
\item{read_count_df}{Data frame with asv, sample, replicate, and read_count columns.}

\item{compare_all}{Boolean: if TRUE calculate the Renkonen distance among all pairs
of sample-replicates. Only between replicates of the same sample otherwise.}
}
\value{
Data frame with the following columns:
sample1,sample2,replicate1,replicate2,renkonen_d,sample_comp
(within, if sample1 and sample2 are identical, between otherwise).
}
\description{
Calculate the Renkonen distance between pairs of sample-replicates.
}
\examples{
\dontrun{
MakeRenkonenDistances(read_count_df, compare_all=FALSE)
}
}
