% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{Merge}
\alias{Merge}
\title{Merge}
\usage{
Merge(
  fastqinfo,
  fastq_dir,
  vsearch_path = "",
  outdir = "",
  fastq_ascii = 33,
  fastq_maxdiffs = 10,
  fastq_maxee = 1,
  fastq_minlen = 50,
  fastq_maxlen = 500,
  fastq_minmergelen = 50,
  fastq_maxmergelen = 1000,
  fastq_maxns = 0,
  fastq_truncqual = 10,
  fastq_minovlen = 50,
  fastq_allowmergestagger = F,
  sep = ",",
  compress = F,
  quiet = T
)
}
\arguments{
\item{fastqinfo}{data frame or csv file with column: tag_fw,primer_fw,tag_rv,primer_rv,sample,sample_type(mock/negative/real),habitat(optional),replicate,fastq_fw,fastq_rv}

\item{fastq_dir}{directory with input fastq files (listed in fastqinfo_df$fastq_fw and fastqinfo_df$fastq_rv)}

\item{vsearch_path}{path to vsearch executables}

\item{outdir}{output directory}

\item{fastq_ascii}{\link{33/64} ASCII character number used as the basis for the FASTQ quality score; default: 33}

\item{fastq_maxdiffs}{maximum number of non-matching nucleotides allowed in the overlap region; default:10}

\item{fastq_maxee}{discard sequences with more than the specified number of expected errors; default:}

\item{fastq_minlen}{discard sequences with less than fastq_minlen bases; default: 50}

\item{fastq_maxlen}{discard sequences with more than fastq_maxlen bases; default: 500}

\item{fastq_minmergelen}{minimum length of the merged sequence; default: 50}

\item{fastq_maxmergelen}{maximum length of the merged sequence; default: 1000}

\item{fastq_maxns}{discard sequences with more than fastq_maxns of N’s; default: 0}

\item{fastq_truncqual}{truncate sequences starting from the first base with fastq_truncqual base quality score value or lower; default: 10}

\item{fastq_minovlen}{minimum overlap between the merged reads; default: 50}

\item{fastq_allowmergestagger}{\link{T/F} allow to merge staggered read pairs. Staggered pairs are pairs where the 3’ end of the reverse read has an overhang to the left of the 5’ end of the forward read; default: F}

\item{sep}{separator in input and output csv files; default: ","}

\item{compress}{\link{T/F}; Compress output file to gzip format; Default=F}

\item{quiet}{\link{T/F}; TRUE by default print only warnings and error to STDOUT}
}
\description{
Merge forward are reverse fastq read pairs to fasta.
Output fasta files can be compressed if compress option is used.
The output fastainfo.csv file is similar to the fastqinfo file, but the fastq columns are replaced by a fasta column with the name of the output files
Returns data frame corresponding to the output csv file
}
