% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_functions.R
\name{OptimizeLFNReaCountAndLFNvariant}
\alias{OptimizeLFNReaCountAndLFNvariant}
\title{OptimizeLFNReaCountAndLFNvariant}
\usage{
OptimizeLFNReaCountAndLFNvariant(
  read_count_df,
  known_occurrences = "",
  sep = sep,
  outdir = "",
  min_lfn_read_count_cutoff = 10,
  min_lnf_variant_cutoff = 0.001,
  by_replicate = FALSE,
  lfn_sample_replicate_cutoff = 0.001,
  pcr_error_var_prop = 0.1,
  vsearch_path = "",
  max_mismatch = 1,
  by_sample = T,
  sample_prop = 0.8,
  min_replicate_number = 2
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{known_occurrences}{file produced by make_known_occurrences function, with known FP and TP}

\item{sep}{separator used in fileinfo and mock_composition}

\item{outdir}{name of the output directory}

\item{min_lfn_read_count_cutoff}{the lowest cutoff value for LFN_read_count function to start from (10 by default). Values from this value to 100 by increments of 5 and tested}

\item{min_lnf_variant_cutoff}{the lowest cutoff value for LFN_variant function to start from (0.001 by default). Values from this value to 0.05 by increments of 0.001 and tested}

\item{by_replicate}{T/F (False by default); see LFN_variant function}

\item{lfn_sample_replicate_cutoff}{cutoff value for LFN_sample_replicate (see LFN_sample_replicate function; default 0.001)}

\item{pcr_error_var_prop}{cutoff value for FilterPCRerror (see FilterPCRerror function; default 0.1)}

\item{vsearch_path}{path to vsearch executables}

\item{max_mismatch}{parameter for FilterPCRerror (see FilterPCRerror function; default 1)}

\item{by_sample, parameter}{for FilterPCRerror (see FilterPCRerror function; default T)}

\item{sample_prop}{for FilterPCRerror (see FilterPCRerror function; default 0.8)}

\item{min_replicate_number}{for FilterMinReplicateNumber (see FilterMinReplicateNumber function; default 2)}
}
\description{
Suggest optimal parametres for lfn_read_count_cutoff and lfn_read_count_cutoff.
This script will run LFN_sample_replicate, FilterPCRerror and FilterMinReplicateNumber on the input
read_count_df using parameters set by the user (ideally optimized ones), to eliminate part of the noise.
The the LFN_read_count and LFN_variant is run for a series of parameter value combinations, end after each the number of FN, TP, and FP is reported.
The results are written to the OptimizeLFNReaCountAndLFNvariant.csv.
Users should chose the parameter setting that minimizes, FN and FP.
}
