% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_functions.R
\name{OptimizeLFNsampleReplicate}
\alias{OptimizeLFNsampleReplicate}
\title{OptimizeLFNsampleReplicate}
\usage{
OptimizeLFNsampleReplicate(
  read_count_df,
  mock_composition = "",
  sep = ",",
  outdir = ""
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{mock_composition}{csv file with columns: plate, marker, sample, action (keep/tolerate), asv}

\item{sep}{separator used in mock_composition}

\item{outdir}{name of the output directory}
}
\description{
Prepare a file that lists all expected occurrences in all mock sample replicates their read_counts and the proportion of
read_counts to the total number of reads in the sample-replicate.
The lfn_variant_replicate parameter should be bellow the smallest proportion in order to keep all expected ASVs in the dataset.
}
