% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{OptimizePCRerror}
\alias{OptimizePCRerror}
\title{OptimizePCRerror}
\usage{
OptimizePCRerror(
  read_count,
  mock_composition = "",
  vsearch_path = "vsearch",
  sep = ",",
  outfile = "",
  max_mismatch = 1,
  min_read_count = 2
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{mock_composition}{Data frame or csv file with columns:
sample, action (keep/tolerate), asv.}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{sep}{Field separator character in input and output csv files.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{max_mismatch}{Positive integer: maximum number of mismatches
allowed between two asv to be compared.}

\item{min_read_count}{Positive integer: occurrences with fewer
reads are ignored.}
}
\value{
Data frame with the following columns: sample,expected_read_count,
unexpected_read_count,pcr_error_var_prop,expected_asv_id,unexpected_asv_id,
expected_asv,unexpected_asv
}
\description{
Prepare a data frame that lists pairs of expected and unexpected ASVs
in mock samples with maximum \code{max_mismatch} difference between them.
The \code{pcr_error_var_prop} parameter should be above the highest
\code{pcr_error_var_prop} (unexpected_read_count/expected_read_count)
value in the table.
}
\examples{
\dontrun{
OptimizePCRerror(read_count=read_count_df, 
    mock_composition="data/mock_composition.csv", 
    vsearch_path=vsearch_path, 
    max_mismatch=2, 
    min_read_count=5
    )
}
}
