% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{PairwiseIdentity}
\alias{PairwiseIdentity}
\title{Pairwise identity}
\usage{
PairwiseIdentity(
  asv,
  min_id = 0.8,
  vsearch_path = vsearch,
  num_threads = 0,
  outfile = "",
  sep = ",",
  quiet = TRUE
)
}
\arguments{
\item{asv}{Data frame or csv file with asv and asv_id columns.}

\item{min_id}{Real: Bellow this percentage of identity do not align asv}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{Logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
Data frame with asv pairs and their percentage of identity.
asv pairs with identity bellow min_id are not listed.
}
\description{
Align all pairs of asv with at least min_id similarity and
make a data frame with pairs of asv and their percentage of identity.
asv pairs with identity bellow min_id are not listed.
}
\examples{
\dontrun{
identity_df <- PairwiseIdentity(asv, 
                                min_id = 0.8, 
                                vsearch_path=vsearch, 
                                num_threads=8)
}
}
