% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{PairwiseIdentityPlotPerClusterIdentityThreshold}
\alias{PairwiseIdentityPlotPerClusterIdentityThreshold}
\title{Make a density plot}
\usage{
PairwiseIdentityPlotPerClusterIdentityThreshold(
  read_count,
  identity_min = 0.9,
  identity_max = 0.99,
  identity_increment = 0.01,
  min_id = 0.8,
  vsearch_path = "vsearch",
  num_threads = 0,
  outfile = "",
  plotfile = "",
  sep = ",",
  quiet = TRUE
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv, sample, replicate (optional), read_count.}

\item{identity_min}{Numeric. Value between 0 and 1: Minimum identity threshold between asv and centroid.}

\item{identity_max}{Numeric. Value between 0 and 1: Maximum identity threshold between asv and centroid.}

\item{identity_increment}{Numeric. Value between 0 and 1: Identity thresholds vary from identity_min
to identity_max by identity_increment.}

\item{min_id}{Numeric. Value between 0 and 1: Bellow this pairwise identity asv pairs are not aligned
and their identity is not plotted.}

\item{vsearch_path}{Character string: path to vsearch executable.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}

\item{outfile}{Character string: csv file name with the following columns:
pairwise_asv_identity, cluster, clustering_parameter; If empty, no file is written.}

\item{plotfile}{Character string: png file name for the output plot;
If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{Logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
A density plot of pairwise percentage of identities.
}
\description{
Make a density plot pf pairwise percent identities between ASVs within and across
clusters using different clustering identity thresholds.

Cluster all ASVs using cluster_size algorithm of vsearch with a range of identity values.

For each clustering identity threshold, make a density plot of pairwise
percentage of identities between ASVs, using different colors for identities
between ASVs of the same or different clusters.
}
\examples{
\dontrun{
plot <- PairwiseIdentityPlotPerClusterIdentityThreshold(read_count_df, 
                                      identity_min=0.9, 
                                      identity_max=0.99,
                                      identity_increment=0.01,
                                      min_id = 0.8, 
                                      vsearch_path="vsearch", 
                                      num_threads=8,
                                      plotfile="density_plot.png")
}
}
