% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{PairwiseIdentityPlotPerClusterIdThreshold}
\alias{PairwiseIdentityPlotPerClusterIdThreshold}
\title{Pairwise identity density plot for different \% identity thresholds of clustering}
\usage{
PairwiseIdentityPlotPerClusterIdThreshold(
  read_count,
  identity_min = 0.9,
  identity_max = 0.99,
  identity_increment = 0.01,
  min_id = 0.8,
  vsearch_path = "vsearch",
  num_threads = 0,
  outfile = "",
  sep = ",",
  quiet = TRUE
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv, sample, replicate (optional), read_count.}

\item{identity_min}{Real; Minimum identity threshold between asv and centroid.}

\item{identity_max}{Real; Maximum identity threshold between asv and centroid.}

\item{identity_increment}{Real; Identity thresholds vary from identity_min
to identity_max by identity_increment.}

\item{min_id}{Real: Bellow this percentage of identity asv pairs are not aligned
and their identity is not plotted.}

\item{vsearch_path}{Character string: path to vsearch executable.}

\item{num_threads}{Positive integer: Number of CPUs.}

\item{outfile}{Character string: csv file name with the following columns:
identity, cluster, cluster_criterium; If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{Logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
A density plot of pairwise percentage of identities.
}
\description{
Cluster all asv using cluster_size algorithm of vsearch with a range of identity values.
}
\details{
For each identity, make a density plot of pairwise percentage of identities between
asv, using different colors for identities between asv of the same or different
clusters.
}
\examples{
\dontrun{
plot <- PairwiseIdentityPlotPerClusterIdThreshold(read_count_df, 
                                      identity_min=0.9, 
                                      identity_max=0.99,
                                      identity_increment=0.01,
                                      min_id = 0.8, 
                                      vsearch_path="vsearch", 
                                      num_threads=8,
                                      outfile="density_plot.png")
}
}
