% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{PairwiseIdentityPlotPerSwarmD}
\alias{PairwiseIdentityPlotPerSwarmD}
\title{Pairwise identity density plot for different Swarm d values}
\usage{
PairwiseIdentityPlotPerSwarmD(
  read_count,
  swarm_d_min = 1,
  swarm_d_max = 15,
  swarm_d_increment = 1,
  min_id = 0.8,
  vsearch_path = "vsearch",
  swarm_path = "swarm",
  num_threads = 0,
  outfile = "",
  sep = ",",
  quiet = TRUE
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv, sample, replicate (optional), read_count.}

\item{swarm_d_min}{Positive integer: Minimum value of d for Swarm.}

\item{swarm_d_max}{Positive integer: Maximum value of d for Swarm.}

\item{swarm_d_increment}{Positive integer: increase d by swarm_d_increment between
swarm_d_min and swarm_d_max.}

\item{min_id}{Real: Bellow this percentage of identity asv pairs are not aligned
and their identity is not plotted.}

\item{vsearch_path}{Character string: path to vsearch executable.}

\item{swarm_path}{Character string: path to swarm executable.}

\item{num_threads}{Positive integer: Number of CPUs.}

\item{outfile}{Character string: png file name to print the output plot.
If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{Logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
A density plot pairwise percentage of identities.
}
\description{
Cluster by swarm all asv with a range of d values.
For each d, make a density plot of pairwise percentage of identities between
asv, using different colors for identities between asv of the same or different
clusters.
}
\examples{
\dontrun{
plot <- PairwiseIdentityPlotPerSwarmD(read_count_df, 
                                      swarm_d_min=2, 
                                      swarm_d_max=12,
                                      swarm_d_increment=2,
                                      min_id = 0.8, 
                                      vsearch_path="vsearch", 
                                      swarm_path="swarm",
                                      num_threads=8,
                                      outfile="density_plot.png")
}
}
