% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{PlotClusterClasstification}
\alias{PlotClusterClasstification}
\title{Plot Cluster Classification according to taxa}
\usage{
PlotClusterClasstification(
  read_count,
  taxa,
  clustering_method = "swarm",
  cluster_params = c(2, 4, 6, 8, 10),
  vsearch_path = "vsearch",
  swarm_path = "swarm",
  taxlevels = c("species", "genus"),
  outfile = "",
  plotfile = "",
  sep = ",",
  num_threads = 0,
  quiet = TRUE
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, asv, read_count}

\item{taxa}{Data frame or CSV file with the following variables:
asv_id, one columns per taxonomic level. (e.g. domain, phylum, class, order,
family, genus, species).
Missing values within the lineage are not allowed. If the taxonomic assignment
is incomplete due to low resolution, higher-level taxa should be recorded as \code{NA}.}

\item{clustering_method}{character; swarm or vsearch}

\item{cluster_params}{numerical vector of either swarm's d to be used, or
percentage of percentage of identity threshold for the cluster_size algorithm of vsearch.}

\item{vsearch_path}{Character string: path to vsearch executable.}

\item{swarm_path}{Character string: path to swarm executable.}

\item{taxlevels}{Character vector with names of the taxonomic levels to be classed and plotted}

\item{outfile}{Character string: name of the output csv file with the following
columns: classification, number_of_clusters, taxlevel, clustering_parameter;
If empty, no file is written.}

\item{plotfile}{Character string: png file name for the output plot;
If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}

\item{quiet}{Logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
A connected scatterplot of the number of clusters in different classes
(open, closed, hybrid)
}
\description{
Cluster ASV with using different clustering parameters, than classify each
cluster at each cluster setting :
\itemize{
\item closed: ASVs in the cluster are assigned to the same taxon,
and all ASVs of that taxon belong exclusively to this cluster.
\item open: ASVs in the cluster are assigned to the same taxon,
but some ASVs of that taxon are found in other clusters.
\item hybrid: the cluster contains ASVs assigned to more than one taxon.
}
}
\details{
Plot the number of clusters in each class for each setting.
}
\examples{
\dontrun{
plot <- PlotClusterClasstification(read_count, 
                                   taxa,
                                   clustering_method = "swarm"
                                   cluster_params = c(2, 4, 6, 8, 10)
                                   swarm_path= swarm_path,
                                   taxlevels = c(species, genus),
                                   num_threads=8)
}
}
