% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{pool_2df}
\alias{pool_2df}
\title{pool_2df}
\usage{
pool_2df(df1, df2)
}
\arguments{
\item{df1}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{df2}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}
}
\description{
pool 2 count_read_df data frames, and keep only occurrences present in all filters
}
