% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{PoolFilters}
\alias{PoolFilters}
\title{PoolFilters}
\usage{
PoolFilters(..., outfile = "", sep = ",")
}
\arguments{
\item{...}{a list of data frames}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{sep}{separator in csv files; default: ","}
}
\description{
pool all count_read_df data frames, and keep only occurrences present in all filters
}
