% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{PoolReplicates}
\alias{PoolReplicates}
\title{Pool replicates sample by sample}
\usage{
PoolReplicates(
  read_count_df,
  digits = 0,
  write_csv = F,
  outdir = "",
  sep = ","
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{digits}{round the mean read counts to digits}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}
}
\description{
Take the mean non-zero read counts over replicates for each sample and asv.
Returns a data frame with the following columns: asv, plate, marker, sample, mean_read_count (over replicates)
}
