% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{RandomSampleFastaLinux}
\alias{RandomSampleFastaLinux}
\title{RandomSampleFastaLinux}
\usage{
RandomSampleFastaLinux(
  fasta,
  outfile,
  n = 1e+06,
  vsearch_path = "vsearch",
  compress_method = "R",
  pigz_path = "pigz",
  randseed = NULL,
  quiet = TRUE,
  num_threads = 0
)
}
\arguments{
\item{fasta}{Character string: Path to the input FASTA file. Can be gzipped.}

\item{outfile}{Character string: Path to the output FASTA file. If it ends with \code{.gz},
the output will be gzip-compressed.}

\item{n}{Positive integer: Number of sequences to randomly select.}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{compress_method}{Character or logical. Compression method: \code{"pigz"}, \code{"gzip"}, or \code{"R"}.
\code{"pigz"} requires \code{pigz} to be installed and in the system path (or \code{pigz_path} specified).
\code{"gzip"} is Linux-only.
\code{"R"} uses \code{R.utils}, which is cross-platform but slower.
Speed: R.utils < gzip (only linux) < pig}

\item{pigz_path}{Character string: Path to \code{pigz} executable. Only needed is pigz
is used for file compression, and it is not in the PATH.}

\item{randseed}{Positive integer or NULL: seed for random sampling.
NULL or 0 by default means to use a pseudo-random seed.
A given non-zero seed produces always the same result.}

\item{quiet}{Logical: If TRUE, suppress informational messages; only warnings and errors are shown.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}
}
\value{
Invisibly the number of sequences in the output file.
}
\description{
Randomly select \code{n} sequences from an input FASTA file using \code{fastx_subsample} from
VSEARCH.
}
\details{
The input FASTA can be uncompressed or gzipped. The output compression is
determined automatically from the output file name (i.e., ends with \code{.gz} → gzip).

If the number of sequences in the input file is less than or equal to the
requested number (\code{n}), the input file is simply copied to the output,
respecting the compression inferred from the output file name.
}
\examples{
\dontrun{
RandomSampleFastaLinux(
  fasta = "all_sequences.fasta.gz",
  outfile = "subset_100.fasta.gz",
  vsearch_path="vsearch",
  n = 100,
  randseed = 123,
  quiet = FALSE
)
}
}
