% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_functions.R
\name{RandomSeq_backup}
\alias{RandomSeq_backup}
\title{RandomSeq_backup}
\usage{
RandomSeq_backup(fastainfo_df, fasta_dir = "", outdir = "", n)
}
\arguments{
\item{fastainfo_df}{data frame with a 'fasta' column containing input file names; files can be compressed in gz and zip format}

\item{fasta_dir}{directory that contains the input fasta files}

\item{outdir}{directory for the output files}

\item{n}{integer; the number of randomly selected sequences}
}
\description{
Random select n sequences from each input fasta file. The output is the same compression type (if any) as the input
}
