% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{RandomSeq_windows}
\alias{RandomSeq_windows}
\title{RandomSeq_windows}
\usage{
RandomSeq_windows(
  fastainfo_df,
  fasta_dir = "",
  outdir = "",
  n,
  randseed = 0,
  compress = T
)
}
\arguments{
\item{fastainfo_df}{data frame with a 'fasta' column containing input file names; files can be compressed in gz and zip format}

\item{fasta_dir}{directory that contains the input fasta files}

\item{outdir}{directory for the output files}

\item{n}{integer; the number of randomly selected sequences}

\item{randseed}{positive integer; seed for random sampling; 0 by default means to use a pseudo-random seed, a given non-zero seed produce always the same result}

\item{compress}{\link{T/F} is TRUE, the output file will be compressed}
}
\description{
Random select n sequences from each input fasta file. The output is the same compression type (if any) as the input
}
