% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{RandomSeqWindows}
\alias{RandomSeqWindows}
\title{RandomSeqWindows}
\usage{
RandomSeqWindows(
  fastainfo,
  fasta_dir = "",
  outdir = "",
  n,
  randseed = 0,
  compress = T,
  sep = ","
)
}
\arguments{
\item{fastainfo}{data frame or csv file with a 'fasta' column containing input file names; files can be compressed in gz and zip format}

\item{fasta_dir}{directory that contains the input fasta files}

\item{outdir}{directory for the output files}

\item{n}{integer; the number of randomly selected sequences}

\item{randseed}{positive integer; seed for random sampling; 0 by default means to use a pseudo-random seed, a given non-zero seed produce always the same result}

\item{compress}{\link{T/F} is TRUE, the output file will be compressed}

\item{sep}{separator used in csv files; default: ","}
}
\description{
Random select n sequences from each input fasta file. The output is the same compression type (if any) as the input
}
