% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_test.R
\name{read_asv_table}
\alias{read_asv_table}
\title{read_asv_table}
\usage{
read_asv_table(filename, sep = ",")
}
\arguments{
\item{filename}{name of the input file including full path;
columns: asv, seq_id, plate.marker.sample.replicate columns containing read counts}

\item{sep}{separator; default ","}
}
\description{
Read asv table in wid format to a data frem in long format
}
