% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{read_asv_table_sample}
\alias{read_asv_table_sample}
\title{read_asv_table_sample}
\usage{
read_asv_table_sample(filename, sep = ",")
}
\arguments{
\item{filename}{name of the input file including full path; columns: asv, seq_id, plate.marker.sample columns containing read counts}

\item{sep}{separator; default ","}
}
\description{
Read asv table in wid format to a data frem in long format
}
