% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{read_blast4out}
\alias{read_blast4out}
\title{read_blast4out}
\usage{
read_blast4out(filename)
}
\arguments{
\item{filename}{Tab separeted file with merged and centroid asv_ids in the first 2 columns}
}
\value{
data frame with merged_id and centroid_id
}
\description{
read output of vsearch --cluster_size to df
}
