% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{read_blast_results}
\alias{read_blast_results}
\title{Read BLAST results into a data frame}
\usage{
read_blast_results(file)
}
\arguments{
\item{file}{Character string specifying the BLAST output file. The file must
be tab-separated and contain the columns: \code{qseqid}, \code{pident}, \code{qcovhsp}, \code{staxids}.}
}
\value{
Data frame with columns: \code{qseqid}, \code{pident}, \code{qcovhsp}, \code{staxids}.
}
\description{
Import BLAST output into a structured data frame. If a hit contains multiple
taxonomy IDs (\code{staxids}), each taxid is split into a separate row.
}
\examples{
\dontrun{
read_blast_results(file = "blastout.txt")
}

}
