% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{read_fasta_count_reads}
\alias{read_fasta_count_reads}
\title{Read a fasta file, demultiplex reads and count them}
\usage{
read_fasta_count_reads(file)
}
\arguments{
\item{file}{tsv file with columns: plate    marker    sample    replicate    file}
}
\description{
Read only sequences from input fasta file.
Transform all nt to upper case.
Demultiplex reads to ASVs.
Count the number of reads of each ASVs.
Return a tibble with asv and nb_reads.
}
