% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_treatment.R
\name{read_fasta_local}
\alias{read_fasta_local}
\title{read_fasta_local}
\usage{
read_fasta_local(filename, use_id = T, case = "", dereplicate = F)
}
\arguments{
\item{filename}{input fasta file with full path; can be  uncompressed o compressed in gz format only}

\item{use_id}{\link{T/F} if TRUE returns data frame with seq_id and sequence columns; not compatible with dereplicate==T, where the seq_ids are lost; Default: T}

\item{case}{\link{"uc"/"lc"/""} write sequences with upper or lower case letters; Default: "", do not change the case}

\item{dereplicate}{\link{T/F} if TRUE, returns a data frame with unique sequences in the sequence column and read counts in the read_count column; Default: F}
}
\description{
Read a fasta file to a dataframe
}
