% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{read_fasta_seq}
\alias{read_fasta_seq}
\title{Read sequences from a FASTA file.

Input FASTA files can be gzip-compressed or uncompressed.}
\usage{
read_fasta_seq(filename = filename, dereplicate = F)
}
\arguments{
\item{filename}{Character string specifying the input FASTA file (including full path).}

\item{dereplicate}{Logical. If \code{TRUE}, return ASVs with read counts instead of individual reads.}
}
\value{
Data frame. If \code{dereplicate = FALSE}, one read per row is returned. If
\code{dereplicate = TRUE}, a data frame with \code{asv} and \code{read_count} columns is returned.
}
\description{
Read sequences from a FASTA file.

Input FASTA files can be gzip-compressed or uncompressed.
}
\examples{
\dontrun{
read_df <- read_fasta_seq(filename = "data/test.fasta", dereplicate = FALSE)
asv_df <- read_fasta_seq(filename = "data/test.fasta", dereplicate = TRUE)
}

}
