% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{read_fastas_from_fileinfo}
\alias{read_fastas_from_fileinfo}
\title{Read all fasta files in fileinfo file to a data frame}
\usage{
read_fastas_from_fileinfo(
  sortedinfo_df,
  dir = "",
  outfile = "",
  sep = ",",
  asv_list = ""
)
}
\arguments{
\item{sortedinfo_df}{data frame with columns:  sample,  replicate, filename, (optional: sample_type(mock/negative/real), habitat)}

\item{dir}{name of the directory with fasta files}

\item{outfile}{Name of the output csv file with "asv", "sample", "replicate", "read_count" as columns; if NULL, only a data frame is returned}

\item{sep}{separator in csv files; default: ","}
}
\description{
Read all fasta file in sortedinfo_df data frame (sample	replicate	filename).
Dereplicate reads to ASVs.
Count the number of reads of each ASVs in each sample-replicate.
Returns a df ("asv", "sample", "replicate", "read_count")
}
