% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_functions.R
\name{read_fastas_from_fileinfo_backup}
\alias{read_fastas_from_fileinfo_backup}
\title{Read all fasta files in fileinfo file to a data frame}
\usage{
read_fastas_from_fileinfo_backup(
  fileinfo_df,
  dir = "",
  write_csv = F,
  outdir = NA,
  sep = ","
)
}
\arguments{
\item{fileinfo_df}{data frame with columns:  plate, marker, sample,  replicate, filename, (optional: sample_type(mock/negative/real), habitat)}

\item{dir}{name of the directory with fasta file}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}
}
\description{
Read all fasta file in fileinfo_df data frame (plate	marker	sample	replicate	filename).
Dereplicate reads to ASVs.
Count the number of reads of each ASVs in each plate-marker-sample-replicate.
Returns a df ("asv", "plate", "marker", "sample", "replicate", "read_count").
}
