% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{read_fastas_from_sortedinfo}
\alias{read_fastas_from_sortedinfo}
\title{Read all fasta files in sortedinfo file to a data frame}
\usage{
read_fastas_from_sortedinfo(
  sortedinfo,
  dir = "",
  outfile = "",
  sep = ",",
  asv_list = "",
  updated_asv_list = "",
  quiet = T
)
}
\arguments{
\item{sortedinfo}{data frame or csv file with columns:  sample,  replicate, fasta, (optional: sample_type(mock/negative/real), habitat)}

\item{dir}{name of the directory with fasta files}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{sep}{separator in csv files; default: ","}

\item{asv_list}{name of the file, containing asvs and asv_ids from earlier analyses. Optional. It is used to homogenize asv_ids between different data sets}

\item{updated_asv_list}{name of the output file, containing the updated ASVs. Optional.}

\item{quiet}{\link{T/F}; TRUE by default print only warnings and error to STDOUT}
}
\description{
Read all fasta file in sortedinfo_df data frame (sample	replicate	file name).
Dereplicate reads to ASVs.
Count the number of reads of each ASVs in each sample-replicate.
Add a unique asv_id to each asv. If asv_list is given (containing earlier asv and asv_ids), uses already existing asv_id when possible, and adds new uniq asv_id, if asv is new
If updated_asv_list is given, writes an updated file containing all asv and asv_id is asv_list and sortedinfo_df
Returns a df ("asv_id", "sample", "replicate", "read_count", asv)
}
