% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{remove_leading_ones}
\alias{remove_leading_ones}
\title{Remove leading ones from a vector}
\usage{
remove_leading_ones(row)
}
\arguments{
\item{row}{Vector of taxids.}
}
\value{
Vector of taxids with leading \code{1} values removed and replaced by \code{NA} at the end.
}
\description{
Remove all leading \code{1} values from a vector, shift remaining values to the left,
and pad the end with \code{NA} to preserve the original length.
}
\examples{
\dontrun{
remove_leading_ones(row)
}

}
