% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_graph.R
\name{renkonen_all}
\alias{renkonen_all}
\title{renkonen_all}
\usage{
renkonen_all(read_count_df)
}
\arguments{
\item{read_count_df}{data frame with asv, sample, replicate, and read_count columns}
}
\description{
Calculate the Renkonen distance among all pairs of sample-replicates
Returns a data frame with the following columns: sample1,sample2,replicate1,replicate2,renkonen_d,sample_comp (within, if sample1 and sample2 are identical, between otherwise)
}
