% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{renkonen_dist}
\alias{renkonen_dist}
\title{Calculate Renkonen distance}
\usage{
renkonen_dist(df1, df2)
}
\arguments{
\item{df1, df2}{Data frames with \code{asv} and \code{read_count} columns.}
}
\value{
Numeric value between 0 and 1 representing the Renkonen distance
between the two samples.
}
\description{
Compute the Renkonen distance between two ASV abundance profiles.

The Renkonen distance is based on the absolute differences in relative
abundances of shared ASVs and ranges from 0 (identical compositions) to 1
(completely different compositions).
}
\examples{
\dontrun{
df1 <- read_count_df \%>\%
  dplyr::filter(sample == "tpos1") \%>\%
  dplyr::group_by(asv) \%>\%
  dplyr::summarise(read_count = sum(read_count))

df2 <- read_count_df \%>\%
  dplyr::filter(sample == "tnegtag") \%>\%
  dplyr::group_by(asv) \%>\%
  dplyr::summarise(read_count = sum(read_count))

renkonen_dist(df1, df2)
}

}
