% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Reverse complement a sequence}
\usage{
reverse_complement(sequence)
}
\arguments{
\item{sequence}{Character string specifying a DNA sequence.}
}
\value{
Character string. The reverse complement of the input sequence.
}
\description{
Compute the reverse complement of a DNA sequence. IUPAC ambiguity codes are accepted.
}
\examples{
reverse_complement(sequence = "AAATGCRC")

}
