% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{run_clustersize}
\alias{run_clustersize}
\title{run_clustersize}
\usage{
run_clustersize(
  read_count_df,
  id = 0.97,
  vsearch_path = "vsearch",
  num_threads = 0,
  quiet = TRUE
)
}
\arguments{
\item{read_count_df}{Data frame with the following variables:
asv_id, sample, replicate (optional), read_count, asv.}

\item{id}{Real; Minimum identity between asv and centroid.}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{num_threads}{Positive integer: Number of CPUs.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
read_count data frame, with ASV of the same cluster, sample, (replicate)
grouped to the same line
}
\description{
Run vsearch clustersize on all ASV of the input data frame and pool
ASV of the same cluster under the asv_id of the centroid.
}
\examples{
\dontrun{
clustered_df <- run_clustersize(read_count_df, id=0.97)
}
}
