% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{run_system2}
\alias{run_system2}
\title{Run a command using system2}
\usage{
run_system2(path, args, quiet = FALSE)
}
\arguments{
\item{path}{Character string specifying the path to the executable.}

\item{args}{Character vector of arguments passed to the command.}

\item{quiet}{Logical. If \code{TRUE}, suppress informational messages and
only display warnings or errors.}
}
\description{
Execute a system command using \code{system2}, with optional control over
the verbosity of the output.
}
\examples{
\dontrun{
run_system2(path = "ls", args = c("-all"), quiet = FALSE)
}

}
