% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{select_sequences}
\alias{select_sequences}
\title{select_sequences}
\usage{
select_sequences(file, n = 100, randseed = 0)
}
\arguments{
\item{file}{input fasta file; can be  uncompressed or compressed in gz format (zip files are not supported)}

\item{n}{number of sequences to be selected}

\item{randseed}{positive integer; seed for random sampling; 0 by default means to use a pseudo-random seed, a given non-zero seed produce always the same result}
}
\description{
Select n random sequences from the input file and returns a data frame with two columns, headers and sequences.
}
