% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_treatment.R
\name{SortReads}
\alias{SortReads}
\title{SortReads}
\usage{
SortReads(
  fastainfo_df,
  fastadir,
  outdir = "",
  cutadapt_path = "",
  vsearch_path = "",
  check_reverse = F,
  tag_to_end = T,
  primer_to_end = T,
  cutadapt_error_rate = 0.1,
  cutadapt_minimum_length = 50,
  cutadapt_maximum_length = 500,
  sep = ",",
  compress = 0
)
}
\arguments{
\item{fastainfo_df}{data frame with column: tag_fw,primer_fw,tag_rv,primer_rv,plate,marker,sample,sample_type(mock/negative/real),habitat(optional),replicate,fasta}

\item{fastadir}{directory with input fasta files (listed in fastainfo_df$fasta)}

\item{outdir}{output directory}

\item{cutadapt_path}{path to cutadapt executables}

\item{vsearch_path}{path to vsearch executables}

\item{check_reverse}{\link{T/F} if TRUE, ckeck the reverse comlementary sequences of the input fasta as well; default: F}

\item{tag_to_end}{tags are at the extremity of the reads (starting at the first base); default: T}

\item{primer_to_end}{primers follow directly the tags (no heterogeneity spacer); default: T}

\item{cutadapt_error_rate}{maximum proportion of errors between primers and reads (for tags, exact match is required); default: 0.1}

\item{cutadapt_minimum_length}{minimum length of the trimmed sequence; default: 50}

\item{cutadapt_maximum_length}{maximum length of the merged sequence; default: 500}

\item{sep}{separator in input and output csv files; default: ","}

\item{compress}{\link{gz/zip}; compress output to gz or zip files. Do not compress by default;}
}
\description{
Demultiplex each input fasta file using the tag combinations at the extremities of the merged reads.
Trim primers from demultiplexed reads.
Output fasta files can be compressed if compress option is used.
The output fileinfo.csv file is similar to the fastainfo file, but the but do not have tag and primer columns.
Returns data frame of corresponding to the output csv file.
}
