% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{SortReads_no_reverse}
\alias{SortReads_no_reverse}
\title{SortReads_no_reverse}
\usage{
SortReads_no_reverse(
  fastainfo_df,
  fastadir,
  outdir = "",
  cutadapt_path = "",
  tag_to_end = T,
  primer_to_end = T,
  cutadapt_error_rate = 0.1,
  cutadapt_minimum_length = 50,
  cutadapt_maximum_length = 500,
  sep = ",",
  compress = F
)
}
\arguments{
\item{fastainfo_df}{data frame with column: tag_fw,primer_fw,tag_rv,primer_rv,sample,sample_type(mock/negative/real),habitat(optional),replicate,fasta}

\item{fastadir}{directory with input fasta files (listed in fastainfo_df$fasta)}

\item{outdir}{output directory}

\item{cutadapt_path}{path to cutadapt executables}

\item{tag_to_end}{tags are at the extremity of the reads (starting at the first base); default: T}

\item{primer_to_end}{primers follow directly the tags (no heterogeneity spacer); default: T}

\item{cutadapt_error_rate}{maximum proportion of errors between primers and reads (for tags, exact match is required); default: 0.1}

\item{cutadapt_minimum_length}{minimum length of the trimmed sequence; default: 50}

\item{cutadapt_maximum_length}{maximum length of the merged sequence; default: 500}

\item{sep}{separator in input and output csv files; default: ","}

\item{compress}{\link{T/F}; compress output to gzip files.}
}
\description{
Same as SortReads, but do not check the reverse complement of the sequences.
Demultiplex each input fasta file using the tag combinations at the extremities of the merged reads.
Trim primers from demultiplexed reads.
Output fasta files can be compressed if compress option is used.
The output sortedinfo.csv file is similar to the fastainfo file, but the but do not have tag and primer columns.
Returns data frame of corresponding to the output csv file.
}
