% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{SummarizeBy}
\alias{SummarizeBy}
\title{SummarizeBy}
\usage{
SummarizeBy(dir, feature, grouped_by, outfile = "", sep = ",")
}
\arguments{
\item{dir}{Character string: directory containing the output files
of intermediate filtering steps. File names should start with a number
followed by an underscore (e.g. 5_LFNsampleReplicate.csv).}

\item{feature}{Character string with the following possible values:
"asv_id", "asv", "sample", "replicate", "read_count".}

\item{grouped_by}{Character string with the following possible values:
"asv_id", "asv", "sample", "replicate". Group data by this variable.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Data frame with input file names in columns, \code{grouped_by} values in
lines, and the number of features or the read counts in the cells.
}
\description{
Summarizes the output of each intermediate filtering steps.
Takes all files in the directory, that start with a number.
For each file, groups the lines by \code{grouped_by} variable,
and gets the number of distinct values of \code{feature} for
each group, or if \code{feature} is \code{read_count}, gets the total
number of reads of each group.
}
\examples{
\dontrun{
SummarizeBy(dir="vtamR_test/out_mfzr", feature="asv", grouped_by="sample")
SummarizeBy(dir="vtamR_test/out_mfzr", feature="read_count", grouped_by="sample")
}
}
