% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{Swarm}
\alias{Swarm}
\title{Run swarm}
\usage{
Swarm(
  read_count,
  outfile = "",
  swarm_path = "swarm",
  num_threads = 0,
  swarm_d = 1,
  fastidious = T,
  sep = ",",
  by_sample = T,
  quiet = T
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv, sample, replicate (optional), read_count.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{swarm_path}{Character string: path to swarm executables.}

\item{num_threads}{Positive integer: Number of CPUs.}

\item{swarm_d}{Positive integer: d parameter for swarm.
Maximum number of differences allowed between two ASVs,
meaning that two ASVs will be grouped if they have d (or less) differences.}

\item{fastidious}{logical: when working with d = 1, perform a second
clustering pass to reduce the number of small clusters.}

\item{sep}{Field separator character in input and output csv files.}

\item{by_sample}{logical: run swarm separately for each sample.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
Data frame with the same structure as the input, but ASVs of
the same cluster pooled to one row.
}
\description{
Runs swarm \url{https://github.com/torognes/swarm}
on input read_count data frame or csv file.
Pools variants of the same cluster and sums read counts of the
underlying ASVs.
}
\details{
Swarm can be run sample by sample (by_sample=T) or for the whole data set
in one go.
}
\examples{
\dontrun{
Swarm(read_count=read_count_df, swarm_path=swarm_path, num_threads=4, by_sample=T)
}
}
