% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxassign_functions.R
\name{TaxAssign}
\alias{TaxAssign}
\title{TaxAssign}
\usage{
TaxAssign(
  df,
  ltg_params_df = "",
  taxonomy = "",
  blast_db = "",
  blast_path = "",
  outdir = "",
  num_threads = 8
)
}
\arguments{
\item{df}{a data frame contining and asv column}

\item{ltg_params_df}{data frame with a list of ercentage of identity values (pid) and associated parameters (pcov,phit,taxn,seqn,refres,ltgres)}

\item{taxonomy}{file containing the following columns: tax_id,parent_tax_id,rank,name_txt,old_tax_id(has been mered to tax_id),taxlevel (8: species, 7: genus, 6: family, 5: order, 4: class, 3: phylum, 2: kingdom, 1: superkingdom, 0: root)}

\item{blast_db}{BLAST database}

\item{blast_path}{path to BLAST executable}

\item{outdir}{name of the output directory}

\item{taxres_df}{data frame with the following columns: asv,ltg_taxid,ltg_name,ltg_rank,ltg_rank_index,superkingdom_taxid,
superkingdom,kingdom_taxid,kingdom,phylum_taxid,phylum,class_taxid,class,order_taxid,order,family_taxid,family,genus_taxid,genus,species_taxid,species,pid,
pcov,phit,taxn,seqn,refres,ltgres}
}
\description{
Find LTG for each asv in the input dataframe
}
