% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_assign_fun.R
\name{TaxAssignRDP}
\alias{TaxAssignRDP}
\title{TaxAssignRDP}
\usage{
TaxAssignRDP(
  asv,
  dir = NULL,
  max_memory = 1,
  confidence = 0.8,
  rm_chloroplast = TRUE,
  outfile = "",
  quiet = TRUE,
  sep = ","
)
}
\arguments{
\item{asv}{Data frame or csv file containing an asv and asv_id columns.}

\item{dir}{Directory with the data for an existing classifier
(created with trainRDP()). If NULL, default classifier trained with
the data shipped with RDP is used. See RDP documentation (?rdp)}

\item{max_memory}{Positive integer: RAM available for rRDP in Gb.}

\item{confidence}{Numeric; Value between 0-1. Minimum confidence value, to accept
the taxonomic assignment}

\item{rm_chloroplast}{Logical; If true, reset the lineage to NA if class is
Chloroplast}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Data frame with the following columns:
asv_id,domain,kingdom,phylum,class,order,family,genus
}
\description{
Assign 16S bacteria/Archaea sequences to taxa using the rRDP package
}
\details{
This function needs the rRDP and rRDPData packages.

By default, the 16S Bacterial/Archaea RDP database is used from rRDPData.
(http://sourceforge.net/projects/rdp-classifier/).
Otherwise the directory name containing the a trained classifier object created
with trainRDP(), can be provided.
}
\examples{
\dontrun{
taxa <- TaxAssignRDP(asv=read_count_df, confidence=0.7, max_memory=8, rm_chloroplast=FALSE)
}
}
