% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_vtamR.R
\name{test_make_known_occurrences}
\alias{test_make_known_occurrences}
\title{test_make_known_occurrences}
\usage{
test_make_known_occurrences(test_dir = "~/vtamR/vtamR_test/", sep = ";")
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{sep}{separator in csv files}
}
\description{
Run make_known_occurrences on a test file and compare the output to expected results
}
