% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_test.R
\name{test_MakeKnownOccurrences}
\alias{test_MakeKnownOccurrences}
\title{test_MakeKnownOccurrences}
\usage{
test_MakeKnownOccurrences(test_dir = "vtamR_test/", sep = ",")
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{sep}{separator in csv files}
}
\description{
Run MakeKnownOccurrences on a test file and compare the output to expected results
}
