% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_vtamR.R
\name{test_merge_and_sortreads}
\alias{test_merge_and_sortreads}
\title{test_merge_and_sortreads}
\usage{
test_merge_and_sortreads(
  vtam_dir = vtam_dir,
  vsearch_path = "",
  cutadapt_path = ""
)
}
\arguments{
\item{vtam_dir}{directory of vtamR}

\item{vsearch_path}{path to vsearch executables}

\item{cutadapt_path}{path to cutadapt executables}
}
\description{
Compare the Merge and SortReads (using default values of vtam) of vtamR to precoputed file obtained by vtam
}
