% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_test.R
\name{Test_MergeSortReads}
\alias{Test_MergeSortReads}
\title{Test_MergeSortReads}
\usage{
Test_MergeSortReads(
  test_dir = "vtamR_test/",
  vsearch_path = "vsearch",
  cutadapt_path = "cutadapt",
  delete_tmp = T,
  sep = ",",
  quiet = T
)
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{vsearch_path}{path to vsearch executables}

\item{cutadapt_path}{path to cutadapt executables}

\item{delete_tmp}{Boolean.  Delete output folder}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\description{
Compare the Merge and SortReads output (using default values of vtam) of vtamR
to pre-computed files obtained by vtam
}
