% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_vtamR.R
\name{test_taxassign}
\alias{test_taxassign}
\title{test_taxassign}
\usage{
test_taxassign(
  test_dir = "~/vtamR/vtamR_test/",
  sep = sep,
  blast_path = blast_path,
  blast_db = blast_db,
  taxonomy = taxonomy
)
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{sep}{separator in csv files}

\item{blast_path}{path to BLAST executable}

\item{blast_db}{BLAST database}

\item{taxonomy}{file containing the following columns: tax_id,parent_tax_id,rank,name_txt,old_tax_id(has been mered to tax_id),taxlevel (8: species, 7: genus, 6: family, 5: order, 4: class, 3: phylum, 2: kingdom, 1: superkingdom, 0: root)}
}
\description{
Run TaxAssign on a test file and compare the output to mkLTG results
}
