% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{TrimPrimer}
\alias{TrimPrimer}
\title{TrimPrimer}
\usage{
TrimPrimer(
  fastainfo,
  fasta_dir = "",
  outdir = "",
  compress = F,
  cutadapt_path = "",
  vsearch_path = "",
  check_reverse = F,
  primer_to_end = T,
  cutadapt_error_rate = 0.1,
  cutadapt_minimum_length = 50,
  cutadapt_maximum_length = 500,
  quiet = T
)
}
\arguments{
\item{fastainfo}{csv file of data frame with the following columns: tag_fw,primer_fw,tag_rv,primer_rv,sample,sample_type,habitat,replicate,fasta,read_count}

\item{fasta_dir}{name of the directory with the fasta files to be trimmed}

\item{outdir}{name of the directory for the trimmed fasta files to be trimmed}

\item{compress}{\link{T/F}; compress output to gzip format; Default=F}

\item{cutadapt_path}{path to cutadapt executable}

\item{vsearch_path}{path to vsearch executable}

\item{check_reverse}{\link{T/F} if TRUE, check the reverse complementary sequences of the input fasta as well; default: F}

\item{primer_to_end}{primers follow directly the tags (no heterogeneity spacer); default: T}

\item{cutadapt_error_rate}{maximum proportion of errors between primers and reads (for tags, exact match is required); default: 0.1}

\item{cutadapt_minimum_length}{minimum length of the trimmed sequence; default: 50}

\item{cutadapt_maximum_length}{maximum length of the merged sequence; default: 500}

\item{quiet}{\link{T/F}; TRUE by default print only warnings and error to STDOUT}
}
\description{
Trim primers from each fasta file in input fastainfo csv file of data frame. Keep only trimmed reads.
Can check both orientation
Count the number of reads in the output files
}
