% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{update_taxids}
\alias{update_taxids}
\title{Update obsolete taxids}
\usage{
update_taxids(df, old_taxid)
}
\arguments{
\item{df}{Data frame with the following columns: \code{qseqid}, \code{pident}, \code{qcovhsp}, \code{staxids}.}

\item{old_taxid}{Data frame with columns \code{tax_id} and \code{old_tax_id} used to map
deprecated taxids to valid ones.}
}
\value{
Data frame with updated \code{staxids}, where obsolete taxids have been replaced
by current valid taxids.
}
\description{
Replace outdated NCBI taxonomy IDs with their current valid equivalents.
This step ensures that merged or deprecated taxids are mapped to updated identifiers.
}
\examples{
\dontrun{
update_taxids(df = blastout_df, old_taxid = old_taxid_df)
}

}
