% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{write_asvtable}
\alias{write_asvtable}
\title{Write ASV table}
\usage{
write_asvtable(
  read_count_samples_df,
  outfile,
  asv_tax = NULL,
  fileinfo = "",
  add_empty_samples = F,
  add_sums_by_sample = F,
  add_sums_by_asv = F,
  add_expected_asv = F,
  mock_composition = "",
  sep = ","
)
}
\arguments{
\item{read_count_samples_df}{data frame with the following variables: asv_id, sample, read_count, asv}

\item{outfile}{name of the output csv file including path}

\item{asv_tax}{optional: data frame with taxonomic assignments}

\item{fileinfo}{tsv file with columns: sample, sample_type, habitat, replicate, file; only necessary if add_empty_samples==T or add_expected_asv==T}

\item{add_empty_samples}{\link{T/F} add a column for each samples in the original data set, even if they do not have reads after filtering}

\item{add_sums_by_sample}{\link{T/F} add a line with the total number of reads in each sample, and another with the number of ASVs in each sample}

\item{add_sums_by_asv}{\link{T/F} add a column with the total number of reads for each ASV, and another with the number of samples, where the ASV is present}

\item{add_expected_asv}{\link{T/F} add a column for each mock sample where keep and tolerate ASVs are marked}

\item{mock_composition}{csv file with the following columns: sample,action,asv; action can take the following values: keep/tolerate; only necessary if add_expected_asv==T}

\item{sep}{separator used in the I/O csv files}
}
\description{
Samples in columns, ASVs in lines
}
