% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{write_cutadapt_adapter_fastq}
\alias{write_cutadapt_adapter_fastq}
\title{Make a FASTA file with forward or reverse adapters}
\usage{
write_cutadapt_adapter_fastq(
  fastqinfo_df,
  fastq_file,
  outdir,
  tag_to_end = TRUE
)
}
\arguments{
\item{fastqinfo_df}{Data frame with columns: \code{tag_fw}, \code{tag_rv}, \code{fastq_fw} and \code{fastq_rv}.}

\item{outdir}{Character string specifying the output directory.}

\item{tag_to_end}{Logical. If \code{TRUE}, tags are assumed to be located at the
extremities of reads.}

\item{fastq}{Character string specifying the \code{fastq_fw} file to be demultiplexed
(must be present in the \code{fastq_fw} column of \code{fastqinfo_df}).}
}
\value{
Vector with the output fasta files, or \code{NA} if all tags
are \code{NA} in \code{fastqinfo_df} for the given \code{fastq} file.
}
\description{
Create two FASTA files containing forward or reverse tags
formatted for \code{cutadapt}.
This file is used by \code{demultiplex_and_trim_fastq} to demultiplex input FASTQ files pairs.
}
\examples{
\dontrun{
write_cutadapt_adapter_fastq(
  fastqinfo_df = fastqinfo_df, 
  fastq_file = "fastq_file", 
  tag_to_end = FALSE, 
  outdir = "data/out"
)
}

}
