% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{write_df_to_fasta}
\alias{write_df_to_fasta}
\title{write_df_to_fasta}
\usage{
write_df_to_fasta(df, out, compress = F)
}
\arguments{
\item{df}{input data frame with the following columns: header, sequence}

\item{out}{out fasta filename; modified according to compress}

\item{compress}{\link{T/F} is TRUE, the output file will be compressed}
}
\description{
Writes input data frame to a fasta file. Can produce gz compressed file or uncompressed files.
The output filename extention is corrected according to the compression
}
