% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{write_fasta}
\alias{write_fasta}
\title{write_fasta}
\usage{
write_fasta(sequences, filename, seq_as_id = F)
}
\arguments{
\item{sequences}{list of sequences}

\item{filename}{output file name, including path}

\item{seq_as_id}{T/F; It TRUE, use sequences as seqID (FALSE by default)}
}
\description{
Write a fasta file using the sequences as ID
}
