% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{write_fasta_df}
\alias{write_fasta_df}
\title{write_fasta_df}
\usage{
write_fasta_df(df, outfile, read_count = FALSE)
}
\arguments{
\item{df}{Data frame with asv_id, asv and read_count (optional) columns}

\item{outfile}{character string: output fasta file name}

\item{read_count}{Boolean; If TRUE  definition line is >label;size=### format}
}
\value{
outfile fasta file with definition lines in >label;size=### format or >label format
}
\description{
Write a fasta file with definition lines in >label;size=### format
}
