% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{write_fasta_rc}
\alias{write_fasta_rc}
\title{write_fasta_rc}
\usage{
write_fasta_rc(df, outfile)
}
\arguments{
\item{df}{Data frame with asv_id, read_count and asv columns}

\item{outfile}{haracter string: output fasta file name}
}
\value{
outfile fasta file with definition lines in >label;size=### format
}
\description{
Write a fasta file with definition lines in >label;size=### format
}
