% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{write_fasta_seq_as_id}
\alias{write_fasta_seq_as_id}
\title{write_fasta_seq_as_id}
\usage{
write_fasta_seq_as_id(sequences, filename)
}
\arguments{
\item{sequences}{list of sequences}

\item{filename}{output file name, including path}
}
\description{
Write a fasta file using the sequences as ID
}
