% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_helper_functions.R
\name{write_log}
\alias{write_log}
\title{Complete the log and write it to a CSV file}
\usage{
write_log(log, file = NULL, sep = ",")
}
\arguments{
\item{log}{A data frame returned by \code{collect_log()}, containing the calling
function name, argument names, argument values, and the start time.}

\item{file}{Character string specifying the path to the CSV log file. If
\code{NULL}, no log file is written.}

\item{sep}{Character string specifying the field separator used in input and
output CSV files.}
}
\value{
The completed log data frame, returned invisibly.
}
\description{
Completes the log data frame returned by \code{collect_log()} by adding the end
time and execution time, then appends the log to a CSV file.
}
\keyword{internal}
