% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{add_ids}
\alias{add_ids}
\title{add_ids}
\usage{
add_ids(read_count_df, asv_list = asv_list, updated_asv_list = "", sep = ",")
}
\arguments{
\item{read_count_df}{data frame with columns:  asv, sample,  replicate, read_count}

\item{asv_list}{name of the file, containing asvs and asv_ids from earlier analyses. Optional. It is used to homogenize asv_ids between different data sets}

\item{updated_asv_list}{name of the output file, containing the updated ASVs. Optional.}

\item{sep}{separator in csv files; default: ","}
}
\description{
Add asv_ids to a data frame that have an asv column. Can take into account already existing asv and asv_id (from earlier dataset)
If update_asv_list, the input asv_list id completed by new asvs and asv_ids
Returns a data frame with as asv_id column on top of the input.
}
