% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{add_ids}
\alias{add_ids}
\title{add_ids}
\usage{
add_ids(
  read_count_df,
  asv_list = asv_list,
  updated_asv_list = "",
  sep = ",",
  quiet = T
)
}
\arguments{
\item{read_count_df}{data frame with columns:  asv, sample,  replicate, read_count}

\item{asv_list}{name of the file, containing asvs and asv_ids from earlier analyses. Optional. It is used to homogenize asv_ids between different data sets}

\item{updated_asv_list}{name of the output file, containing the updated ASVs. Optional.}

\item{sep}{separator in csv files; default: ","}

\item{quiet}{\link{T/F}; TRUE by default print only warnings and error to STDOUT}
}
\description{
Add asv_ids to a data frame that have an asv column. Can take into account already existing asv and asv_id (from earlier data set) present in asv_list file
If updated_asv_list file name is given, the input asv_list is completed by new asvs and asv_ids
Returns a data frame with asv_id column added to the input data frame.
}
