% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxassign_functions.R
\name{adjust_ltgres}
\alias{adjust_ltgres}
\title{adjust_ltgres}
\usage{
adjust_ltgres(taxres_df, tax_df)
}
\arguments{
\item{taxres_df}{data frame with the following columns: asv,ltg_taxid,ltg_name,ltg_rank,ltg_rank_index,superkingdom_taxid,
superkingdom,kingdom_taxid,kingdom,phylum_taxid,phylum,class_taxid,class,order_taxid,order,family_taxid,family,genus_taxid,genus,species_taxid,species,pid,
pcov,phit,taxn,seqn,refres,ltgres}

\item{tax_df}{data frame with the following columns: tax_id, parent_tax_id, rank, name_txt, taxlevel (8: species, 7: genus, 6: family, 5: order, 4: class, 3: phylum, 2: kingdom, 1: superkingdom, 0: root)}
}
\description{
If the ltg has a higher resolution than the ltgres parameter, adjust the ltg and stop lineage at ltgres level; Returns the adjusted data frame
}
