% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_graph.R
\name{Barplot_ReadCountBySample}
\alias{Barplot_ReadCountBySample}
\title{Barplot_ReadCountBySample}
\usage{
Barplot_ReadCountBySample(
  read_count_df,
  sample_types = "",
  sample_replicate = T,
  sep = ",",
  x_axis_label_size = 6
)
}
\arguments{
\item{read_count_df}{data frame with sample, read_count and replicate (optional) columns}

\item{sample_types}{file with sample and sample_type (real/mock/negative) columns}

\item{sample_replicate}{Boolean. If TRUE barplot is made by sample-replicates,
by sample otherwise}

\item{sep}{separator used in csv files}

\item{x_axis_label_size}{size of labels in x axis}
}
\description{
Create barplot with the number of reads in each sample or sample-replicate;
If information is given on sample types, bars are colored in function of them
}
