% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{calculate_renkonen_dist}
\alias{calculate_renkonen_dist}
\title{Calculate Renkonen Distance}
\usage{
calculate_renkonen_dist(df1, df2)
}
\arguments{
\item{df1, df2}{Data frames with asv and read_count columns.}
}
\value{
Renkonen distance (Real; 0-1) between the two date frames.
}
\description{
Calculate renkonen distance between two data frames with ASVs and read counts.
}
\examples{
df1 <- read_count_df \%>\%
  filter(sample=="tpos1") \%>\%
  group_by(asv) \%>\%
  summarize(read_count = sum(read_count))
df2 <- read_count_df \%>\%
  filter(sample=="tnegtag") \%>\%
  group_by(asv) \%>\%
  summarize(read_count = sum(read_count))
calculate_renkonen_dist(df1, df2)
}
