% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{check_dir}
\alias{check_dir}
\title{Check directory}
\usage{
check_dir(dir)
}
\arguments{
\item{dir}{directory name}
}
\description{
Create dir if does not exists.
Add slash to the end of the directory name.
}
