% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{check_dir}
\alias{check_dir}
\title{Check directory}
\usage{
check_dir(dir)
}
\arguments{
\item{dir}{Character string naming a directory.}
}
\description{
Check if directory exists, and create it if not.
}
\examples{
\dontrun{
check_dir(dir="data")
}
}
