% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{check_file_exists}
\alias{check_file_exists}
\title{check_file_exists}
\usage{
check_file_exists(dir = dir, file_list = file_list)
}
\arguments{
\item{dir}{name of the directory containig the input files}

\item{file_list}{vector of file names}
}
\description{
Check if all files in the input vector exist
}
