% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{check_file_info}
\alias{check_file_info}
\title{Check File formats}
\usage{
check_file_info(
  file,
  dir = "",
  file_type = "fastqinfo",
  sep = ",",
  quiet = FALSE
)
}
\arguments{
\item{file}{Character string: name of the input file to be checked or
data frame with the content of the input file.}

\item{dir}{Character string: Name of the directory containing the files
in \code{fastq_fw}, \code{fastq_rv} and \code{fasta} columns in the input file.}

\item{file_type}{Character string with the following possible values:
"fastqinfo", "fastainfo", "sampleinfo", "mock_composition", "known_occurrences",
"read_count", "read_count_sample", "asv_list".}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
Error message if problem with the files and stop the script.
}
\description{
Check the format and coherence of different file types:
\itemize{
\item Check if all obligatory columns are present (all file_type)
\item Check if all sample names are alpha-numerical
\item Check if asv, tag_fw, tag_rv primer_fw, primer_rv have only IUPAC codes
\item Check if read_count is numerical
\item Check sample type, habitat homogeneity across replicates
(fastqinfo, fastainfo, sampleinfo)
\item Check if fastq file pairs are coherent (e.g. 1 to 1 relation; fastqinfo)
\item Check if tag combinations are unique within a file(pair)s (fastqinfo, fastainfo)
\item Check action (mock_composition, known_occurrences)
\item Check if 1 to 1 relation between asv_id ad asv (read_count,read_count_sample,asv_list)
}
}
\examples{
\dontrun{
check_file_info(file="input/sampleinfo.csv", dir="fasta", file_type="sampleinfo")
check_file_info(file=sampleinfo_df, dir="fasta", file_type="read_count_sample")
}
}
