% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{check_heading}
\alias{check_heading}
\title{check_heading}
\usage{
check_heading(list1, list2, file = "")
}
\arguments{
\item{list1}{vector}

\item{list2}{vector}

\item{file}{file name to write to error message. Optional}

\item{outfile}{name of the output file; optional, it is not given results are returned in a data frame, but no file is written}
}
\description{
Compare the 2 input vectors, and returns elements of list1 that are not on list2
}
