% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{classify_control_occurrences}
\alias{classify_control_occurrences}
\title{Classify control occurrences and compute performance metrics}
\usage{
classify_control_occurrences(
  read_count,
  sampleinfo,
  mock_composition,
  sep = ",",
  known_occurrences = NULL,
  false_negatives = NULL,
  performance_metrics = NULL,
  habitat_proportion = 0.5,
  quiet = TRUE,
  log_file = NULL
)
}
\arguments{
\item{read_count}{Data frame or CSV file with the following variables:
\code{asv_id}, \code{sample}, \code{replicate} (optional), \code{read_count}, \code{asv}.}

\item{sampleinfo}{Data frame or CSV file with columns:
\code{sample}, \code{sample_type} (\code{mock}/\code{negative}/\code{real}), \code{habitat}, \code{replicate} (optional),
and optionally \code{file}.}

\item{mock_composition}{Data frame or CSV file with columns:
\code{sample}, \code{action} (\code{keep}/\code{tolerate}), \code{asv}.}

\item{sep}{Field separator character used in input and output CSV files.}

\item{known_occurrences}{Character string specifying output CSV file for known occurrences
(expected occurrences in mocks and false positives). If NULL, no file is written.}

\item{false_negatives}{Character string specifying output CSV file for false negatives.
If NULL, no file is written.}

\item{performance_metrics}{Character string specifying output CSV file for performance metrics.
If NULL, no file is written.}

\item{habitat_proportion}{Numeric between 0 and 1. For each ASV, if the proportion of reads
within a habitat is below this threshold, it is considered an artifact in all samples
of that habitat.}

\item{quiet}{Logical. If \code{TRUE}, suppress informational messages and only show warnings or errors.}

\item{log_file}{Character string specifying the path to the CSV log file.
If \code{NULL}, no log file is written.}
}
\value{
A list containing:
\itemize{
\item \code{known_occurrences_df}: sample, action, asv_id, asv
\item \code{false_negatives_df}: sample, action, asv, asv_id
\item \code{performance_metrics_df}: TP, FP, FN, Precision, Sensitivity
}
}
\description{
Identify expected and unexpected occurrences in control samples and summarize
pipeline performance using true positives, false positives, and false negatives.
}
\details{
This function also computes key evaluation metrics including precision and sensitivity.
}
\examples{
\dontrun{
results <- classify_control_occurrences(
  read_count_samples_df,
  sampleinfo = sampleinfo_df,
  mock_composition = "data/mock_composition.csv",
  habitat_proportion = 0.7
)
known_occurrences_df <- results[[1]]
false_negatives_df <- results[[2]]
performance_metrics_df <- results[[3]]
}

}
