% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{cluster_asv}
\alias{cluster_asv}
\title{Cluster ASVs Using Swarm or VSEARCH}
\usage{
cluster_asv(
  read_count,
  group = TRUE,
  by_sample = FALSE,
  method = "swarm",
  path = "",
  num_threads = 0,
  swarm_d = 1,
  fastidious = T,
  identity = 0.97,
  outfile = "",
  sep = ",",
  quiet = T
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv, sample, replicate (optional), read_count.}

\item{group}{Logical; If TRUE, ASVs of the same cluster are pooled to one row.
if FALSE, input Df is simply completed by a cluser_id column.}

\item{by_sample}{Logical: run clustering separately for each sample.}

\item{method}{Character string: program to use for clustering: swarm or vsearch}

\item{path}{Character string: path to swarm or vsearch executables.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}

\item{swarm_d}{Positive integer: d parameter for swarm (if applicable).
Maximum number of differences allowed between two ASVs,
meaning that two ASVs will be grouped if they have d (or less) differences.}

\item{fastidious}{Logical: When clustering with swarm and working with d = 1,
perform a second clustering pass to reduce the number of small clusters.}

\item{identity}{Numeric. Value between 0 and 1: the identity threshold used for
clustering when the cluster_size algorithm of vsearch is used.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
read_count_df: same structure as the input, but ASVs of
the same cluster pooled to one row if group==TRUE, or rows kept as in the input
and an additional cluster_id column is added if group==FALSE.
}
\description{
This function can \strong{cluster} ASVs into mOTUs or perform \strong{denoising}.
When using Swarm with \code{d = 1}, the clustering result corresponds to denoising.

The function runs either Swarm or VSEARCH's \code{cluster_size} command on the ASVs
in the input data frame. Each ASV is assigned a cluster, and two output formats
are available, controlled by the \code{group} argument.
\itemize{
\item If \code{group = TRUE}, ASVs in the same cluster are aggregated into a single row.
In this case, the \code{asv_id} and \code{asv} columns contain the identifier and
sequence of the cluster's centroid, and \code{read_count} is summed across ASVs.
Sample and replicate information is retained.
\item If \code{group = FALSE}, the function returns the original data frame with an
additional column, \code{cluster_id}. Each row corresponds to one ASV.
}

Clustering can be performed on the entire data set or separately for each sample
(\code{by_sample} argument).

\strong{Note:} If \code{by_sample = TRUE} and \code{group = FALSE}, the same \code{asv_id} may be
assigned different \code{cluster_id}s in different samples.
}
\examples{
\dontrun{
read_count_df <- cluster_asv(read_count=read_count_df, group=TRUE,method="vsearch",
by_sample=TRUE, path=swarm_path, 
num_threads=4)
}
}
