% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{cluster_swarm}
\alias{cluster_swarm}
\title{Cluster all input ASV by swarm}
\usage{
cluster_swarm(
  read_count,
  swarm_d = 1,
  fastidious = FALSE,
  swarm_path = "swarm",
  num_threads = 0,
  outfile = "",
  sep = ",",
  quiet = TRUE
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, asv, read_count.}

\item{swarm_d}{Positive integer: d for Swarm.}

\item{fastidious}{logical: when working with d = 1, perform a second
clustering pass to reduce the number of small clusters.}

\item{swarm_path}{Character string: path to swarm executable.}

\item{num_threads}{Positive integer: Number of CPUs.}

\item{outfile}{Character string: name of the output csv file}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{Logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
Data frame with the following columns: asv_id, cluster_id
}
\description{
Cluster all input ASV by swarm and return a data frame with asv_id, cluster_id
columns
}
\examples{
\dontrun{
plot <- cluster_swarm(read_count_df, 
                     swarm_d=7, 
                     swarm_path="swarm",
                     num_threads=8)
}
}
