% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{concatenate_files}
\alias{concatenate_files}
\title{Concatenate contents of files with identical names}
\usage{
concatenate_files(dirs, outdir, pattern = "\\\\.", quiet = TRUE)
}
\arguments{
\item{dirs}{A character vector of input directories to search for files.}

\item{outdir}{A character string specifying the output directory where pooled
files will be written.}

\item{pattern}{A regular expression used to select filenames within the input
directories.}

\item{quiet}{Logical: if TRUE, suppress informational messages; only warnings and errors are shown.}
}
\value{
Invisibly data frame of files and directories
}
\description{
Reads files from multiple directories and concatenates the contents of files
sharing the same filename into a single output file.
}
\details{
For each unique filename matching \code{pattern}, all corresponding files found in
\code{dirs} are read in order and merged into a file of the same name in \code{outdir}.

This function works both on uncompressed and gz compressed files.
}
\examples{
\dontrun{
concatenate_files(
  dirs = c("run1", "run2", "run3"),
  outdir = "pooled",
  pattern = "\\\\.fastq$"
)
}

}
