% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{count_reads_file}
\alias{count_reads_file}
\title{count_reads_file}
\usage{
count_reads_file(file, file_type = "")
}
\arguments{
\item{file}{filename including path}

\item{file_type}{\link{fasta/fastq} if not specified, the number of lines is returned}
}
\description{
Count the number of sequences in a fasta or fastq file, or  he number of line in other files
Efficient in linux-like systems, but slow on Windows for large files
Can handle gz compressed and uncompressed files, but not zipped files
}
