% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_treatment.R
\name{count_seq}
\alias{count_seq}
\title{count_seq}
\usage{
count_seq(dir = "", file)
}
\arguments{
\item{dir}{directory that contains the input file}

\item{file}{fasta file, without it's path}
}
\description{
Count the number of sequences in the input fasta file. Input can be uncompressed, gz, or zip file
}
