% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{count_seq}
\alias{count_seq}
\title{Count sequences in fasta}
\usage{
count_seq(file)
}
\arguments{
\item{file}{Character string: input fasta file (including path).}
}
\value{
The number of sequences in the input file.
}
\description{
Count the number of sequences in the input fasta file.
}
\details{
Input can be uncompressed or gzip compressed file,
but other compression types are not supported.
For linux-like systems, uses the bash commands grep and wc and it is quick.
For other operating systems it uses count.fields, and it is slower.
}
\examples{
\dontrun{
n <- count_seq(file="data/test.fasta")
}
}
