% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{decompress_file}
\alias{decompress_file}
\title{Decompress a gzipped file}
\usage{
decompress_file(filename = "", remove_input = F)
}
\arguments{
\item{filename}{Character string: gzip compressed input file.}

\item{remove_input}{logical: Remove the input compressed file.}
}
\value{
Character string: output decompressed file.
}
\description{
Decompress the input gzipped file.
If remove_input is TRUE, deleted the compressed input file.
}
\examples{
\dontrun{
outfile <- decompress_file(filename="data/test.fasta.gz", remove_input=T)
}
}
