% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{demultiplex_fastq_pairs_strand_plus}
\alias{demultiplex_fastq_pairs_strand_plus}
\title{Demultiplex fastq file pairs and trim tags and primers (reverse strand not checked)}
\usage{
demultiplex_fastq_pairs_strand_plus(
  fastqinfo,
  fastq_dir,
  outdir,
  cutadapt_path = "cutadapt",
  num_threads = 0,
  tag_to_end = T,
  primer_to_end = T,
  cutadapt_error_rate = 0.1,
  sep = ",",
  compress = F,
  quiet = T
)
}
\arguments{
\item{fastqinfo}{Data frame or path to a CSV file with the following columns:
\code{tag_fw}, \code{primer_fw}, \code{tag_rv}, \code{primer_rv},
\code{sample}, \code{sample_type} (mock/negative/real),
\code{habitat} (optional), \code{replicate}, \code{fastq_fw},
\code{fastq_rv}}

\item{fastq_dir}{Character string specifying the directory containing input
FASTQ files (listed in the \code{fastq_fw} and \code{fastq_rv} columns of \code{fastqinfo}).}

\item{outdir}{Character string specifying the output directory.}

\item{cutadapt_path}{Character string specifying the path to the
\code{cutadapt} executable.}

\item{num_threads}{Positive integer specifying the number of CPU threads to
use. If \code{0}, all available CPUs are used.}

\item{tag_to_end}{Logical. If \code{TRUE}, tags are assumed to be located at the
extremities of reads (starting at the first base).}

\item{primer_to_end}{Logical. If \code{TRUE}, primers are assumed to follow
directly after tags (i.e., no heterogeneity spacer).}

\item{cutadapt_error_rate}{Numeric value between 0 and 1 specifying the
maximum allowed error rate between primers and reads (exact match is
required for tags).}

\item{sep}{Character string specifying the field separator used in input and
output CSV files.}

\item{compress}{Logical. If \code{TRUE}, compress output files using gzip.}

\item{quiet}{Logical. If \code{TRUE}, suppress informational messages and
only display warnings or errors.}
}
\value{
Data frame similar to the input \code{fastqinfo} file,
but contains the output fastq file names and read counts.
}
\description{
Same as \code{demultiplex_fastq_pairs}, but without checking the reverse-complement
of the sequences.

FASTQ file pairs are first demultiplexed by requiring a perfect match between
the tag sequence and the 5' end of the read. The resulting files are then
trimmed to remove primer sequences using less stringent matching parameters
(controlled by \code{cutadapt_error_rate}). A match between the primer and the 5'
end of the read is required for the read to be trimmed and retained. Matching
of the 3' primer is optional

Input files can be compressed or uncompressed. Output compression is
controlled by \code{compress}.
}
\examples{
\dontrun{
fastqinfo_df <- demultiplex_fastq_pairs_strand_plus(
  fastqinfo = fastqinfo_df,
  fastq_dir = "data/fastq",
  outdir = "data/fastq_demultiplexed",
  tag_to_end = TRUE,
  primer_to_end = TRUE,
  sep = ","
)
}

}
