% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{denoise_by_swarm}
\alias{denoise_by_swarm}
\title{Denoise data using SWARM}
\usage{
denoise_by_swarm(
  read_count,
  by_sample = FALSE,
  split_clusters = FALSE,
  min_abundance_ratio = 0.2,
  min_read_count = 10,
  swarm_path = "swarm",
  num_threads = 0,
  swarm_d = 1,
  fastidious = TRUE,
  outfile = "",
  sep = ",",
  quiet = TRUE
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv, sample, replicate (optional), read_count.}

\item{by_sample}{Logical. If \code{TRUE}, clustering is performed separately for
each sample.}

\item{split_clusters}{Logical. If \code{TRUE}, apply post-processing to split
clusters based on relatively abundant ASVs compared to the centroid.}

\item{min_abundance_ratio}{Numeric. Used when \code{split_clusters = TRUE}.
Minimum ratio of an ASV's read count relative to the centroid's read count
required for the ASV to define a new cluster.}

\item{min_read_count}{Numeric. Used when \code{split_clusters = TRUE}.
Minimum absolute read count required for an ASV to be considered for splitting.}

\item{swarm_path}{Character string: path to swarm executable.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}

\item{swarm_d}{Positive integer: d parameter for swarm (if applicable).
Maximum number of differences allowed between two ASVs,
meaning that two ASVs will be grouped if they have d (or less) differences.}

\item{fastidious}{Logical: When clustering with swarm and working with d = 1,
perform a second clustering pass to reduce the number of small clusters.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
A data frame with the same structure as the input, where ASVs
belonging to the same cluster are pooled (summed) within each
sample–replicate combination. If \code{split_clusters = TRUE}, clusters may be
further subdivided according to the splitting criteria.
}
\description{
This function performs \strong{denoising} using SWARM.
}
\details{
Clustering can be applied to the entire dataset or performed separately
for each sample using the \code{by_sample} argument.

By default, read counts of ASVs belonging to the same cluster are summed
within each sample–replicate combination. Optionally, an experimental
post-processing step (\code{split_clusters = TRUE}) can be applied to further
refine clusters. In this step, clusters are split based on the relative
abundance of ASVs compared to the centroid (see \code{?split_swarm_clusters}).
This option is only available when clustering is performed by sample
(\code{by_sample = TRUE}).
}
\examples{
\dontrun{
read_count_df <- denoise_by_swarm(
  read_count = read_count_df,
  by_sample = TRUE,
  split_clusters = TRUE
)
}
}
