% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{detect_false_negatives}
\alias{detect_false_negatives}
\title{Make Missing Occurrences}
\usage{
detect_false_negatives(
  read_count_samples,
  mock_composition,
  sep = ",",
  out = "",
  quiet = TRUE
)
}
\arguments{
\item{read_count_samples}{Data frame or csv file with the following variables:
asv, sample, read_count.}

\item{mock_composition}{Data frame or csv file with columns:
sample, action (keep/tolerate), asv.}

\item{sep}{Field separator character in input and output csv files.}

\item{out}{Character string: output file. If empty, no file is written.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
Data frame with the following columns:
sample, action, asv, asv_id
}
\description{
Prepare a data fram that list all expected occurrences that are missing
(False negatives).
}
\examples{
\dontrun{
detect_false_negatives(read_count_samples=read_count_samples_df, 
    mock_composition="data/mock_composition.csv"
    )
}
}
