% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_helper_functions.R
\name{download_osf}
\alias{download_osf}
\title{Download from OSF}
\usage{
download_osf(filename, url, dest_dir = ".", untar = TRUE, quiet = FALSE)
}
\arguments{
\item{filename}{character string specifying the name of the file to download.}

\item{url}{URL of the Zenodo record.}

\item{dest_dir}{path to the directory where the file should be downloaded.}

\item{untar}{logical: If TRUE, the archive will be extracted after download.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
full path to the directory where the file was downloaded.
}
\description{
Downloads file from \href{https://osf.io/}{OSF}
Extracts the .tar.gz archive, if necessary
}
\examples{
\dontrun{
download_db(filename = "COInr_for_vtam_2025_05_23_dbV5.tar.gz",
   url = "https://osf.io/download/jyhz6/",
   dest_dir = "~/vtamR/OSF",
   untar = TRUE,
   quiet = FALSE
   )
}
}
