% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_helper_functions.R
\name{download_zenodo}
\alias{download_zenodo}
\title{Download from Zenodo}
\usage{
download_zenodo(filename, url, dest_dir = ".", untar = TRUE, quiet = FALSE)
}
\arguments{
\item{filename}{character string specifying the name of the file to download.}

\item{url}{URL of the Zenodo record (copy from the browser address bar).}

\item{dest_dir}{path to the directory where the file should be downloaded.}

\item{untar}{logical: If TRUE, the archive will be extracted after download.}

\item{quiet}{logical: If TRUE, suppress informational messages and only}
}
\value{
full path to the directory where the file was downloaded.
show warnings or errors.
}
\description{
Downloads file from Zenodo.
Extracts the .tar.gz archive, if necessary
}
\examples{
\dontrun{
download_db(filename = "COInr_2025_05_23.tar.gz",
   url = "https://zenodo.org/records/15515860",
   dest_dir = "~/vtamR",
   untar = TRUE,
   quiet = FALSE
   )
}
}
