% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{expand_rows}
\alias{expand_rows}
\title{expand_rows}
\usage{
expand_rows(row)
}
\arguments{
\item{row}{a row of a data frame with the following columns: qseqid,pident,qcovhsp,staxids}
}
\description{
split he taxid column to different taxids in the input dataframe
Make one line for each taxid
Return a data frame with the following columns: "qseqid","pident","qcovhsp","staxids"
}
