% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{expand_rows}
\alias{expand_rows}
\title{Expand Rows}
\usage{
expand_rows(row)
}
\arguments{
\item{row}{Row of a data frame with the following columns:
qseqid,pident,qcovhsp,staxids}
}
\value{
Data frame with the following columns: qseqid,pident,qcovhsp,staxids
}
\description{
Split the taxid column to different taxids in the input data frame.
Make one line for each taxid.
}
\examples{
\dontrun{
expand_rows(row)
}
}
