% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{filter_asv_global}
\alias{filter_asv_global}
\title{Filter out ASVs with Low Total Read Count}
\usage{
filter_asv_global(read_count, cutoff = 10, outfile = NULL, sep = ",")
}
\arguments{
\item{read_count}{Data frame or path to a CSV file with the following columns:
\code{asv_id}, \code{sample}, \code{replicate}, \code{read_count}, \code{asv}.}

\item{cutoff}{Positive integer specifying the minimum total number of reads
required for an ASV to be retained. ASVs below this threshold are removed.}

\item{outfile}{Character string specifying the CSV file to write the output
data frame. If NULL, no file is written.}

\item{sep}{Character string specifying the field separator used in input and
output CSV files.}
}
\value{
Filtered \code{read_count} data frame.
}
\description{
Remove ASVs with fewer than a given number of reads across the entire dataset.
}
\examples{
\dontrun{
filtered_read_count_df <- filter_asv_global(read_count_df, cutoff = 2)
}

}
