% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{filter_indel}
\alias{filter_indel}
\title{Filter ASVs by frame consistency (indel detection)}
\usage{
filter_indel(read_count, outfile = NULL, sep = ",", log_file = NULL)
}
\arguments{
\item{read_count}{Data frame or path to a CSV file with the following columns:
\code{asv_id}, \code{sample}, \code{replicate}, \code{read_count}, \code{asv}.}

\item{outfile}{Character string specifying the CSV file to write the output
data frame. If NULL, no file is written.}

\item{sep}{Character string specifying the field separator used in input and
output CSV files.}

\item{log_file}{Character string specifying the path to the CSV log file.
The path is resolved with the following priority:
\enumerate{
\item the \code{log_file} argument, if explicitly provided by the user;
\item the package-level option/variable storing a default log path
(if set);
\item \code{"vtamR_log.csv"} in the current working directory, used as a
last resort if neither of the above is set.
}
To disable logging entirely, set \code{log_file = NA}.}
}
\value{
Filtered \code{read_count} data frame.
}
\description{
Remove ASVs whose length is not compatible with the dominant reading frame
of the dataset. Specifically, ASVs are filtered out if their length modulo 3
differs from that of the majority of ASVs, which can indicate indel errors.
}
\examples{
\dontrun{
filtered_read_count_df <- filter_indel(read_count_df)
}

}
