% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{filter_replicate}
\alias{filter_replicate}
\title{Filter replicates based on Renkonen distance}
\usage{
filter_replicate(
  read_count,
  outfile = NULL,
  cutoff = NA,
  renkonen_distance_quantile = 0.9,
  sep = ","
)
}
\arguments{
\item{read_count}{Data frame or path to a CSV file with the following columns:
\code{asv_id}, \code{sample}, \code{replicate}, \code{read_count}, \code{asv}.}

\item{outfile}{Character string specifying the CSV file to write the output
data frame. If NULL, no file is written.}

\item{cutoff}{Numeric value between 0 and 1 specifying the maximum acceptable
Renkonen distance for retaining replicates. Replicates exceeding this threshold
relative to most others in the same sample are removed.}

\item{renkonen_distance_quantile}{Numeric value between 0 and 1 specifying a
quantile-based cutoff. If \code{cutoff} is not provided, this quantile of the
Renkonen distance distribution is used to define the threshold (e.g., 0.9
corresponds to the 90th percentile).}

\item{sep}{Character string specifying the field separator used in input and
output CSV files.}
}
\value{
Filtered \code{read_count} data frame with inconsistent replicates removed.
}
\description{
Remove replicates that are inconsistent with other replicates of the same sample
based on Renkonen distance. Replicates that show high dissimilarity to most other
replicates are considered outliers and are filtered out.
}
\examples{
\dontrun{
filtered_read_count_df <- filter_replicate(read_count_df, cutoff = 0.6)
filtered_read_count_df <- filter_replicate(
  read_count_df, 
  renkonen_distance_quantile = 0.9
)
}

}
