% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{FilterChimera}
\alias{FilterChimera}
\title{FilterChimera}
\usage{
FilterChimera(
  read_count_df,
  write_csv = F,
  outdir = "",
  vsearch_path = "",
  by_sample = T,
  sample_prop = 0.8,
  abskew = abskew
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}

\item{vsearch_path}{path to vsearch executable; can be empty if vsearch in the the PATH}

\item{by_sample}{T/F, if T ASVs are flagged as chimera separately for each sample}

\item{sample_prop}{if by_sample=T, the ASV deleted if they are flagged as chimera in at least sample_prop of the samples among the sample they are present}

\item{abskew}{A chimera must be at least abskew times less frequent that the parental ASVs}
}
\description{
Filter out Chimeras
}
