% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{FilterChimera}
\alias{FilterChimera}
\title{FilterChimera}
\usage{
FilterChimera(
  read_count,
  outfile = "",
  vsearch_path = "",
  by_sample = T,
  sample_prop = 0.8,
  abskew = abskew,
  sep = ","
)
}
\arguments{
\item{read_count}{data frame or csv file with the following variables: asv_id, sample, replicate, read_count, asv}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{vsearch_path}{path to vsearch executable; can be empty if vsearch in the the PATH}

\item{by_sample}{T/F, if T ASVs are flagged as chimera separately for each sample}

\item{sample_prop}{if by_sample=T, the ASV deleted if they are flagged as chimera in at least sample_prop of the samples among the sample they are present}

\item{abskew}{A chimera must be at least abskew times less frequent that the parental ASVs}
}
\description{
Filter out Chimeras.
Returns the filtered read_count_df data frame
}
