% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{FilterCodonStop}
\alias{FilterCodonStop}
\title{FilterCodonStop}
\usage{
FilterCodonStop(
  read_count_df,
  write_csv = F,
  outdir = outdir,
  genetic_code = 5,
  sep = ","
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}

\item{genetic_code}{genetic code number from https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=cgencodes}
}
\description{
Filter out all ASVs, if there is a codon stop in all three frames of the direct strand.
Returns the filtered read_count_df data frame.
}
