% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{FilterCodonStop}
\alias{FilterCodonStop}
\title{FilterCodonStop}
\usage{
FilterCodonStop(read_count, outfile = "", genetic_code = 5, sep = ",")
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{genetic_code}{Positive integer: genetic code number from \href{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=cgencodes}{NCBI}.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Filtered read_count_df data frame.
}
\description{
Filter out all ASVs, if there is a codon stop in all three reading
frames of the direct strand.
}
\examples{
\dontrun{
filtered_read_count_df <- FilterCodonStop(read_count_df, genetic_code=5)
}
}
