% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{FilterIndel}
\alias{FilterIndel}
\title{FilterIndel}
\usage{
FilterIndel(read_count_df, write_csv = F, outdir, sep = ",")
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}
}
\description{
Filter out all ASVs, if the modulo 3 of their length is not the same as that of the majority of the ASVs
}
