% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{FilterIndel}
\alias{FilterIndel}
\title{FilterIndel}
\usage{
FilterIndel(read_count, outfile = "", sep = ",")
}
\arguments{
\item{read_count}{data frame or csv file with the following variables: asv_id, sample, replicate, read_count, asv}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}
}
\description{
Filter out all ASVs, if the modulo 3 of their length is not the same as that of the majority of the ASVs
}
