% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{FilterMinReplicate}
\alias{FilterMinReplicate}
\title{FilterMinReplicate}
\usage{
FilterMinReplicate(read_count, cutoff = 2, outfile = "", sep = ",")
}
\arguments{
\item{read_count}{data frame or csv file with the following variables: asv_id, sample, replicate, read_count, asv}

\item{cutoff}{Minimum number of replicates; default=2}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}
}
\description{
Filter out all occurrences where the asv in not present in at least cutoff number of replicates.
Returns the filtered read_count_df data frame.
}
