% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{FilterMinReplicateNumber}
\alias{FilterMinReplicateNumber}
\title{FilterMinReplicateNumber}
\usage{
FilterMinReplicateNumber(
  read_count_df,
  cutoff = 2,
  write_csv = F,
  outdir = NA,
  sep = ","
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{cutoff}{Minimum number of replicates; default=2}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}
}
\description{
Filter out all occurrences where the asv in not present in cutoff replicates.
Returns the filtered read_count_df data frame.
}
