% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{FilterPCRerror}
\alias{FilterPCRerror}
\title{FilerPCRerror}
\usage{
FilterPCRerror(
  read_count_df,
  outfile = "",
  vsearch_path = "",
  pcr_error_var_prop = 0.1,
  max_mismatch = 1,
  by_sample = T,
  sample_prop = 0.8,
  sep = ","
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv_id, sample, replicate, read_count, asv}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{vsearch_path}{path to vsearch executable; can be empty if vsearch in the the PATH}

\item{pcr_error_var_prop}{if the proportion of read_counts of two similar ASVs is less or equal to pcr_error_var_prop, the less abundant is flagged as a PCR error}

\item{max_mismatch}{maximum number of mismatches (gaps included) to consider two ASVs as similar}

\item{by_sample}{T/F, if T ASVs are flagged as an PCR error separately for each sample}

\item{sample_prop}{if by_sample=T, the ASV must be flagged as an PCRerror in sample_prop of the cases to be eliminated}
}
\description{
Filter out an ASVs if it is very similar (max_mismatch) to another more frequent ASV (pcr_error_var_prop).
The whole data set can be analyzed at once (by_sample=F) or sample by sample.
Returns the filtered read_count_df data frame.
}
