% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{FilterPCRerror}
\alias{FilterPCRerror}
\title{FilerPCRerror}
\usage{
FilterPCRerror(
  read_count_df,
  write_csv = F,
  outdir = "",
  vsearch_path = "",
  pcr_error_var_prop = 0.1,
  max_mismatch = 1,
  by_sample = T,
  sample_prop = 0.8,
  sep = ","
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}

\item{vsearch_path}{path to vsearch executable; can be empty if vsearch in the the PATH}

\item{pcr_error_var_prop}{if the proportion of read_counts of two similar ASVs is less or equal to pcr_error_var_prop, the less abundant is flagged as a PCR error}

\item{max_mismatch}{maximum number of mismatches (gaps included) to consider two ASVs as similar}

\item{by_sample}{T/F, if T ASVs are flagged as an PCR error separately for each sample}

\item{sample_prop}{if by_sample=T, the ASV must be flagged as an PCRerror in sample_prop of the cases to be eliminated}
}
\description{
Filter out an ASVs if it is very similar (max_mismatch) to another more frequent ASV (pcr_error_var_prop).
The whole dataset can be analyzed at once (by_sampe=F) or sample by sample.
Returns the filtered read_count_df data frame.
}
