% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{FilterRenkonen}
\alias{FilterRenkonen}
\title{FilterRenkonen}
\usage{
FilterRenkonen(
  read_count_df,
  outfile = "",
  cutoff = NA,
  renkonen_distance_quantile = 0.9,
  sep = ","
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv_id, sample, replicate, read_count, asv}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{cutoff}{\link{0-1}  Filter out all replicates that have renkonen distances above cutoff to most other replicates of the sample}

\item{renkonen_distance_quantile}{If cutoff value is not given, use the renkonen_distance_quantile to determine cutoff value (e.g. with 0.9 as renkonen_distance_quantile, 90\% of the distances are bellow cutoff)}
}
\description{
Filter out all replicates that have renkonen distances above cutoff to most other replicates of the sample.
Returns the filtered read_count_df data frame
}
