% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_functions.R
\name{flag_from_mock}
\alias{flag_from_mock}
\title{flag_from_mock}
\usage{
flag_from_mock(occurrence_df, mock_composition = "", fileinfo_df, sep = ",")
}
\arguments{
\item{occurrence_df}{data frame with the following variables: asv, plate, marker, sample, mean_read_count, sample_type, habitat, action}

\item{mock_composition}{csv file with columns: plate, marker, sample, action (keep/tolerate), asv}

\item{fileinfo_df}{csv file with columns: plate, marker, sample, sample_type(mock/negative/real), habitat}

\item{sep}{separator used in mock_composition}
}
\description{
Flag all occurrences mock samples.
Expected variants as 'keep', unexpected ASVs as 'delete', tolerate ASVs as NA.
Those are ASVs that can be in the mock, but the filtering should not be optimized to keep them.
(e.g. badly amplified species present in the mock)
}
