% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{flag_from_mock}
\alias{flag_from_mock}
\title{flag_from_mock}
\usage{
flag_from_mock(occurrence_df, mock_composition = "", sortedinfo_df, sep = ",")
}
\arguments{
\item{occurrence_df}{data frame with the following variables: asv_id, sample, read_count, asv, sample_type, habitat, action}

\item{mock_composition}{csv file with columns: sample, action (keep/tolerate), asv}

\item{sortedinfo_df}{csv file with columns: sample, sample_type(mock/negative/real), habitat}

\item{sep}{separator used in csv files}
}
\description{
Flag all occurrences in mock samples.
Expected variants as 'keep', unexpected ASVs as 'delete', tolerate ASVs as NA.
Tolerate ASVs are ASVs that can be in the mock, but the filtering should not be optimized to keep them. (e.g. badly amplified species present in the mock)
}
