% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{flagChimera}
\alias{flagChimera}
\title{flagChimera}
\usage{
flagChimera(unique_asv_df, outdir = "", vsearch_path = "", abskew = 2)
}
\arguments{
\item{unique_asv_df}{data frame with the following variables: asv, read_count}

\item{outdir}{name of the output directory for temporary files}

\item{vsearch_path}{path to vsearch executable; can be empty if vsearch in the the PATH}

\item{abskew}{A chimera must be at least abskew times less frequent that the parental ASVs}
}
\description{
Select chimeras in a dataframe of unique ASVs
}
