% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{flagPCRerror_vsearch}
\alias{flagPCRerror_vsearch}
\title{flagPCRerror_vsearch}
\usage{
flagPCRerror_vsearch(
  unique_asv_df,
  outdir = "",
  vsearch_path = "",
  pcr_error_var_prop = 0.1,
  max_mismatch = 1
)
}
\arguments{
\item{unique_asv_df}{data frame with the following variables: asv, read_count; ASVs should be unique}

\item{outdir}{name of the output directory for temporary files}

\item{vsearch_path}{path to vsearch executable; can be empty if vsearch in the the PATH}

\item{pcr_error_var_prop}{if the proportion of read_counts of two similar ASVs is bellow pcr_error_var_prop, the less abundant is flagged as a PCR error}

\item{max_mismatch}{maximum number of mismatches (gaps included) to consider two ASVs as similar}
}
\description{
Identify potential PCRerrors: ASVs very similar (max_mismatch) to another more frequent ASV (pcr_error_var_prop) in a data frame.
Adds a column to the input data frame, with 1 if sequence is a probable PCR error, an d0 otherwise
}
