% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxassign_functions.R
\name{get_lineage_ids}
\alias{get_lineage_ids}
\title{get_lineage_ids}
\usage{
get_lineage_ids(taxids, tax_df)
}
\arguments{
\item{taxids}{vector taxIDs (taxonomic IDs)}

\item{tax_df}{data frame with the following columns: tax_id, parent_tax_id, rank, name_txt, taxlevel (8: species, 7: genus, 6: family, 5: order, 4: class, 3: phylum, 2: kingdom, 1: superkingdom, 0: root)}
}
\description{
Get the complete taxonomic lineage of each taxID in the input vector; return taxIDs of each taxa in the lineage
}
