% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{get_ranked_lineages}
\alias{get_ranked_lineages}
\title{Get Ranked Lineages}
\usage{
get_ranked_lineages(taxids, tax_df, fill_lineage = TRUE)
}
\arguments{
\item{taxids}{Vector of taxIDs.}

\item{tax_df}{Data frame with the following columns:
tax_id, parent_tax_id, rank, name_txt, taxlevel
(8: species, 7: genus, 6: family, 5: order, 4: class, 3: phylum,
2: kingdom, 1: superkingdom, 0: root).}

\item{fill_lineage}{Boolean. If TRUE, fill in missing higher-level taxa
in the lineage using the name of the next known lower-level taxon, prefixed
by the current taxonomic level
(e.g., No_kingdom_Chrysophyceae if kingdom is missing but class is known).}
}
\value{
Data frame with the ranked lineages of taxids.
Columns: ltg_taxid,ltg_name,ltg_rank,ltg_rank_index,
superkingdom_taxid,superkingdom,kingdom_taxid,kingdom,phylum_taxid,phylum,
class_taxid,class,order_taxid,order,family_taxid,family,
genus_taxid,genus,species_taxid,species).
}
\description{
For each input taxid, get the major taxonomic ranks of their lineage.
}
\examples{
taxids <- c(9593,9606)
tax_df <- read.delim(taxonomy="xxxxxxxx", header=T, sep="\t", fill=T, quote="")
get_ranked_lineages(taxids, tax_df)
}
