% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{get_stat}
\alias{get_stat}
\title{get_stat}
\usage{
get_stat(read_count, stat_df = NULL, stage = "", params = NA, outfile = NULL)
}
\arguments{
\item{read_count}{Data frame or path to a CSV file containing the following
variables: \code{asv_id}, \verb{sample``, }replicate,
\code{read_count}, \code{asv}.}

\item{stat_df}{A data frame containing the following variables:

\if{html}{\out{<div class="sourceCode">}}\preformatted{`parameters` Parameters used for the analysis
`asv_count` Number of ASVs detected
`read_count` Total number of reads
`sample_count` Number of samples analyzed
`sample_replicate_count` Number of replicates per sample
}\if{html}{\out{</div>}}

If provided, the function appends a new row to \code{stat_df}. If not provided,
a new data frame with a single row is initialized and returned.}

\item{stage}{Character string specifying the name of the filtering step. It
is used as the row name in \code{stat_df}.}

\item{params}{Character string containing concatenated parameter values used
for the filtering step.}

\item{outfile}{Character string specifying the name of a CSV file to write
the updated data frame. If NULL, no file is written.}
}
\value{
Data frame. The updated \code{stat_df} with an additional row.
}
\description{
Compute summary statistics (reads, variants, samples, and replicates) and
append them to a statistics data frame.
}
\examples{
\dontrun{
get_stat(
  read_count_df,
  stat_df,
  stage = "filter_occurrence_variant",
  params = "0.002;by_replicate=TRUE"
)

get_stat(
  read_count_df,
  stat_df = NULL,
  stage = "filter_indel",
  params = "0.002;by_replicate=TRUE",
  outfile = "out/ReadCount_stat.csv"
)
}

}
