% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{get_stat}
\alias{get_stat}
\title{Get read, variant, sample and replicate counts}
\usage{
get_stat(read_count_df, stat_df, stage, params = NA)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{stat_df}{data frame with the following variables: parameters, asv_count, read_count, read_count, sample_count, sample_replicate_count}

\item{stage}{name of the filtering step; it is used for labeling the line in the stat_df}

\item{params}{parameter value (or their concatenation if more than one) used for the filtering step}
}
\description{
Get read, variant, sample and replicate counts
}
