% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{GetStat}
\alias{GetStat}
\title{GetStat
Get read, variant, sample and replicate counts.
Complete the stat_df with the above statistics.}
\usage{
GetStat(read_count, stat_df, stage = "", params = NA, out = "")
}
\arguments{
\item{read_count}{data frame or csv file with the following variables: asv_id, sample, replicate, read_count, asv}

\item{stat_df}{data frame with the following variables: parameters, asv_count, read_count, read_count, sample_count, sample_replicate_count}

\item{stage}{name of the filtering step; it is used for labeling the line in the stat_df}

\item{params}{parameter value (or their concatenation if more than one) used for the filtering step}

\item{out}{name of the output file}
}
\description{
GetStat
Get read, variant, sample and replicate counts.
Complete the stat_df with the above statistics.
}
