% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{LFN_read_count}
\alias{LFN_read_count}
\title{LFN_read_count}
\usage{
LFN_read_count(read_count_df, cutoff = 10, write_csv = F, outdir = NA)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{cutoff}{minimum number of reads for an occurrence; default=10; default=10}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}
}
\description{
Eliminate occurrences with less than cutoff reads
}
