% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{LFNglobalReadCount}
\alias{LFNglobalReadCount}
\title{LFNglobalReadCount}
\usage{
LFNglobalReadCount(read_count, cutoff = 10, outfile = "", sep = ",")
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{cutoff}{Positive integer: minimum number of reads for an ASV in the
whole data set. Bellow this cutoff, the ASV is deleted from the data set.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Filtered read_count_df data frame.
}
\description{
Eliminate ASVs with less than cutoff reads in the data set.
}
\examples{
filtered_read_count_df <- LFNglobalReadCount(read_count_df, cutoff=2)
}
