% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{make_adapter_fasta}
\alias{make_adapter_fasta}
\title{make_adapter_fasta}
\usage{
make_adapter_fasta(fastainfo_df, fasta_file, tag_to_end, outdir)
}
\arguments{
\item{fastainfo_df}{data frame with column: tag_fw,tag_rv,fasta}

\item{fasta_file}{name of the fasta file for wich tag-combination should be written to a tags.fasta file}

\item{tag_to_end}{\link{T/F} if T use anchored search (tags are at the extremity of the sequence)}

\item{outdir}{output directory}
}
\description{
Make a fasta file with tag combinations
}
