% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{match_variants_to_mock_species}
\alias{match_variants_to_mock_species}
\title{match_variants_to_mock_species}
\usage{
match_variants_to_mock_species(
  read_count,
  fas,
  taxonomy = "",
  blast_path = "blastn",
  sampleinfo = "",
  outdir = NULL,
  num_threads = 0,
  sep = ",",
  quiet = TRUE
)
}
\arguments{
\item{read_count}{Data frame or csv file containing an asv_id, sample, asv and
read_count columns.}

\item{fas}{Character string; name of a FASTA file containing sequences from species
expected in the mock samples or from closely related taxa. The FASTA header line
should follow this format: \verb{>HQ563207.1 taxID=1592914}, where \code{taxID} is a valid
NCBI taxonomic identifier (https://www.ncbi.nlm.nih.gov/taxonomy).}

\item{taxonomy}{Character string; path to a TSV file containing the following columns:
\code{tax_id}, \code{parent_tax_id}, \code{rank}, \code{name_txt}, \code{old_tax_id} (merged into \code{tax_id}),
and \code{taxlevel} (8: species, 7: genus, 6: family, 5: order, 4: class, 3: phylum,
2: kingdom, 1: domain, 0: root).
A taxonomy file distributed with the COInr database is suitable for Eukaryotes (any marker).}

\item{blast_path}{Character string: path to BLAST executable.}

\item{sampleinfo}{Data frame or csv file containing sample and sample_type columns.}

\item{outdir}{Character string: A directory to write output files.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
Data frame with the following columns: sample, action, asv, taxon, asv_id
}
\description{
Identify potential expected variants in mock samples.
}
\details{
This function performs the following steps:
\itemize{
\item Creates a small BLAST database from sequences corresponding to the species
expected in the mock samples, or from closely related taxa. The sequences
should partially cover the amplified region (70\% coverage) but may be shorter or longer.
\item Assigns taxonomy to all ASVs detected in the mock samples using this custom database.
\item Selects the most abundant ASV for each taxon.
\item Generates a \code{mock_composition} template file that users should review and, if necessary, edit.
}

The output directory will contain two files:
\itemize{
\item \emph{mock_taxassign.csv}:
Contains the results of the taxonomic assignment, including the total number of reads
detected in the mock samples and the number of mock samples in which each ASV is present.
\item \emph{mock_composition_template_to_check.csv}:
A template for the \code{mock_composition} file.
It includes the most abundant sequence for each \code{ltg_name} identified in the taxonomic
assignment output, repeated for each mock sample.
When working with different mock samples that have distinct compositions, users should
remove the lines corresponding to taxa not expected in a given sample.

Note: the file does \emph{not} include sequences for species in the custom database that
did not show significant similarity to any ASV. This may occur if the species was not
amplified in the mock samples or if the reference sequence in the custom database is erroneous.
}
}
\examples{
\dontrun{
match_variants_to_mock_species(read_count=read_count_df, fas=xxxxx, taxonomy="xxxxxx", 
blast_path=blast_path, sampleinfo=sampleinfo)
}
}
