% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_helper_functions.R
\name{miem_bioinformatics}
\alias{miem_bioinformatics}
\title{Generate a MIEM bioinformatics report}
\usage{
miem_bioinformatics(log, r_versions, outfile = "", sep = ",")
}
\arguments{
\item{log}{A data.frame containing the vtamR log or the path to a CSV
log file.}

\item{r_versions}{A data.frame containing installed R package versions
or the path to a CSV file.}

\item{outfile}{Optional output CSV file.}

\item{sep}{Field separator used when reading/writing CSV files.}
}
\value{
A data.frame with three columns:
\itemize{
\item Step
\item Information
\item Reporting Requirements
}
}
\description{
Generate a MIEM (Minimum Information for an eDNA Metabarcoding Study)
reporting table from a vtamR log file.
}
\details{
The function extracts software versions, pipeline parameters and
taxonomic assignment settings from a vtamR log and summarizes them
in a table suitable for inclusion in the Supplementary Material of
a manuscript.
}
