% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{OptimizeLFNreadCountLFNvariant}
\alias{OptimizeLFNreadCountLFNvariant}
\title{OptimizeLFNreadCountLFNvariant}
\usage{
OptimizeLFNreadCountLFNvariant(
  read_count,
  known_occurrences,
  sep = ",",
  outfile = "",
  min_lfn_read_count_cutoff = 10,
  max_lfn_read_count_cutoff = 100,
  increment_lfn_read_count_cutoff = 5,
  min_lnf_variant_cutoff = 0.001,
  max_lnf_variant_cutoff = 0.01,
  increment_lnf_variant_cutoff = 0.001,
  by_replicate = FALSE,
  min_replicate_number = 2,
  quiet = T
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{known_occurrences}{Data frame or file produced by
\code{MakeKnownOccurrences} function, with known False Positives and True Positives.}

\item{sep}{Field separator character in input and output csv files.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{min_lfn_read_count_cutoff}{Positive integer: the lowest cutoff value for
\code{LFNreadCount} function.}

\item{max_lfn_read_count_cutoff}{Positive integer: the highest cutoff value
for \code{LFNreadCount} function.}

\item{increment_lfn_read_count_cutoff}{Positive integer: values from
\code{min_lfn_read_count_cutoff} to \code{max_lfn_read_count_cutoff}
are tested by \code{increment_lfn_read_count_cutof} of increment.}

\item{min_lnf_variant_cutoff}{Real (0-1): the lowest cutoff value for
\code{LFNvariant} function.}

\item{max_lnf_variant_cutoff}{Real (0-1): the highest value for \code{LFNvariant}
function.}

\item{increment_lnf_variant_cutoff}{Real (0-1): values from
\code{min_lnf_variant_cutoff} to \code{max_lnf_variant_cutoff} are tested by
\code{increment_lnf_variant_cutoff} increment.}

\item{by_replicate}{logical: : compare read count of the occurrence to the
read counts of the ASV-replicate in \code{LFNvariant} function.}

\item{min_replicate_number}{Positive integer: minimum number of replicates
for \code{FilterMinReplicate}.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
data frame with the following columns:
lfn_read_count_cutoff, lnf_variant_cutoff, FN, TP, FP.
}
\description{
Suggest optimal parameters for \code{lfn_read_count_cutoff} and \code{lnf_variant_cutoff}
functions.
The the \code{LFNreadCount} and \code{LFNvariant} is run for a series of parameter value
combinations followed by \code{FilterMinReplicate}.
For each parameter combination, the number of FN, TP, and FP is reported.
Users should chose the parameter setting that minimizes, FN and FP.
}
\examples{
\dontrun{
OptimizeLFNreadCountLFNvariant(read_count_df, 
    known_occurrences="data/known_occurrences.csv", 
    min_lfn_read_count_cutoff=10,
     max_lfn_read_count_cutoff=50, 
     increment_lfn_read_count_cutoff=10, 
     min_lnf_variant_cutoff=0.001, 
     max_lnf_variant_cutoff=0.005, 
     increment_lnf_variant_cutoff=0.001
     )
}
}
