% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{OptimizeLFNsampleReplicate}
\alias{OptimizeLFNsampleReplicate}
\title{OptimizeLFNsampleReplicate}
\usage{
OptimizeLFNsampleReplicate(
  read_count_df,
  mock_composition = "",
  sep = ",",
  outfile = ""
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv_id, plate, marker, sample, replicate, read_count, asv}

\item{mock_composition}{csv file with columns: sample, action (keep/tolerate), asv}

\item{sep}{separator used csv files}

\item{outfile}{name of the output file; Optional; If empty the results are not written to a file}
}
\description{
Prepare a data frame that lists all expected occurrences in all mock sample replicates their read_counts and the proportion of
read_counts to the total number of reads in the sample-replicate.
The lfn_variant_replicate parameter should be bellow the smallest proportion in order to keep all expected ASVs in the data set.
}
