% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{OptimizePCRError}
\alias{OptimizePCRError}
\title{OptimizePCRError}
\usage{
OptimizePCRError(
  read_count,
  mock_composition = "",
  sep = ",",
  outfile = "",
  max_mismatch = 1,
  min_read_count = 2
)
}
\arguments{
\item{read_count}{data frame or csv file with the following variables: asv_id, sample, replicate, read_count, asv}

\item{mock_composition}{data frame or csv file with columns: sample, action (keep/tolerate), asv}

\item{sep}{separator used in mock_composition}

\item{outfile}{name of the output file; Optional; If empty the results are not written to a file}

\item{max_mismatch}{maximum number of mismatches allowed between two asv where one of the asvs is considered as a PCRerror}

\item{min_read_count}{occurrences under this read_count limits are ignored}
}
\description{
Prepare a data frame that lists pairs of expected and unexpected ASVs in mock samples maximum max_mismatch difference between them.
The pcr_error_var_prop parameter should be above the highest pcr_error_var_prop  (unexpected_read_count/expected_read_count) value in the table.
}
