% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_functions.R
\name{OptimizePCRError}
\alias{OptimizePCRError}
\title{OptimizePCRError}
\usage{
OptimizePCRError(
  read_count_df,
  mock_composition = "",
  sep = ",",
  outdir = "",
  max_mismatch = 1,
  min_read_count = 2
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{mock_composition}{csv file with columns: plate, marker, sample, action (keep/tolerate), asv}

\item{sep}{separator used in mock_composition}

\item{outdir}{name of the output directory}

\item{max_mismatch}{maximum number of mismatches allowed between two asv where one of the asvs is considered as a PCRerror}

\item{min_read_count}{occurrences under this read_count limits are ignored}
}
\description{
Prepare a file that lists all pairs of expected and unexpected ASVs in mock samples with a single difference between them,
and their read counts. The pcr_error_var_prop parameter should be above the highest pcr_error_var_prop value in the table.
}
