% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{PairwiseIdentity}
\alias{PairwiseIdentity}
\title{Pairwise identity}
\usage{
PairwiseIdentity(
  asv,
  min_id = 0.8,
  vsearch_path = vsearch,
  quiet = TRUE,
  num_threads = 0
)
}
\arguments{
\item{asv}{data frame or csv file with asv and asv_id columns.}

\item{min_id}{bellow this percentage of identity do not aligne asv}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}
}
\value{
data frame with asv pairs and their percentage of identity.
asv pairs with identity bellow min_id are not listed.
}
\description{
Align all pairs of asv with at least min_id similarity and
make a data frame with pairs of asv and their percentage of identity.
asv pairs with identity bellow min_id are not listed.
}
\examples{
\dontrun{
check_dir(path="data")
}
}
