% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_graph.R
\name{plot_renkonen_distance_density}
\alias{plot_renkonen_distance_density}
\title{Plot Renkonen Distance Density}
\usage{
plot_renkonen_distance_density(df, plotfile = "", log_file = NULL)
}
\arguments{
\item{df}{Data frame with the following columns:
\code{sample1}, \code{sample2}, \code{replicate1}, \code{replicate2}, \code{renkonen_d}.
This output can be produced using \code{compute_renkonen_distances()}.}

\item{plotfile}{Character string: name of the output PNG file. If empty,
no file is written.}

\item{log_file}{Character string specifying the path to the CSV log file.
The path is resolved with the following priority:
\enumerate{
\item the \code{log_file} argument, if explicitly provided by the user;
\item the package-level option/variable storing a default log path
(if set);
\item \code{"vtamR_log.csv"} in the current working directory, used as a
last resort if neither of the above is set.
}
To disable logging entirely, set \code{log_file = NA}.}
}
\value{
A density plot object.
}
\description{
Creates a density plot of Renkonen distances between pairs of samples or
sample–replicates.
}
