% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_graph.R
\name{plot_renkonen_distance_density}
\alias{plot_renkonen_distance_density}
\title{plot_renkonen_distance_density}
\usage{
plot_renkonen_distance_density(df, plotfile = "")
}
\arguments{
\item{df}{data frame with the following columns:
sample1,sample2,replicate1,replicate2,renkonen_d
(can be produced by compute_renkonen_distances)}

\item{plotfile}{Character string: png file name for the output plot;
If empty, no file is written.}
}
\value{
Density plot
}
\description{
Create density plot with Renkonen distances between pairs of sample-replicates
}
