% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{pool_by_cluster}
\alias{pool_by_cluster}
\title{Pool ASVs of the same cluster}
\usage{
pool_by_cluster(read_count_df, cluster_df)
}
\arguments{
\item{read_count_df}{Data frame with the following variables:
asv, sample, replicate (optional), read_count.}

\item{cluster_df}{Data frame with the following variables:
asv_id, cluster_id.}
}
\value{
Data frame: same structure as the input (read_count_df), but ASVs of
the same cluster pooled to one row.
}
\description{
Pools variants of the same cluster and sums read counts of the
underlying ASVs.
}
\examples{
\dontrun{
pool_by_cluster(read_count=read_count_df, cluster_df=cluster_df)
}
}
