% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{pool_datasets}
\alias{pool_datasets}
\title{pool_datasets}
\usage{
pool_datasets(files, outfile = "", sep = ",", fun = "mean")
}
\arguments{
\item{files}{data frame with the following variables: file (name of input files), marker; Input files must have asv, sample and mean_read_count columns}

\item{outfile}{name of the output file; Optional; If empty the results are not written to a file}

\item{sep}{separator used in csv files}

\item{fun}{\link{mean/sum} ; For each cluster either the mean of the sum of the read counts and calculated over the underlying ASVs Default: "mean"}
}
\description{
Take several output files, each in long format containing asv_id, sample, mean_read_count and asv columns
Pool the different data sets, if all have the same marker
If more than one marker, ASVs identical on their overlapping regions are pooled into clusters, and different asvs of the same cluster are pooled under the centroîd (longest ASV).
Pooling can take the mean of the read counts of the ASV (default) or their sum.
}
