% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{pool_datasets}
\alias{pool_datasets}
\title{pool_datasets}
\usage{
pool_datasets(
  files,
  outdir = "",
  sep = ",",
  mean_over_markers = T,
  write_csv = F
)
}
\arguments{
\item{files}{data frame with the following variables: file (name of input files), marker; Input files must have asv, sample and mean_read_count columns}

\item{outdir}{name of the output directory}

\item{sep}{separaton used in csv files}

\item{mean_over_markers}{\link{T/F} If TRUE, the mean read count is calculated over different ASVs of each cluster. Sum otherwise. Default: TRUE}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}
}
\description{
Take several output files, each has long format, ASVs and mean_read_count over replicates
Pool the different dataset, if all have the sale marker
If more than one marker, ASVs identical on their overlapping regions are pooled into clusters, and different asvs of the same cluster aye pooled under the centroid (longest ASV).
Pooling can take the mean of the read counts of the ASV (default) or their sum.
}
