% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{pool_datasets}
\alias{pool_datasets}
\title{pool_datasets}
\usage{
pool_datasets(
  files,
  outfile = "",
  asv_with_centroids = "",
  sep = ",",
  mean_over_markers = T,
  vsearch_path = "",
  quiet = T
)
}
\arguments{
\item{files}{data frame with the following variables: file (name of input files), marker; Input files must have asv, sample and read_count columns}

\item{outfile}{name of the output file with asv_id, sample, read_count and asv columns, after pooling ASVs identical on their corresponding regions together; Optional}

\item{asv_with_centroids}{name of the output file containing the same information as the concatenated input files, completed by centroid_id and centroid columns; Optional}

\item{sep}{separator used in csv files}

\item{mean_over_markers}{\link{T/F} If TRUE, the mean read count is calculated over different ASVs of each cluster. Sum otherwise. Default: TRUE}

\item{vsearch_path}{path to vsearch executables}

\item{quiet}{\link{T/F}; TRUE by default print only warnings and error to STDOUT}
}
\description{
Take several input files, each in long format containing asv_id, sample, read_count and asv columns
Pool the different data sets, if all have the same marker
If more than one markers, ASVs identical on their overlapping regions are pooled into groups, and different ASVs of the same group are pooled under the centroid (longest ASV of the group).
Pooling can take the mean of the read counts of the ASV (default) or their sum.
Return a pooled data frame.
}
