% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{pool_filters}
\alias{pool_filters}
\title{Retain occurrences shared across multiple filtered datasets}
\usage{
pool_filters(..., outfile = NULL, sep = ",", log_file = NULL)
}
\arguments{
\item{...}{Data frames with the following columns:
\code{asv_id}, \code{sample}, \code{replicate} (optional), \code{read_count}, \code{asv}.}

\item{outfile}{Character string specifying the CSV file to write the output
data frame. If NULL, no file is written.}

\item{sep}{Character string specifying the field separator used in input and
output CSV files.}

\item{log_file}{Character string specifying the path to the CSV log file.
If \code{NULL}, no log file is written.}
}
\value{
Filtered \code{read_count_df} data frame containing only shared occurrences.
}
\description{
Pool multiple \code{read_count_df} data frames and retain only occurrences
present in all filters.
}
\examples{
\dontrun{
filtered_read_count_df <- pool_filters(read_count_df1, read_count_df2)
}

}
