% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{pool_filters}
\alias{pool_filters}
\title{pool_filters}
\usage{
pool_filters(..., outfile = "", sep = ",")
}
\arguments{
\item{...}{Data frames with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Filtered read_count_df data frame.
}
\description{
Pool all count_read_df data frames, and keep only occurrences
present in all filters.
}
\examples{
\dontrun{
filtered_read_count_df <- pool_filters(read_count_df1, read_count_df2)
}
}
