% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{pool_LFN}
\alias{pool_LFN}
\title{pool_LFN}
\usage{
pool_LFN(..., write_csv = F, outdir = NA, sep = ",")
}
\arguments{
\item{...}{a list of data frames}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}
}
\description{
pool all count_read_df data frames, and keep only occurrences present in all filters
}
