% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R, R/vtamR_fun.R
\name{pool_markers}
\alias{pool_markers}
\title{Pool data from different markers}
\usage{
pool_markers(
  files,
  marker_ids = NULL,
  outfile = "",
  asv_with_centroids = "",
  method = "mean",
  vsearch_path = "vsearch",
  num_threads = 0,
  sep = ",",
  quiet = T
)

pool_markers(
  files,
  marker_ids = NULL,
  outfile = "",
  asv_with_centroids = "",
  method = "mean",
  vsearch_path = "vsearch",
  num_threads = 0,
  sep = ",",
  quiet = T
)
}
\arguments{
\item{files}{Character vector specifying the file paths of the data sets to pool.
Each file must have the same format, with columns: \code{asv_id}, \code{sample},
\code{replicate} (optional), \code{read_count}, and \code{asv}.}

\item{marker_ids}{Character vector specifying marker IDs.
Markers ID must be positive non-zero integers. They should be in the same order
as the csv file. Marker ids are post-fixed to the asv_id, to avoid confusion
between the same asv_ids across different markers. If not provided, the asv_id
of the different markers must be all unique among all ASVs of all markers.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{asv_with_centroids}{Character string: csv file name of the output file
containing the same information as the concatenated input files,
completed by centroid_id and centroid columns.}

\item{method}{Character string specifying how read counts from identical
sample–replicates should be aggregated. Must be one of "mean",
"max", "sum", or "min".}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}

\item{...}{Data frames with the following variables:
marker, asv_id, sample, replicate (optional), read_count, asv.}
}
\value{
Data frame with asv_id, sample, replicate (optional), read_count, asv columns.

Data frame with asv_id, sample, replicate (optional), read_count, asv columns.
where the asvs of the same group-sample-replicate are pooled into one line.
read_count is the mean, max, min or sum over the pooled asvs.
}
\description{
Take two or more input files containing filtered results of the same samples
from different but strongly overlapping markers.
Files are in long format with asv_id, sample, replicate (optional),
read_count and asv columns.

ASVs identical on their overlapping
regions are pooled into groups, and different ASVs of the same group
are pooled under the centroid (longest ASV of the group). The asv_id are
prefixed by the marker, to avoid confounding different ASVs of different
markers, with the same id.
Pooling can take the mean of the read counts of the ASV (default), their sum
or maximum.

Take several input files, each in long format containing
asv_id, sample, replicate (optional), read_count and asv columns.

ASVs identical on their overlapping regions are pooled into groups,
and different ASVs of the same group are pooled under the centroid
(longest ASV of the group). The asv_id are
prefixed by the marker, to avoid confounding different ASVs of different
markers, with the same id.
Pooling can take the mean, max, min, or sum of the read counts of the asv.
}
\examples{
\dontrun{
markers_pooled <- pool_markers(df_mfzr, df_zfzr, method="mean")
}
\dontrun{
files <- data.frame(file=c("vtamR_test/out_mfzr/14_PoolReplicates.csv", 
    "vtamR_test/out_zfzr/14_PoolReplicates.csv"),
    marker=c("MFZR", "ZFZR"))
PoolDatasets(files, vsearch_path=vsearch_path)
}
}
