% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fun.R
\name{pool_markers}
\alias{pool_markers}
\title{Pool data from different markers}
\usage{
pool_markers(
  ...,
  method = "mean",
  outfile = "",
  asv_with_centroids = "",
  sep = ",",
  vsearch_path = "vsearch",
  num_threads = 0,
  quiet = T
)
}
\arguments{
\item{...}{Data frames with the following variables:
marker, asv_id, sample, replicate (optional), read_count, asv.}

\item{method}{Character string specifying how read counts should be pooled.
Must be one of "mean", "max" or "sum".}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{asv_with_centroids}{Character string: csv file name of the output file
containing the same information as the concatenated input files,
completed by centroid_id and centroid columns. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
Data frame with asv_id, sample, replicate (optional), read_count, asv columns.
}
\description{
Take two or more input files containing filtered results of the same samples
from different but strongly overlapping markers.
Files are in long format with asv_id, sample, replicate (optional),
read_count and asv columns.

ASVs identical on their overlapping
regions are pooled into groups, and different ASVs of the same group
are pooled under the centroid (longest ASV of the group). The asv_id are
prefixed by the marker, to avoid confounding different ASVs of different
markers, with the same id.
Pooling can take the mean of the read counts of the ASV (default), their sum
or maximum.
}
\examples{
\dontrun{
markers_pooled <- pool_markers(df_mfzr, df_zfzr, method="mean")
}
}
