% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{pool_replicates}
\alias{pool_replicates}
\title{Pool replicates sample by sample}
\usage{
pool_replicates(
  read_count,
  method = "mean",
  digits = 0,
  outfile = "",
  sep = ","
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv, cluster_id (optional).}

\item{method}{Character string specifying how replicate read counts should be pooled.
Must be one of "mean", "max", "sum" or "min".}

\item{digits}{Positive integer: round the mean read counts to digits.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Data frame with the following columns:
asv, sample, read_count (over replicates), cluster_id (optional)
}
\description{
For each sample, replicates are pooled. The read count for a given
sample and ASV can be calculated as the mean of non-zero read counts
across replicates, or alternatively as the sum, maximum, or minimum.
}
\examples{
\dontrun{
pool_replicates(read_count_df)
}
}
