% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{PoolReplicates}
\alias{PoolReplicates}
\title{Pool replicates sample by sample}
\usage{
PoolReplicates(read_count, digits = 0, outfile = "", sep = ",")
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate, read_count, asv.}

\item{digits}{Positive integer: round the mean read counts to digits.}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Data frame with the following columns:
asv, sample, read_count (over replicates)
}
\description{
Take the mean non-zero read counts over replicates for each sample and asv.
}
\examples{
\dontrun{
PoolReplicates(read_count_df)
}
}
