% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{PoolReplicates}
\alias{PoolReplicates}
\title{Pool replicates sample by sample}
\usage{
PoolReplicates(read_count_df, digits = 0, outfile = "", sep = ",")
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv_id, sample, replicate, read_count, asv}

\item{digits}{round the mean read counts to digits}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}
}
\description{
Take the mean non-zero read counts over replicates for each sample and asv.
Returns a data frame with the following columns: asv, plate, marker, sample, read_count (over replicates)
}
