% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{RandomSeq}
\alias{RandomSeq}
\title{RandomSeq}
\usage{
RandomSeq(
  fastainfo_df,
  fasta_dir = "",
  outdir = "",
  vsearch_path = "",
  n,
  randseed = 0,
  compress = F,
  sep = ","
)
}
\arguments{
\item{fastainfo_df}{data frame with a 'fasta' column containing input file names; files can be compressed in gz and zip format}

\item{fasta_dir}{directory that contains the input fasta files}

\item{outdir}{directory for the output files}

\item{vsearch_path}{path to vsearch}

\item{n}{integer; the number of randomly selected sequences}

\item{randseed}{positive integer; seed for random sampling; 0 by default means to use a pseudo-random seed, a given non-zero seed produce always the same result}

\item{compress}{\link{T/F}; If TRUE, output files are compressed in the same format as the input. Otherwise the output is uncompressed;}

\item{separator}{used un csv files; "," by default}
}
\description{
Random select n sequences from each input fasta file.
Do not work on Windows! If using Windows, please, use RandomSeqWindows
}
