% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{read_blast_res}
\alias{read_blast_res}
\title{read_blast_res}
\usage{
read_blast_res(file)
}
\arguments{
\item{file}{name of the output of BLAST; tab separated colums: qseqid,pident,qcovhsp,staxids}
}
\description{
Read BLAST result to a data frame; If more than one taxid for a hit, make a separate line for each taxid
Return a data frame with the following columns: "qseqid","pident","qcovhsp","staxids"
}
