% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{read_fasta_seq}
\alias{read_fasta_seq}
\title{read_fasta_seq}
\usage{
read_fasta_seq(filename = filename, dereplicate = F)
}
\arguments{
\item{filename}{name of the input file including full path}

\item{dereplicate}{\link{T/F} if T, return asvs with read counts}
}
\description{
Read sequences from a fasta file. Fasta can be gz compressed or uncompressed
Returns a either a data frame with read read in a line, or a data frame with asv and read counts (if dereplicate==T)
}
