% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{read_fasta_to_df}
\alias{read_fasta_to_df}
\title{read_fasta_to_df}
\usage{
read_fasta_to_df(file)
}
\arguments{
\item{file}{input fasta file; can be  uncompressed or compressed in gz format (zip files are not supported)}
}
\description{
Read a fasta file to a data frame with two columns: headers and sequences
}
