% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_functions.R
\name{read_fastas_from_fileinfo}
\alias{read_fastas_from_fileinfo}
\title{Read all fasta files in fileinfo file to a data frame}
\usage{
read_fastas_from_fileinfo(file, dir, write_csv = F, outdir = NA)
}
\arguments{
\item{file}{tsv file with columns: plate    marker    sample    replicate    file}

\item{dir}{name of the directory with fasta file}

\item{write_csv}{T/F; write read_counts to csv file; default=FALSE}

\item{outdir}{name of the output directory}
}
\description{
Read all fasta file in fileinfo file (plate	marker	sample	replicate	filename).
Demultiplex reads to ASVs.
Count the number of reads of each ASVs in each sample-replicate.
Returns a df ("asv", "plate", "marker", "sample", "replicate", "read_count").
}
