% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{read_fastas_from_fileinfo}
\alias{read_fastas_from_fileinfo}
\title{Read all fasta files in fileinfo file to a data frame}
\usage{
read_fastas_from_fileinfo(
  sortedinfo_df,
  dir = "",
  outfile = "",
  sep = ",",
  asv_list = "",
  updated_asv_list = ""
)
}
\arguments{
\item{sortedinfo_df}{data frame with columns:  sample,  replicate, file name, (optional: sample_type(mock/negative/real), habitat)}

\item{dir}{name of the directory with fasta files}

\item{outfile}{Name of the output csv file with asv_id, sample, replicate, read_count and asv as columns; if no file name provided, only a data frame is returned}

\item{sep}{separator in csv files; default: ","}

\item{asv_list}{name of the file, containing asvs and asv_ids from earlier analyses. Optional. It is used to homogenize asv_ids between different data sets}

\item{updated_asv_list}{name of the output file, containing the updated ASVs. Optional.}
}
\description{
Read all fasta file in sortedinfo_df data frame (sample	replicate	file name).
Dereplicate reads to ASVs.
Count the number of reads of each ASVs in each sample-replicate.
Returns a df ("asv", "sample", "replicate", "read_count")
}
