% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_treatment.R
\name{reverse_complement}
\alias{reverse_complement}
\title{reverse_complement}
\usage{
reverse_complement(sequence)
}
\arguments{
\item{sequence}{DNA sequence}
}
\description{
Reverse complement a DNA sequence. IUPAC codes are accepted
}
