% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{run_blast}
\alias{run_blast}
\title{run_blast}
\usage{
run_blast(
  df,
  blast_db = "",
  blast_path = "",
  outdir = "",
  qcov_hsp_perc = 70,
  perc_identity = 70,
  num_threads = 8
)
}
\arguments{
\item{df}{a data frame containing and asv column}

\item{blast_db}{BLAST DB incuding path}

\item{blast_path}{path to BLAST executables}

\item{outdir}{output directory}

\item{qcov_hsp_perc}{minimum query coverage}

\item{perc_identity}{minimum percentage of identity}

\item{num_threads}{number of threads}
}
\description{
Run BLAST
}
