% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{run_swarm}
\alias{run_swarm}
\title{run_swarm}
\usage{
run_swarm(
  read_count_df,
  swarm_path = "",
  num_threads = 1,
  swarm_d = 1,
  fastidious = T
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{swarm_path}{Path to th swarm executable (Default: TRUE)}

\item{num_threads}{Number of CPUs}

\item{swarm_d}{positive integer, d parameter for swarm (1 by default); maximum number of differences allowed between two amplicons, meaning that two amplicons will be grouped if they have d (or less) differences.}

\item{fastidious}{\link{T/F} when working with d = 1, perform a second clustering pass to reduce the number of small clusters (Default: TRUE)}
}
\description{
Run swarm (https://github.com/torognes/swarm) on input read_count_df data frame, pool variants of the same cluster sum reads of the underlying ASVs
Return a data frame with the same structure as the input
}
