% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{run_system2}
\alias{run_system2}
\title{Run a command using system2}
\usage{
run_system2(path, args, quiet = FALSE)
}
\arguments{
\item{path}{Character string naming a path to a program.}

\item{args}{vector; Arguments of the command.}

\item{quiet}{Logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\description{
Run a command using system2 and handles quite or verbose output
}
\examples{
\dontrun{
run_system2(path="ls", args=c("-all"), quiet=FALSE)
}
}
