% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_treatment.R
\name{select_sequences}
\alias{select_sequences}
\title{select_sequences}
\usage{
select_sequences(fasta_dir = "", file, outdir = "", random_integers)
}
\arguments{
\item{fasta_dir}{path to the input fasta files}

\item{file}{input fasta file; can be  uncompressed o compressed in gz format only}

\item{outdir}{directory for the output files; same compression as the input file}

\item{random_integers}{vector of random integers between 1 and the number of sequences in the input file}
}
\description{
Select sequences from the input file that correspond to the vector of integers
}
