% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{smart_gzip}
\alias{smart_gzip}
\title{Compress or uncompress file}
\usage{
smart_gzip(
  file,
  outfile = "",
  remove = FALSE,
  pigz_path = "pigz",
  pigz = FALSE,
  num_threads = 0,
  quiet = TRUE,
  compress = FALSE
)
}
\arguments{
\item{file}{Character string: Input file name.}

\item{outfile}{Character string: Output file name. If empty, automatically derived from input.}

\item{remove}{Logical: Remove input file after successful operation.}

\item{pigz_path}{Character string: Path to \code{pigz} executable.}

\item{pigz}{Logical: Use pigz for compression/decompression. If FALSE uses \code{R.utils}.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}

\item{quiet}{Logical: If \code{TRUE}, suppress informational messages.}

\item{compress}{Logical: If \code{TRUE}, compress (\code{gzip}); if \code{FALSE}, decompress (\code{gunzip}).}
}
\value{
Character string: Name of the output file including path.
}
\description{
Compress or uncompress a file using \code{pigz} (if available, faster) or fallback to \code{R.utils}.
}
\examples{
\dontrun{
# Decompress a .gz file
smart_gzip("data.fastq.gz", compress = FALSE)

# Compress a file
smart_gzip("data.fastq", compress = TRUE)
}
}
