% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{suggest_sample_cutoff}
\alias{suggest_sample_cutoff}
\title{Suggest cutoff for \code{filter_occurrence_sample}}
\usage{
suggest_sample_cutoff(
  read_count,
  mock_composition,
  sep = ",",
  outfile = NULL,
  log_file = NULL
)
}
\arguments{
\item{read_count}{Data frame or CSV file with the following variables:
\code{asv_id}, \code{sample}, \code{replicate}, \code{read_count}, \code{asv}.}

\item{mock_composition}{Data frame or CSV file with columns:
\code{sample}, \code{action} (\code{keep} or \code{tolerate}), \code{asv}.}

\item{sep}{Field separator character used in input and output CSV files.}

\item{outfile}{Character string specifying the output CSV file. If NULL,
no file is written.}

\item{log_file}{Character string specifying the path to the CSV log file.
If \code{NULL}, no log file is written.}
}
\value{
Data frame with the following columns: \code{sample}, \code{replicate}, \code{action},
\code{asv_id}, \code{read_count}, \code{read_count_sample_replicate}, \code{sample_cutoff}, \code{asv}.
}
\description{
Compute read count proportions of expected ASV occurrences in mock
sample-replicates to help determine an appropriate cutoff for
\code{filter_occurrence_sample}.
}
\details{
The cutoff should be set below the smallest observed proportion to ensure
that all expected ASVs are retained in the dataset.
}
\examples{
\dontrun{
suggest_sample_cutoff(
  read_count_df,
  mock_composition = "data/mock_composition.csv"
)
}

}
