% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{SummarizeBy}
\alias{SummarizeBy}
\title{SummarizeBy}
\usage{
SummarizeBy(
  dir,
  sep = ",",
  outfile = "",
  feature = "asv",
  grouped_by = "sample"
)
}
\arguments{
\item{dir}{directory containing the output files of intermediate filtering steps; file names should start with a number followed by an underscore (e.g. 5_LFNsampleReplicate.csv)}

\item{sep}{separator used in csv files}

\item{outfile}{name of the output file; optional, it is not given results are returned in a data frame, but no file is written}

\item{feature}{\link{asv_id/asv/sample/replicate/read_count} if read_count, get the total number of reads for each of the grouped values, otherwise get the number of distinct values of feature}

\item{grouped_by}{\link{asv_id/asv/sample/replicate} group data by this variable}

\item{value}{values of feature that should selected. That output data frame contains all lines where value is present in feature in the input files}
}
\description{
Summarizes the output of each intermediate filtering steps.
Takes all files in the dir, that start with a number.
For each file, groups the lines by grouped_by variable, and gets the number of distinct values of feature for each group, or if feature is read_count, gets the total number of reads of each group
}
