% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{swarm}
\alias{swarm}
\title{swarm}
\usage{
swarm(
  read_count_df,
  outfile = "",
  swarm_path = "",
  num_threads = 1,
  swarm_d = 1,
  fastidious = T,
  write_csv = F,
  sep = ",",
  by_sample = T
)
}
\arguments{
\item{read_count_df}{data frame with the following variables: asv, plate, marker, sample, replicate, read_count}

\item{outfile}{name of the output file with asv_id, sample, replicate, read_count and asv columns; Optional; If empty the results are not written to a file}

\item{swarm_path}{Path to th swarm executable (Default: TRUE)}

\item{num_threads}{Number of CPUs}

\item{swarm_d}{positive integer, d parameter for swarm (1 by default); maximum number of differences allowed between two amplicons, meaning that two amplicons will be grouped if they have d (or less) differences.}

\item{fastidious}{\link{T/F} when working with d = 1, perform a second clustering pass to reduce the number of small clusters (Default: TRUE)}

\item{write_csv}{\link{T/F}; write read_counts to csv file; default=FALSE}

\item{sep}{separator for the output file}

\item{by_sample}{\link{T/F}, if TRUE, swarm run separately for each sample}
}
\description{
Run swarm (https://github.com/torognes/swarm) on input read_count_df data frame, pool variants of the same cluster sum reads of the underlying ASVs
Return a data frame with the same structure as the input
Swarm can be run sample by sample (by_sample=T) or for the whole data set in ine go
}
