% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{TaxAssign}
\alias{TaxAssign}
\title{TaxAssign}
\usage{
TaxAssign(
  asv,
  ltg_params = "",
  taxonomy = "",
  blast_db = "",
  blast_path = "",
  outfile = "",
  num_threads = 1,
  tax_sep = "\\t",
  sep = ","
)
}
\arguments{
\item{ltg_params}{csv file or data frame with a list of percentage of identity values (pid) and associated parameters (pcov,phit,taxn,seqn,refres,ltgres)}

\item{taxonomy}{file containing the following columns: tax_id,parent_tax_id,rank,name_txt,old_tax_id(has been mered to tax_id),taxlevel (8: species, 7: genus, 6: family, 5: order, 4: class, 3: phylum, 2: kingdom, 1: superkingdom, 0: root)}

\item{blast_db}{BLAST database}

\item{blast_path}{path to BLAST executable}

\item{outfile}{Name of the output csv file with the following columns: asv,ltg_taxid,ltg_name,ltg_rank,ltg_rank_index,superkingdom_taxid,superkingdom,kingdom_taxid,kingdom,phylum_taxid,phylum,class_taxid,class,order_taxid,order,family_taxid,family,genus_taxid,genus,species_taxid,species,pid,pcov,phit,taxn,seqn,refres,ltgres; if no file name provided, only a data frame is returned}

\item{num_threads}{Number of CPUs}

\item{tax_sep}{separator used in taxonomy file}

\item{sep}{separator used in csv files}

\item{df}{csv file or data frame containing and asv column}

\item{outdir}{name of the output directory}
}
\description{
Find LTG for each asv in the input dataframe
}
