% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_test.R
\name{Test_Filters}
\alias{Test_Filters}
\title{Test_Filters}
\usage{
Test_Filters(
  test_dir = "vtamR_test/",
  vsearch_path = "vsearch",
  swarm_path = "swarm",
  sep = ",",
  delete_tmp = T,
  quiet = T
)
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{vsearch_path}{path to vsearch executables}

\item{swarm_path}{Character string: path to swarm executables.}

\item{sep}{separator of the csv files}

\item{delete_tmp}{Boolean. Delete output folder}

\item{quiet}{Boolean: print only warnings and errors to STDOUT.}
}
\description{
Run different filtering steps on a test input and compare results to expected
(pre-computed) output
}
