% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_vtamR.R
\name{test_filters}
\alias{test_filters}
\title{test_filters}
\usage{
test_filters(test_dir = "~/vtamR/vtamR_test/", vsearch_path = "", sep = ",")
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{vsearch_path}{path to vsearch executables}

\item{sep}{separator of the csv files}
}
\description{
Run different filtering steps on a test input and compare results to expected (pre-computed) output
}
