% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_test.R
\name{test_Filters}
\alias{test_Filters}
\title{test_Filters}
\usage{
test_Filters(
  test_dir = "vtamR_test/",
  vsearch_path = "",
  swarm_path = "",
  sep = ","
)
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{vsearch_path}{path to vsearch executables}

\item{sep}{separator of the csv files}
}
\description{
Run different filtering steps on a test input and compare results to expected (pre-computed) output
}
