% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_test.R
\name{Test_MakeKnownOccurrences}
\alias{Test_MakeKnownOccurrences}
\title{Test_MakeKnownOccurrences}
\usage{
Test_MakeKnownOccurrences(test_dir = "vtamR_test/", sep = ",", delete_tmp = T)
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{sep}{separator in csv files}

\item{delete_tmp}{\link{T/F}  Delete output folder}
}
\description{
Run MakeKnownOccurrences on a test file and compare the output to expected results
}
