% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_test.R
\name{test_Merge_and_SortReads}
\alias{test_Merge_and_SortReads}
\title{test_Merge_and_SortReads}
\usage{
test_Merge_and_SortReads(
  test_dir = "vtamR_test/",
  vsearch_path = "",
  cutadapt_path = ""
)
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{vsearch_path}{path to vsearch executables}

\item{cutadapt_path}{path to cutadapt executables}
}
\description{
Compare the Merge and SortReads output (using default values of vtam) of vtamR to pre-computed files obtained by vtam
}
