% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_test.R
\name{Test_Optimize}
\alias{Test_Optimize}
\title{Test_Optimize}
\usage{
Test_Optimize(
  test_dir = "vtamR_test/",
  vsearch_path = "vsearch",
  delete_tmp = T,
  sep = ",",
  quiet = T
)
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{vsearch_path}{path to vsearch executable}

\item{delete_tmp}{Boolean.  Delete output folder}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{Boolean: print only warnings and errors to STDOUT.}
}
\description{
Run OptimizePCRerror, OptimizeLFNsampleReplicate and
OptimizeLFNreadCountLFNvariant on a test file and compare the output to expected results
}
