% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_test.R
\name{test_TaxAssign}
\alias{test_TaxAssign}
\title{test_TaxAssign}
\usage{
test_TaxAssign(
  test_dir = "vtamR_test/",
  sep = ",",
  blast_path = blast_path,
  blast_db = "vtamR_test/test/db_test/COInr_reduced",
  taxonomy = "vtamR_test/test/db_test/taxonomy_reduced.tsv",
  num_threads = 1
)
}
\arguments{
\item{test_dir}{directory of the test files (Default "~/vtamR/vtamR_test/")}

\item{sep}{separator in csv files}

\item{blast_path}{path to BLAST executable}

\item{blast_db}{BLAST database}

\item{taxonomy}{file containing the following columns: tax_id,parent_tax_id,rank,name_txt,old_tax_id (has been merged to another tax_id),taxlevel (8: species, 7: genus, 6: family, 5: order, 4: class, 3: phylum, 2: kingdom, 1: superkingdom, 0: root)}

\item{num_threads}{Number of CPUs}
}
\description{
Run TaxAssign on a test file and compare the output to mkLTG results
}
