% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{trim_primers}
\alias{trim_primers}
\title{Trim primers from fasta files}
\usage{
trim_primers(
  fastainfo,
  fasta_dir = "",
  outdir = "",
  compress = F,
  compress_method = "R",
  pigz_path = "pigz",
  cutadapt_path = "cutadapt",
  vsearch_path = "vsearch",
  check_reverse = F,
  primer_to_end = T,
  cutadapt_error_rate = 0.1,
  cutadapt_minimum_length = 50,
  cutadapt_maximum_length = 500,
  sep = ",",
  quiet = T
)
}
\arguments{
\item{fastainfo}{Data frame or csv file with the following columns:
tag_fw,primer_fw,tag_rv,primer_rv,sample,sample_type,habitat,replicate,fasta,read_count}

\item{fasta_dir}{Character string: directory of the fasta files to be trimmed}

\item{outdir}{Character string: output directory for the trimmed fasta files.}

\item{compress}{logical: Compress output files to gzip format.}

\item{compress_method}{Character or logical. Compression method: \code{"pigz"}, \code{"gzip"}, or \code{"R"}.
\code{"pigz"} requires \code{pigz} to be installed and in the system path (or \code{pigz_path} specified).
\code{"gzip"} is Linux-only.
\code{"R"} uses \code{R.utils}, which is cross-platform but slower.
Speed: R.utils < gzip (only linux) < pig}

\item{pigz_path}{Character string: Path to \code{pigz} executable. Only needed is pigz
is used for file compression, and it is not in the PATH.}

\item{cutadapt_path}{Character string: path to cutadapt executables.}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{check_reverse}{logical: if TRUE, check the reverse complementary sequences
of the input fasta as well.}

\item{primer_to_end}{logical: primers follow directly the tags (no heterogeneity spacer).}

\item{cutadapt_error_rate}{Numeric. Value between 0 and 1: maximum proportion of errors
between primers and reads (for tags, exact match is required).}

\item{cutadapt_minimum_length}{Positive integer: minimum length of the
trimmed sequence.}

\item{cutadapt_maximum_length}{Positive integer: maximum length of the
trimmed sequence.}

\item{sep}{Field separator character in input and output csv files.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
fastainfo_df the input data frame, with updated files names and sequence counts.
}
\description{
Trim primers from each fasta file in input fastainfo data frame or csv file.
Keep only trimmed reads.
Can check both orientations.
Count the number of reads in the output files.
}
\examples{
\dontrun{
fastainfo_df <- trim_primers(fastainfo, 
    fasta_dir="data/fasta", 
    outdir="out", 
    compress=T, 
    check_reverse=T, 
    primer_to_end=F, 
    cutadapt_error_rate=0.1, 
    cutadapt_minimum_length=300, 
    cutadapt_maximum_length=350, 
    quiet=T
    )
}
}
