% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{trim_primers_file}
\alias{trim_primers_file}
\title{Trim primers}
\usage{
trim_primers_file(
  fasta,
  outfile,
  primer_fw,
  primer_rv,
  cutadapt_path = "cutadapt",
  vsearch_path = "vsearch",
  compress_method = "R",
  pigz_path = "pigz",
  num_threads = 0,
  check_reverse = F,
  primer_to_end = T,
  cutadapt_error_rate = 0.1,
  cutadapt_minimum_length = 50,
  cutadapt_maximum_length = 500,
  quiet = T
)
}
\arguments{
\item{fasta}{Character string: input fasta file (icluding path).}

\item{outfile}{Character string: output fasta file (icluding path).}

\item{primer_fw}{Character string: forward primer (IUPAC ambiguity codes are accepted).}

\item{primer_rv}{Character string: reverse primer (IUPAC ambiguity codes are accepted).}

\item{cutadapt_path}{Character string: path to cutadapt executables.}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{compress_method}{Character or logical. Compression method: \code{"pigz"}, \code{"gzip"}, or \code{"R"}.
\code{"pigz"} requires \code{pigz} to be installed and in the system path (or \code{pigz_path} specified).
\code{"gzip"} is Linux-only.
\code{"R"} uses \code{R.utils}, which is cross-platform but slower.
Speed: R.utils < gzip (only linux) < pig
Only needed if \code{check_reverse} is TRUE.}

\item{pigz_path}{Character string: Path to \code{pigz} executable. Only needed is pigz
is used for file compression, and it is not in the PATH.}

\item{num_threads}{Positive integer: Number of CPUs. If 0, use all available CPUs.}

\item{check_reverse}{logical: if TRUE, check the reverse complementary
sequences of the input fasta as well.}

\item{primer_to_end}{logical: primers follow directly the tags
(no heterogeneity spacer).}

\item{cutadapt_error_rate}{Numeric. Value between 0 and 1: maximum proportion of errors
between primers and reads (for tags, exact match is required).}

\item{cutadapt_minimum_length}{Positive integer: minimum length of the
trimmed sequence.}

\item{cutadapt_maximum_length}{Positive integer: maximum length of the
trimmed sequence.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
NULL; Produces an output fasta file.
}
\description{
Trim primers from the input fasta file

Input fasta can be uncompressed or gzip compressed file,
but other compression types are not supported.
The outfut file compression is determined from the output filename.
}
\examples{
\dontrun{
trim_primers_file(fasta="data/test.fasta", 
    outfile="out/test_trimmed.fasta", 
    primer_fw="TCCACTAATCACAARGATATTGGTAC", 
    primer_rv="WACTAATCAATTWCCAAATCCTCC", 
    check_reverse=T, 
    primer_to_end=T, 
    cutadapt_minimum_length=300,
    cutadapt_maximum_length=400
    )
}
}
