% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{TrimPrimer}
\alias{TrimPrimer}
\title{Trim primers from fasta files}
\usage{
TrimPrimer(
  fastainfo,
  fasta_dir = "",
  outdir = "",
  compress = F,
  cutadapt_path = "cutadapt",
  vsearch_path = "vsearch",
  check_reverse = F,
  primer_to_end = T,
  cutadapt_error_rate = 0.1,
  cutadapt_minimum_length = 50,
  cutadapt_maximum_length = 500,
  quiet = T
)
}
\arguments{
\item{fastainfo}{Data frame of csv file with the following columns:
tag_fw,primer_fw,tag_rv,primer_rv,sample,sample_type,habitat,replicate,fasta,read_count}

\item{fasta_dir}{Character string: directory of the fasta files to be trimmed}

\item{outdir}{Character string: output directory for the trimmed fasta files.}

\item{compress}{logical: Compress output files to gzip format.}

\item{cutadapt_path}{Character string: path to cutadapt executables.}

\item{vsearch_path}{Character string: path to vsearch executables.}

\item{check_reverse}{logical: if TRUE, check the reverse complementary sequences
of the input fasta as well.}

\item{primer_to_end}{logical: primers follow directly the tags (no heterogeneity spacer).}

\item{cutadapt_error_rate}{Real (0-1): maximum proportion of errors
between primers and reads (for tags, exact match is required).}

\item{cutadapt_minimum_length}{Positive integer: minimum length of the
trimmed sequence.}

\item{cutadapt_maximum_length}{Positive integer: maximum length of the
trimmed sequence.}

\item{quiet}{logical: If TRUE, suppress informational messages and only
show warnings or errors.}
}
\value{
fastainfo_df the input data frame, with updated files names and sequence counts.
}
\description{
Trim primers from each fasta file in input fastainfo data frame or csv file.
Keep only trimmed reads.
Can check both orientations.
Count the number of reads in the output files.
}
\examples{
\dontrun{
fastainfo_df <- TrimPrimer(fastainfo, 
    fasta_dir="data/fasta", 
    outdir="out", 
    compress=T, 
    check_reverse=T, 
    primer_to_end=F, 
    cutadapt_error_rate=0.1, 
    cutadapt_minimum_length=300, 
    cutadapt_maximum_length=350, 
    quiet=T
    )
}
}
