% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{TrimPrimer_OneFile}
\alias{TrimPrimer_OneFile}
\title{TrimPrimer_OneFile}
\usage{
TrimPrimer_OneFile(
  fasta,
  outfile = "",
  primer_fw = "",
  primer_rv = "",
  cutadapt_path = "",
  vsearch_path = "",
  check_reverse = F,
  primer_to_end = T,
  cutadapt_error_rate = 0.1,
  cutadapt_minimum_length = 50,
  cutadapt_maximum_length = 500
)
}
\arguments{
\item{fasta}{input fasta file (icluding path);  can be uncompressed or gz, bz2 compressed}

\item{outfile}{input fasta file (icluding path); can be uncompressed  or gz compressed, the other compression types are not supported}

\item{primer_fw}{forward primer (IUPAC ambiguity codes are accepted)}

\item{primer_rv}{reverse primer (IUPAC ambiguity codes are accepted)}

\item{cutadapt_path}{path to cutadapt executables}

\item{vsearch_path}{path to vsearch executables}

\item{check_reverse}{\link{T/F} if TRUE, ckeck the reverse comlementary sequences of the input fasta as well; default: F}

\item{primer_to_end}{primers follow directly the tags (no heterogeneity spacer); default: T}

\item{cutadapt_error_rate}{maximum proportion of errors between primers and reads (for tags, exact match is required); default: 0.1}

\item{cutadapt_minimum_length}{minimum length of the trimmed sequence; default: 50}

\item{cutadapt_maximum_length}{maximum length of the merged sequence; default: 500}
}
\description{
Trim primers from the input fasta file
}
