% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxassign_functions.R
\name{update_taxids}
\alias{update_taxids}
\title{update_taxids}
\usage{
update_taxids(df, old_taxid)
}
\arguments{
\item{df}{data frame with the following columns: qseqid,pident,qcovhsp,staxids}

\item{old_taxid}{data frame with the following columns:  tax_id,old_tax_id}
}
\description{
Replaces old taxids by valid ones
}
