% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{UpdateASVlist}
\alias{UpdateASVlist}
\title{Update ASV list}
\usage{
UpdateASVlist(df, asv_list, outfile, sep = ",")
}
\arguments{
\item{df}{Data frame with columns:
asv_id,asv}

\item{asv_list}{Character string: name of the input CSV file with columns:
asv_id,asv.}

\item{outfile}{Character string: output csv file name.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
Updated csv file with all unique asv_id - asv pairs
in the input data frame and csv file.
If there is a conflict within or between the input data stops with
an error message.
}
\description{
Pools unique asv - asv_id pairs in the input data frame
and in the input file (asv_list).
}
\details{
The input file is typically a csv file containing ASVs with their asv_ids
seen in earlier data sets. This function is used to homogenize
asv_ids between different data sets.
}
\examples{
UpdateASVlist(df=read_count_df, asv_list="data/asv_list.csv", outfile="out/updated_asv_list.csv")
}
