% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{UpdateASVlist}
\alias{UpdateASVlist}
\title{UpdateASVlist}
\usage{
UpdateASVlist(read_count_df, asv_list = asv_list, outfile = "", sep = ",")
}
\arguments{
\item{read_count_df}{data frame with columns:  asv_id, sample,  replicate, read_count, asv}

\item{asv_list}{name of the file, containing asvs and asv_ids from earlier analyses. Optional. It is used to homogenize asv_ids between different data sets}

\item{outfile}{Name of the output file; if empty, write a new file using the name asv_list, completed by the number of seconds from 01/01/1970}

\item{sep}{separator in csv files; default: ","}
}
\description{
Pools unique asv - asv_id combinations in the input data frame and asv - asv_id combinations in the input file (asv_list)
The input file is typically a csv file containing asv seen in earlier runs with their asv_id.
If there is a conflict within or between the input data quits with a error message. Otherwise write the updated asv list with their asv_id to the outfile
}
