% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{write_asv_table}
\alias{write_asv_table}
\title{Write ASV table}
\usage{
write_asv_table(
  read_count,
  outfile = "",
  asv_tax = NULL,
  sampleinfo = "",
  pool_replicates = FALSE,
  method = "mean",
  add_empty_samples = FALSE,
  add_sums_by_sample = FALSE,
  add_sums_by_asv = FALSE,
  add_expected_asv = FALSE,
  mock_composition = "",
  sep = ","
)
}
\arguments{
\item{read_count}{Data frame or csv file with the following variables:
asv_id, sample, replicate (optional), read_count, asv, cluster_id (optional)}

\item{outfile}{Character string: csv file name to print the output data
frame if necessary. If empty, no file is written.}

\item{asv_tax}{Data frame or CSV file with taxonomic assignments with the following columns:
asv_id,ltg_taxid,ltg_name,ltg_rank,ltg_rank_index,
domain_taxid,domain,kingdom_taxid,kingdom,
phylum_taxid,phylum,class_taxid,class,order_taxid,order,
family_taxid,family,genus_taxid,genus,species_taxid,species.
pid,pcov,phit,taxn,seqn,refres,ltgres,asv.
If given, the output is completed with taxonomic assignment of each ASV.}

\item{sampleinfo}{Data frame or csv file with columns: sample, sample_type.
Only necessary if add_empty_samples==T or add_expected_asv==T.}

\item{pool_replicates}{Logical: if replicate column is present in the
input data, read counts are aggregated (see method) over replicates (pool_replicates== TRUE),
or kept separately (pool_replicates== FALSE).
In this case, each column in the output df is a
sample.replicate.}

\item{method}{Character string specifying how replicate read counts should be aggregated
Must be one of "mean", "max", "sum" or "min".}

\item{add_empty_samples}{Logical: add a column for each samples
in the original data set, even if they do not have reads after filtering.}

\item{add_sums_by_sample}{Logical: add a line with the total number of reads
in each sample, and another with the number of ASVs in each sample.}

\item{add_sums_by_asv}{Logical: add a column with the total number of reads
for each ASV, and another with the number of samples, where the ASV is present.}

\item{add_expected_asv}{Logical: add a column for each mock sample in which
keep and tolerate ASVs are flagged.}

\item{mock_composition}{Data frame or CSV file with the following columns:
sample,action,asv. Action can take the following values: keep/tolerate.
Only necessary if add_expected_asv==T.}

\item{sep}{Field separator character in input and output csv files.}
}
\value{
An invisible data frame corresponding to the output file.
Columns represent samples, rows represent ASVs, and cells contain
read counts, along with additional information depending on the
input parameters.
}
\description{
Write csv file with samples in columns, ASVs in lines, read_counts in cells.
}
\examples{
\dontrun{
write_asv_table(read_count_samples_df, 
    outfile="out/asv_table.csv", 
    asv_tax=asv_tax, 
    sampleinfo=sampleinfo_df,
    add_empty_samples=T, 
    add_sums_by_sample=T,
    add_sums_by_asv=T, 
    add_expected_asv=T,
    mock_composition="data/mock_compostion.csv"
    )
}
}
