% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{write_cutadapt_adapter_fasta}
\alias{write_cutadapt_adapter_fasta}
\title{Make a FASTA file with adapters}
\usage{
write_cutadapt_adapter_fasta(fastainfo_df, fasta_file, outdir, tag_to_end = T)
}
\arguments{
\item{fastainfo_df}{Data frame with columns: \code{tag_fw}, \code{tag_rv}, \code{fasta}.}

\item{fasta_file}{Character string specifying the FASTA file to be demultiplexed
(must be present in the \code{fasta} column of \code{fastainfo_df}).}

\item{outdir}{Character string specifying the output directory.}

\item{tag_to_end}{Logical. If \code{TRUE}, tags are assumed to be located at the
extremities of reads.}
}
\value{
Path to the generated \code{tags.fasta} file. Returns \code{NA} if all tags
are \code{NA} in \code{fastainfo_df} for the given \code{fasta_file}.
}
\description{
Create a FASTA file containing tag combinations formatted for \code{cutadapt}.
This file is used by \code{demultiplex_and_trim} to demultiplex input FASTA files.
}
\examples{
\dontrun{
write_cutadapt_adapter_fasta(
  fastainfo_df = fastainfo_df, 
  fasta_file = "test.fasta", 
  tag_to_end = FALSE, 
  outdir = "data/out"
)
}

}
