% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtamR_fun.R
\name{write_fasta}
\alias{write_fasta}
\title{Write fasta file}
\usage{
write_fasta(sequences, filename, seq_as_id = F)
}
\arguments{
\item{sequences}{Vector of sequences.}

\item{filename}{Character string: name of the output fasta file.}

\item{seq_as_id}{Boolean: Use sequences as seqID.}
}
\value{
fasta file
}
\description{
Write a vector of sequeces to a fasta file using
either the sequences as ID (seq_as_id==TRUE), or arbitrary sequence IDs
(seq_as_id==FALSE)
}
\examples{
write_fasta(sequences=read_count_df$asv, filename="out/seq.fasta", seq_as_id=T)
}
